chore: emoji data 분리
