+++ emojis/activity.js
@@ -1,0 +1,217 @@
+export default [
+  {
+    codes: '1F383',
+    char: '🎃',
+    name: 'jack-o-lantern',
+    category: 'activity',
+  },
+  {
+    codes: '1F384',
+    char: '🎄',
+    name: 'Christmas tree',
+    category: 'activity',
+  },
+  {
+    codes: '1F386',
+    char: '🎆',
+    name: 'fireworks',
+    category: 'activity',
+  },
+  {
+    codes: '1F387',
+    char: '🎇',
+    name: 'sparkler',
+    category: 'activity',
+  },
+  { codes: '2728', char: '✨', name: 'sparkles', category: 'activity' },
+  { codes: '1F388', char: '🎈', name: 'balloon', category: 'activity' },
+  {
+    codes: '1F389',
+    char: '🎉',
+    name: 'party popper',
+    category: 'activity',
+  },
+  {
+    codes: '1F38A',
+    char: '🎊',
+    name: 'confetti ball',
+    category: 'activity',
+  },
+  {
+    codes: '1F38B',
+    char: '🎋',
+    name: 'tanabata tree',
+    category: 'activity',
+  },
+  {
+    codes: '1F38D',
+    char: '🎍',
+    name: 'pine decoration',
+    category: 'activity',
+  },
+  {
+    codes: '1F38E',
+    char: '🎎',
+    name: 'Japanese dolls',
+    category: 'activity',
+  },
+  {
+    codes: '1F38F',
+    char: '🎏',
+    name: 'carp streamer',
+    category: 'activity',
+  },
+  { codes: '1F380', char: '🎀', name: 'ribbon', category: 'activity' },
+  {
+    codes: '1F381',
+    char: '🎁',
+    name: 'wrapped gift',
+    category: 'activity',
+  },
+  { codes: '1F3C6', char: '🏆', name: 'trophy', category: 'activity' },
+  {
+    codes: '26BD',
+    char: '⚽',
+    name: 'soccer ball',
+    category: 'activity',
+  },
+  { codes: '26BE', char: '⚾', name: 'baseball', category: 'activity' },
+  {
+    codes: '1F3C0',
+    char: '🏀',
+    name: 'basketball',
+    category: 'activity',
+  },
+  {
+    codes: '1F3C8',
+    char: '🏈',
+    name: 'american football',
+    category: 'activity',
+  },
+  {
+    codes: '1F3C9',
+    char: '🏉',
+    name: 'rugby football',
+    category: 'activity',
+  },
+  { codes: '1F3BE', char: '🎾', name: 'tennis', category: 'activity' },
+  { codes: '1F3B3', char: '🎳', name: 'bowling', category: 'activity' },
+  {
+    codes: '26F3',
+    char: '⛳',
+    name: 'flag in hole',
+    category: 'activity',
+  },
+  {
+    codes: '1F3A3',
+    char: '🎣',
+    name: 'fishing pole',
+    category: 'activity',
+  },
+  {
+    codes: '1F3BD',
+    char: '🎽',
+    name: 'running shirt',
+    category: 'activity',
+  },
+  { codes: '1F3BF', char: '🎿', name: 'skis', category: 'activity' },
+
+  {
+    codes: '1F3AF',
+    char: '🎯',
+    name: 'direct hit',
+    category: 'activity',
+  },
+
+  {
+    codes: '1F3B1',
+    char: '🎱',
+    name: 'pool 8 ball',
+    category: 'activity',
+  },
+  {
+    codes: '1F52E',
+    char: '🔮',
+    name: 'crystal ball',
+    category: 'activity',
+  },
+  {
+    codes: '1F3AE',
+    char: '🎮',
+    name: 'video game',
+    category: 'activity',
+  },
+  {
+    codes: '1F3B0',
+    char: '🎰',
+    name: 'slot machine',
+    category: 'activity',
+  },
+  {
+    codes: '1F3B2',
+    char: '🎲',
+    name: 'game die',
+    category: 'activity',
+  },
+  {
+    codes: '2666 FE0F',
+    char: '♦️',
+    name: 'diamond suit',
+    category: 'activity',
+  },
+  {
+    codes: '2663 FE0F',
+    char: '♣️',
+    name: 'club suit',
+    category: 'activity',
+  },
+  {
+    codes: '265F FE0F',
+    char: '♟️',
+    name: 'chess pawn',
+    category: 'activity',
+  },
+  {
+    codes: '1F5BC FE0F',
+    char: '🖼️',
+    name: 'framed picture',
+    category: 'activity',
+  },
+  {
+    codes: '1F9F5',
+    char: '🧵',
+    name: 'thread',
+    category: 'activity',
+  },
+  {
+    codes: '1F9F6',
+    char: '🧶',
+    name: 'yarn',
+    category: 'activity',
+  },
+
+  {
+    codes: '1F004',
+    char: '🀄',
+    name: 'mahjong red dragon',
+    category: 'activity',
+  },
+  {
+    codes: '1F3B4',
+    char: '🎴',
+    name: 'flower playing cards',
+    category: 'activity',
+  },
+  {
+    codes: '1F3AD',
+    char: '🎭',
+    name: 'performing arts',
+    category: 'activity',
+  },
+  {
+    codes: '1F3A8',
+    char: '🎨',
+    name: 'artist palette',
+    category: 'activity',
+  },
+];
