--- emojis/emojis.js
@@ -1,7265 +1,0 @@
-export default [
-  {
-    codes: '1F600',
-    char: '😀',
-    name: 'grinning face',
-    category: 'people',
-  },
-  {
-    codes: '1F604',
-    char: '😄',
-    name: 'grinning face with smiling eyes',
-    category: 'people',
-  },
-  {
-    codes: '1F60A',
-    char: '😊',
-    name: 'smiling face with smiling eyes',
-    category: 'people',
-  },
-  {
-    codes: '1F601',
-    char: '😁',
-    name: 'beaming face with smiling eyes',
-    category: 'people',
-  },
-  {
-    codes: '1F606',
-    char: '😆',
-    name: 'grinning squinting face',
-    category: 'people',
-  },
-  {
-    codes: '1F605',
-    char: '😅',
-    name: 'grinning face with sweat',
-    category: 'people',
-  },
-  {
-    codes: '1F602',
-    char: '😂',
-    name: 'face with tears of joy',
-    category: 'people',
-  },
-  {
-    codes: '1F642',
-    char: '🙂',
-    name: 'slightly smiling face',
-    category: 'people',
-  },
-  {
-    codes: '1F609',
-    char: '😉',
-    name: 'winking face',
-    category: 'people',
-  },
-  {
-    codes: '1F607',
-    char: '😇',
-    name: 'smiling face with halo',
-    category: 'people',
-  },
-  {
-    codes: '1F60D',
-    char: '😍',
-    name: 'smiling face with heart-eyes',
-    category: 'people',
-  },
-  {
-    codes: '1F618',
-    char: '😘',
-    name: 'face blowing a kiss',
-    category: 'people',
-  },
-  {
-    codes: '1F617',
-    char: '😗',
-    name: 'kissing face',
-    category: 'people',
-  },
-  {
-    codes: '1F61A',
-    char: '😚',
-    name: 'kissing face with closed eyes',
-    category: 'people',
-  },
-  {
-    codes: '1F60B',
-    char: '😋',
-    name: 'face savoring food',
-    category: 'people',
-  },
-  {
-    codes: '1F61B',
-    char: '😛',
-    name: 'face with tongue',
-    category: 'people',
-  },
-  {
-    codes: '1F61C',
-    char: '😜',
-    name: 'winking face with tongue',
-    category: 'people',
-  },
-  {
-    codes: '1F61D',
-    char: '😝',
-    name: 'squinting face with tongue',
-    category: 'people',
-  },
-  {
-    codes: '1F911',
-    char: '🤑',
-    name: 'money-mouth face',
-    category: 'people',
-  },
-  {
-    codes: '1F917',
-    char: '🤗',
-    name: 'hugging face',
-    category: 'people',
-  },
-  {
-    codes: '1F92D',
-    char: '🤭',
-    name: 'face with hand over mouth',
-    category: 'people',
-  },
-  {
-    codes: '1F92B',
-    char: '🤫',
-    name: 'shushing face',
-    category: 'people',
-  },
-  {
-    codes: '1F914',
-    char: '🤔',
-    name: 'thinking face',
-    category: 'people',
-  },
-  {
-    codes: '1F910',
-    char: '🤐',
-    name: 'zipper-mouth face',
-    category: 'people',
-  },
-  {
-    codes: '1F928',
-    char: '🤨',
-    name: 'face with raised eyebrow',
-    category: 'people',
-  },
-  {
-    codes: '1F610',
-    char: '😐',
-    name: 'neutral face',
-    category: 'people',
-  },
-  {
-    codes: '1F611',
-    char: '😑',
-    name: 'expressionless face',
-    category: 'people',
-  },
-  {
-    codes: '1F60F',
-    char: '😏',
-    name: 'smirking face',
-    category: 'people',
-  },
-  {
-    codes: '1F612',
-    char: '😒',
-    name: 'unamused face',
-    category: 'people',
-  },
-  {
-    codes: '1F62C',
-    char: '😬',
-    name: 'grimacing face',
-    category: 'people',
-  },
-  {
-    codes: '1F60C',
-    char: '😌',
-    name: 'relieved face',
-    category: 'people',
-  },
-  {
-    codes: '1F614',
-    char: '😔',
-    name: 'pensive face',
-    category: 'people',
-  },
-  {
-    codes: '1F62A',
-    char: '😪',
-    name: 'sleepy face',
-    category: 'people',
-  },
-  {
-    codes: '1F634',
-    char: '😴',
-    name: 'sleeping face',
-    category: 'people',
-  },
-  {
-    codes: '1F637',
-    char: '😷',
-    name: 'face with medical mask',
-    category: 'people',
-  },
-  {
-    codes: '1F635',
-    char: '😵',
-    name: 'dizzy face',
-    category: 'people',
-  },
-  {
-    codes: '1F920',
-    char: '🤠',
-    name: 'cowboy hat face',
-    category: 'people)',
-  },
-  {
-    codes: '1F973',
-    char: '🥳',
-    name: 'partying face',
-    category: 'people)',
-  },
-  {
-    codes: '1F60E',
-    char: '😎',
-    name: 'smiling face with sunglasses',
-    category: 'people',
-  },
-  {
-    codes: '1F615',
-    char: '😕',
-    name: 'confused face',
-    category: 'people',
-  },
-  {
-    codes: '1F61F',
-    char: '😟',
-    name: 'worried face',
-    category: 'people',
-  },
-  {
-    codes: '1F641',
-    char: '🙁',
-    name: 'slightly frowning face',
-    category: 'people',
-  },
-  {
-    codes: '1F62E',
-    char: '😮',
-    name: 'face with open mouth',
-    category: 'people',
-  },
-  {
-    codes: '1F632',
-    char: '😲',
-    name: 'astonished face',
-    category: 'people',
-  },
-  {
-    codes: '1F633',
-    char: '😳',
-    name: 'flushed face',
-    category: 'people',
-  },
-  {
-    codes: '1F626',
-    char: '😦',
-    name: 'frowning face with open mouth',
-    category: 'people',
-  },
-  {
-    codes: '1F628',
-    char: '😨',
-    name: 'fearful face',
-    category: 'people',
-  },
-  {
-    codes: '1F630',
-    char: '😰',
-    name: 'anxious face with sweat',
-    category: 'people',
-  },
-  {
-    codes: '1F625',
-    char: '😥',
-    name: 'sad but relieved face',
-    category: 'people',
-  },
-  {
-    codes: '1F62D',
-    char: '😭',
-    name: 'loudly crying face',
-    category: 'people',
-  },
-  {
-    codes: '1F631',
-    char: '😱',
-    name: 'face screaming in fear',
-    category: 'people',
-  },
-  {
-    codes: '1F616',
-    char: '😖',
-    name: 'confounded face',
-    category: 'people',
-  },
-  {
-    codes: '1F61E',
-    char: '😞',
-    name: 'disappointed face',
-    category: 'people',
-  },
-  {
-    codes: '1F613',
-    char: '😓',
-    name: 'downcast face with sweat',
-    category: 'people',
-  },
-  {
-    codes: '1F629',
-    char: '😩',
-    name: 'weary face',
-    category: 'people',
-  },
-  {
-    codes: '1F62B',
-    char: '😫',
-    name: 'tired face',
-    category: 'people',
-  },
-  {
-    codes: '1F624',
-    char: '😤',
-    name: 'face with steam from nose',
-    category: 'people',
-  },
-  {
-    codes: '1F621',
-    char: '😡',
-    name: 'pouting face',
-    category: 'people',
-  },
-  {
-    codes: '1F620',
-    char: '😠',
-    name: 'angry face',
-    category: 'people',
-  },
-  {
-    codes: '1F608',
-    char: '😈',
-    name: 'smiling face with horns',
-    category: 'people',
-  },
-  { codes: '1F480', char: '💀', name: 'skull', category: 'people' },
-  {
-    codes: '1F4A9',
-    char: '💩',
-    name: 'pile of poo',
-    category: 'people',
-  },
-  { codes: '1F479', char: '👹', name: 'ogre', category: 'people' },
-  { codes: '1F47A', char: '👺', name: 'goblin', category: 'people' },
-  { codes: '1F47B', char: '👻', name: 'ghost', category: 'people' },
-  { codes: '1F47D', char: '👽', name: 'alien', category: 'people' },
-  {
-    codes: '1F47E',
-    char: '👾',
-    name: 'alien monster',
-    category: 'people',
-  },
-  {
-    codes: '1F648',
-    char: '🙈',
-    name: 'see-no-evil monkey',
-    category: 'people',
-  },
-  {
-    codes: '1F649',
-    char: '🙉',
-    name: 'hear-no-evil monkey',
-    category: 'people',
-  },
-  {
-    codes: '1F64A',
-    char: '🙊',
-    name: 'speak-no-evil monkey',
-    category: 'people',
-  },
-  {
-    codes: '1F44B',
-    char: '👋',
-    name: 'waving hand',
-    category: 'people',
-  },
-  {
-    codes: '1F44B 1F3FB',
-    char: '👋🏻',
-    name: 'waving hand: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44B 1F3FC',
-    char: '👋🏼',
-    name: 'waving hand: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44B 1F3FD',
-    char: '👋🏽',
-    name: 'waving hand: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44B 1F3FE',
-    char: '👋🏾',
-    name: 'waving hand: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44B 1F3FF',
-    char: '👋🏿',
-    name: 'waving hand: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270B',
-    char: '✋',
-    name: 'raised hand',
-    category: 'people',
-  },
-  {
-    codes: '270B 1F3FB',
-    char: '✋🏻',
-    name: 'raised hand: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270B 1F3FC',
-    char: '✋🏼',
-    name: 'raised hand: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270B 1F3FD',
-    char: '✋🏽',
-    name: 'raised hand: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270B 1F3FE',
-    char: '✋🏾',
-    name: 'raised hand: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270B 1F3FF',
-    char: '✋🏿',
-    name: 'raised hand: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F44C', char: '👌', name: 'OK hand', category: 'people' },
-  {
-    codes: '1F44C 1F3FB',
-    char: '👌🏻',
-    name: 'OK hand: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44C 1F3FC',
-    char: '👌🏼',
-    name: 'OK hand: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44C 1F3FD',
-    char: '👌🏽',
-    name: 'OK hand: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44C 1F3FE',
-    char: '👌🏾',
-    name: 'OK hand: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44C 1F3FF',
-    char: '👌🏿',
-    name: 'OK hand: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270C FE0F',
-    char: '✌️',
-    name: 'victory hand',
-    category: 'people',
-  },
-  {
-    codes: '270C 1F3FB',
-    char: '✌🏻',
-    name: 'victory hand: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270C 1F3FC',
-    char: '✌🏼',
-    name: 'victory hand: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270C 1F3FD',
-    char: '✌🏽',
-    name: 'victory hand: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270C 1F3FE',
-    char: '✌🏾',
-    name: 'victory hand: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270C 1F3FF',
-    char: '✌🏿',
-    name: 'victory hand: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F448',
-    char: '👈',
-    name: 'backhand index pointing left',
-    category: 'people',
-  },
-  {
-    codes: '1F448 1F3FB',
-    char: '👈🏻',
-    name: 'backhand index pointing left: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F448 1F3FC',
-    char: '👈🏼',
-    name: 'backhand index pointing left: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F448 1F3FD',
-    char: '👈🏽',
-    name: 'backhand index pointing left: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F448 1F3FE',
-    char: '👈🏾',
-    name: 'backhand index pointing left: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F448 1F3FF',
-    char: '👈🏿',
-    name: 'backhand index pointing left: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F449',
-    char: '👉',
-    name: 'backhand index pointing right',
-    category: 'people',
-  },
-  {
-    codes: '1F449 1F3FB',
-    char: '👉🏻',
-    name: 'backhand index pointing right: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F449 1F3FC',
-    char: '👉🏼',
-    name: 'backhand index pointing right: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F449 1F3FD',
-    char: '👉🏽',
-    name: 'backhand index pointing right: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F449 1F3FE',
-    char: '👉🏾',
-    name: 'backhand index pointing right: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F449 1F3FF',
-    char: '👉🏿',
-    name: 'backhand index pointing right: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F446',
-    char: '👆',
-    name: 'backhand index pointing up',
-    category: 'people',
-  },
-  {
-    codes: '1F446 1F3FB',
-    char: '👆🏻',
-    name: 'backhand index pointing up: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F446 1F3FC',
-    char: '👆🏼',
-    name: 'backhand index pointing up: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F446 1F3FD',
-    char: '👆🏽',
-    name: 'backhand index pointing up: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F446 1F3FE',
-    char: '👆🏾',
-    name: 'backhand index pointing up: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F446 1F3FF',
-    char: '👆🏿',
-    name: 'backhand index pointing up: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F447',
-    char: '👇',
-    name: 'backhand index pointing down',
-    category: 'people',
-  },
-  {
-    codes: '1F447 1F3FB',
-    char: '👇🏻',
-    name: 'backhand index pointing down: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F447 1F3FC',
-    char: '👇🏼',
-    name: 'backhand index pointing down: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F447 1F3FD',
-    char: '👇🏽',
-    name: 'backhand index pointing down: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F447 1F3FE',
-    char: '👇🏾',
-    name: 'backhand index pointing down: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F447 1F3FF',
-    char: '👇🏿',
-    name: 'backhand index pointing down: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '261D FE0F',
-    char: '☝️',
-    name: 'index pointing up',
-    category: 'people',
-  },
-  {
-    codes: '261D',
-    char: '☝',
-    name: 'index pointing up',
-    category: 'people',
-  },
-  {
-    codes: '261D 1F3FB',
-    char: '☝🏻',
-    name: 'index pointing up: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '261D 1F3FC',
-    char: '☝🏼',
-    name: 'index pointing up: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '261D 1F3FD',
-    char: '☝🏽',
-    name: 'index pointing up: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '261D 1F3FE',
-    char: '☝🏾',
-    name: 'index pointing up: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '261D 1F3FF',
-    char: '☝🏿',
-    name: 'index pointing up: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F44D', char: '👍', name: 'thumbs up', category: 'people' },
-  {
-    codes: '1F44D 1F3FB',
-    char: '👍🏻',
-    name: 'thumbs up: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44D 1F3FC',
-    char: '👍🏼',
-    name: 'thumbs up: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44D 1F3FD',
-    char: '👍🏽',
-    name: 'thumbs up: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44D 1F3FE',
-    char: '👍🏾',
-    name: 'thumbs up: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44D 1F3FF',
-    char: '👍🏿',
-    name: 'thumbs up: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44E',
-    char: '👎',
-    name: 'thumbs down',
-    category: 'people',
-  },
-  {
-    codes: '1F44E 1F3FB',
-    char: '👎🏻',
-    name: 'thumbs down: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44E 1F3FC',
-    char: '👎🏼',
-    name: 'thumbs down: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44E 1F3FD',
-    char: '👎🏽',
-    name: 'thumbs down: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44E 1F3FE',
-    char: '👎🏾',
-    name: 'thumbs down: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44E 1F3FF',
-    char: '👎🏿',
-    name: 'thumbs down: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270A',
-    char: '✊',
-    name: 'raised fist',
-    category: 'people',
-  },
-  {
-    codes: '270A 1F3FB',
-    char: '✊🏻',
-    name: 'raised fist: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270A 1F3FC',
-    char: '✊🏼',
-    name: 'raised fist: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270A 1F3FD',
-    char: '✊🏽',
-    name: 'raised fist: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270A 1F3FE',
-    char: '✊🏾',
-    name: 'raised fist: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '270A 1F3FF',
-    char: '✊🏿',
-    name: 'raised fist: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44A',
-    char: '👊',
-    name: 'oncoming fist',
-    category: 'people',
-  },
-  {
-    codes: '1F44A 1F3FB',
-    char: '👊🏻',
-    name: 'oncoming fist: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44A 1F3FC',
-    char: '👊🏼',
-    name: 'oncoming fist: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44A 1F3FD',
-    char: '👊🏽',
-    name: 'oncoming fist: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44A 1F3FE',
-    char: '👊🏾',
-    name: 'oncoming fist: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44A 1F3FF',
-    char: '👊🏿',
-    name: 'oncoming fist: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44F',
-    char: '👏',
-    name: 'clapping hands',
-    category: 'people',
-  },
-  {
-    codes: '1F44F 1F3FB',
-    char: '👏🏻',
-    name: 'clapping hands: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44F 1F3FC',
-    char: '👏🏼',
-    name: 'clapping hands: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44F 1F3FD',
-    char: '👏🏽',
-    name: 'clapping hands: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44F 1F3FE',
-    char: '👏🏾',
-    name: 'clapping hands: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F44F 1F3FF',
-    char: '👏🏿',
-    name: 'clapping hands: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64C',
-    char: '🙌',
-    name: 'raising hands',
-    category: 'people',
-  },
-  {
-    codes: '1F64C 1F3FB',
-    char: '🙌🏻',
-    name: 'raising hands: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64C 1F3FC',
-    char: '🙌🏼',
-    name: 'raising hands: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64C 1F3FD',
-    char: '🙌🏽',
-    name: 'raising hands: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64C 1F3FE',
-    char: '🙌🏾',
-    name: 'raising hands: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64C 1F3FF',
-    char: '🙌🏿',
-    name: 'raising hands: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F450',
-    char: '👐',
-    name: 'open hands',
-    category: 'people',
-  },
-  {
-    codes: '1F450 1F3FB',
-    char: '👐🏻',
-    name: 'open hands: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F450 1F3FC',
-    char: '👐🏼',
-    name: 'open hands: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F450 1F3FD',
-    char: '👐🏽',
-    name: 'open hands: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F450 1F3FE',
-    char: '👐🏾',
-    name: 'open hands: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F450 1F3FF',
-    char: '👐🏿',
-    name: 'open hands: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64F',
-    char: '🙏',
-    name: 'folded hands',
-    category: 'people',
-  },
-  {
-    codes: '1F64F 1F3FB',
-    char: '🙏🏻',
-    name: 'folded hands: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64F 1F3FC',
-    char: '🙏🏼',
-    name: 'folded hands: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64F 1F3FD',
-    char: '🙏🏽',
-    name: 'folded hands: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64F 1F3FE',
-    char: '🙏🏾',
-    name: 'folded hands: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64F 1F3FF',
-    char: '🙏🏿',
-    name: 'folded hands: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F485',
-    char: '💅',
-    name: 'nail polish',
-    category: 'people',
-  },
-  {
-    codes: '1F485 1F3FB',
-    char: '💅🏻',
-    name: 'nail polish: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F485 1F3FC',
-    char: '💅🏼',
-    name: 'nail polish: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F485 1F3FD',
-    char: '💅🏽',
-    name: 'nail polish: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F485 1F3FE',
-    char: '💅🏾',
-    name: 'nail polish: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F485 1F3FF',
-    char: '💅🏿',
-    name: 'nail polish: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F4AA',
-    char: '💪',
-    name: 'flexed biceps',
-    category: 'people',
-  },
-  {
-    codes: '1F4AA 1F3FB',
-    char: '💪🏻',
-    name: 'flexed biceps: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F4AA 1F3FC',
-    char: '💪🏼',
-    name: 'flexed biceps: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F4AA 1F3FD',
-    char: '💪🏽',
-    name: 'flexed biceps: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F4AA 1F3FE',
-    char: '💪🏾',
-    name: 'flexed biceps: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F4AA 1F3FF',
-    char: '💪🏿',
-    name: 'flexed biceps: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F442', char: '👂', name: 'ear', category: 'people' },
-  {
-    codes: '1F442 1F3FB',
-    char: '👂🏻',
-    name: 'ear: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F442 1F3FC',
-    char: '👂🏼',
-    name: 'ear: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F442 1F3FD',
-    char: '👂🏽',
-    name: 'ear: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F442 1F3FE',
-    char: '👂🏾',
-    name: 'ear: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F442 1F3FF',
-    char: '👂🏿',
-    name: 'ear: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F443', char: '👃', name: 'nose', category: 'people' },
-  {
-    codes: '1F443 1F3FB',
-    char: '👃🏻',
-    name: 'nose: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F443 1F3FC',
-    char: '👃🏼',
-    name: 'nose: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F443 1F3FD',
-    char: '👃🏽',
-    name: 'nose: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F443 1F3FE',
-    char: '👃🏾',
-    name: 'nose: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F443 1F3FF',
-    char: '👃🏿',
-    name: 'nose: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F440', char: '👀', name: 'eyes', category: 'people' },
-  { codes: '1F445', char: '👅', name: 'tongue', category: 'people' },
-  { codes: '1F444', char: '👄', name: 'mouth', category: 'people' },
-  { codes: '1F476', char: '👶', name: 'baby', category: 'people' },
-  {
-    codes: '1F476 1F3FB',
-    char: '👶🏻',
-    name: 'baby: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F476 1F3FC',
-    char: '👶🏼',
-    name: 'baby: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F476 1F3FD',
-    char: '👶🏽',
-    name: 'baby: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F476 1F3FE',
-    char: '👶🏾',
-    name: 'baby: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F476 1F3FF',
-    char: '👶🏿',
-    name: 'baby: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F9D2 1F3FB',
-    char: '🧒🏻',
-    name: 'child: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F9D2 1F3FC',
-    char: '🧒🏼',
-    name: 'child: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F9D2 1F3FD',
-    char: '🧒🏽',
-    name: 'child: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F9D2 1F3FE',
-    char: '🧒🏾',
-    name: 'child: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F9D2 1F3FF',
-    char: '🧒🏿',
-    name: 'child: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F466', char: '👦', name: 'boy', category: 'people' },
-  {
-    codes: '1F466 1F3FB',
-    char: '👦🏻',
-    name: 'boy: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F466 1F3FC',
-    char: '👦🏼',
-    name: 'boy: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F466 1F3FD',
-    char: '👦🏽',
-    name: 'boy: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F466 1F3FE',
-    char: '👦🏾',
-    name: 'boy: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F466 1F3FF',
-    char: '👦🏿',
-    name: 'boy: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F467', char: '👧', name: 'girl', category: 'people' },
-  {
-    codes: '1F467 1F3FB',
-    char: '👧🏻',
-    name: 'girl: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F467 1F3FC',
-    char: '👧🏼',
-    name: 'girl: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F467 1F3FD',
-    char: '👧🏽',
-    name: 'girl: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F467 1F3FE',
-    char: '👧🏾',
-    name: 'girl: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F467 1F3FF',
-    char: '👧🏿',
-    name: 'girl: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F471 1F3FB',
-    char: '👱🏻',
-    name: 'person: light-skin-tone, blond hair',
-    category: 'people',
-  },
-  {
-    codes: '1F471 1F3FC',
-    char: '👱🏼',
-    name: 'person: medium-light-skin-tone, blond hair',
-    category: 'people',
-  },
-  {
-    codes: '1F471 1F3FD',
-    char: '👱🏽',
-    name: 'person: medium-skin-tone, blond hair',
-    category: 'people',
-  },
-  {
-    codes: '1F471 1F3FE',
-    char: '👱🏾',
-    name: 'person: medium-dark-skin-tone, blond hair',
-    category: 'people',
-  },
-  {
-    codes: '1F471 1F3FF',
-    char: '👱🏿',
-    name: 'person: dark-skin-tone, blond hair',
-    category: 'people',
-  },
-  { codes: '1F468', char: '👨', name: 'man', category: 'people' },
-  {
-    codes: '1F468 1F3FB',
-    char: '👨🏻',
-    name: 'man: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FC',
-    char: '👨🏼',
-    name: 'man: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FD',
-    char: '👨🏽',
-    name: 'man: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FE',
-    char: '👨🏾',
-    name: 'man: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FF',
-    char: '👨🏿',
-    name: 'man: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F469', char: '👩', name: 'woman', category: 'people' },
-  {
-    codes: '1F469 1F3FB',
-    char: '👩🏻',
-    name: 'woman: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FC',
-    char: '👩🏼',
-    name: 'woman: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FD',
-    char: '👩🏽',
-    name: 'woman: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FE',
-    char: '👩🏾',
-    name: 'woman: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FF',
-    char: '👩🏿',
-    name: 'woman: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F474', char: '👴', name: 'old man', category: 'people' },
-  {
-    codes: '1F474 1F3FB',
-    char: '👴🏻',
-    name: 'old man: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F474 1F3FC',
-    char: '👴🏼',
-    name: 'old man: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F474 1F3FD',
-    char: '👴🏽',
-    name: 'old man: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F474 1F3FE',
-    char: '👴🏾',
-    name: 'old man: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F474 1F3FF',
-    char: '👴🏿',
-    name: 'old man: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F475', char: '👵', name: 'old woman', category: 'people' },
-  {
-    codes: '1F475 1F3FB',
-    char: '👵🏻',
-    name: 'old woman: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F475 1F3FC',
-    char: '👵🏼',
-    name: 'old woman: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F475 1F3FD',
-    char: '👵🏽',
-    name: 'old woman: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F475 1F3FE',
-    char: '👵🏾',
-    name: 'old woman: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F475 1F3FF',
-    char: '👵🏿',
-    name: 'old woman: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64D',
-    char: '🙍',
-    name: 'person frowning',
-    category: 'people',
-  },
-  {
-    codes: '1F64D 1F3FB',
-    char: '🙍🏻',
-    name: 'person frowning: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64D 1F3FC',
-    char: '🙍🏼',
-    name: 'person frowning: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64D 1F3FD',
-    char: '🙍🏽',
-    name: 'person frowning: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64D 1F3FE',
-    char: '🙍🏾',
-    name: 'person frowning: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64D 1F3FF',
-    char: '🙍🏿',
-    name: 'person frowning: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64E',
-    char: '🙎',
-    name: 'person pouting',
-    category: 'people',
-  },
-  {
-    codes: '1F64E 1F3FB',
-    char: '🙎🏻',
-    name: 'person pouting: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64E 1F3FC',
-    char: '🙎🏼',
-    name: 'person pouting: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64E 1F3FD',
-    char: '🙎🏽',
-    name: 'person pouting: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64E 1F3FE',
-    char: '🙎🏾',
-    name: 'person pouting: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64E 1F3FF',
-    char: '🙎🏿',
-    name: 'person pouting: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F645',
-    char: '🙅',
-    name: 'person gesturing NO',
-    category: 'people',
-  },
-  {
-    codes: '1F645 1F3FB',
-    char: '🙅🏻',
-    name: 'person gesturing NO: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F645 1F3FC',
-    char: '🙅🏼',
-    name: 'person gesturing NO: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F645 1F3FD',
-    char: '🙅🏽',
-    name: 'person gesturing NO: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F645 1F3FE',
-    char: '🙅🏾',
-    name: 'person gesturing NO: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F645 1F3FF',
-    char: '🙅🏿',
-    name: 'person gesturing NO: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F646',
-    char: '🙆',
-    name: 'person gesturing OK',
-    category: 'people',
-  },
-  {
-    codes: '1F646 1F3FB',
-    char: '🙆🏻',
-    name: 'person gesturing OK: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F646 1F3FC',
-    char: '🙆🏼',
-    name: 'person gesturing OK: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F646 1F3FD',
-    char: '🙆🏽',
-    name: 'person gesturing OK: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F646 1F3FE',
-    char: '🙆🏾',
-    name: 'person gesturing OK: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F646 1F3FF',
-    char: '🙆🏿',
-    name: 'person gesturing OK: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F481',
-    char: '💁',
-    name: 'person tipping hand',
-    category: 'people',
-  },
-  {
-    codes: '1F481 1F3FB',
-    char: '💁🏻',
-    name: 'person tipping hand: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F481 1F3FC',
-    char: '💁🏼',
-    name: 'person tipping hand: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F481 1F3FD',
-    char: '💁🏽',
-    name: 'person tipping hand: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F481 1F3FE',
-    char: '💁🏾',
-    name: 'person tipping hand: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F481 1F3FF',
-    char: '💁🏿',
-    name: 'person tipping hand: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64B',
-    char: '🙋',
-    name: 'person raising hand',
-    category: 'people',
-  },
-  {
-    codes: '1F64B 1F3FB',
-    char: '🙋🏻',
-    name: 'person raising hand: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64B 1F3FC',
-    char: '🙋🏼',
-    name: 'person raising hand: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64B 1F3FD',
-    char: '🙋🏽',
-    name: 'person raising hand: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64B 1F3FE',
-    char: '🙋🏾',
-    name: 'person raising hand: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F64B 1F3FF',
-    char: '🙋🏿',
-    name: 'person raising hand: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F647',
-    char: '🙇',
-    name: 'person bowing',
-    category: 'people',
-  },
-  {
-    codes: '1F647 1F3FB',
-    char: '🙇🏻',
-    name: 'person bowing: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F647 1F3FC',
-    char: '🙇🏼',
-    name: 'person bowing: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F647 1F3FD',
-    char: '🙇🏽',
-    name: 'person bowing: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F647 1F3FE',
-    char: '🙇🏾',
-    name: 'person bowing: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F647 1F3FF',
-    char: '🙇🏿',
-    name: 'person bowing: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46E',
-    char: '👮',
-    name: 'police officer',
-    category: 'people',
-  },
-  {
-    codes: '1F46E 1F3FB',
-    char: '👮🏻',
-    name: 'police officer: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46E 1F3FC',
-    char: '👮🏼',
-    name: 'police officer: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46E 1F3FD',
-    char: '👮🏽',
-    name: 'police officer: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46E 1F3FE',
-    char: '👮🏾',
-    name: 'police officer: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46E 1F3FF',
-    char: '👮🏿',
-    name: 'police officer: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 200D 1F4BB',
-    char: '👨‍💻',
-    name: 'man technologist',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FB 200D 1F4BB',
-    char: '👨🏻‍💻',
-    name: 'man technologist: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FC 200D 1F4BB',
-    char: '👨🏼‍💻',
-    name: 'man technologist: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FD 200D 1F4BB',
-    char: '👨🏽‍💻',
-    name: 'man technologist: medium skin tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FE 200D 1F4BB',
-    char: '👨🏾‍💻',
-    name: 'man technologist: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FF 200D 1F4BB',
-    char: '👨🏿‍💻',
-    name: 'man technologist: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 200D 1F4BB',
-    char: '👩‍💻',
-    name: 'woman technologist',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FB 200D 1F4BB',
-    char: '👩🏻‍💻',
-    name: 'woman technologist: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FC 200D 1F4BB',
-    char: '👩🏼‍💻',
-    name: 'woman technologist: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FD 200D 1F4BB',
-    char: '👩🏽‍💻',
-    name: 'woman technologist: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FE 200D 1F4BB',
-    char: '👩🏾‍💻',
-    name: 'woman technologist: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FF 200D 1F4BB',
-    char: '👩🏿‍💻',
-    name: 'woman technologist: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F482', char: '💂', name: 'guard', category: 'people' },
-  {
-    codes: '1F482 1F3FB',
-    char: '💂🏻',
-    name: 'guard: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F482 1F3FC',
-    char: '💂🏼',
-    name: 'guard: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F482 1F3FD',
-    char: '💂🏽',
-    name: 'guard: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F482 1F3FE',
-    char: '💂🏾',
-    name: 'guard: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F482 1F3FF',
-    char: '💂🏿',
-    name: 'guard: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F477',
-    char: '👷',
-    name: 'construction worker',
-    category: 'people',
-  },
-  {
-    codes: '1F477 1F3FB',
-    char: '👷🏻',
-    name: 'construction worker: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F477 1F3FC',
-    char: '👷🏼',
-    name: 'construction worker: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F477 1F3FD',
-    char: '👷🏽',
-    name: 'construction worker: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F477 1F3FE',
-    char: '👷🏾',
-    name: 'construction worker: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F477 1F3FF',
-    char: '👷🏿',
-    name: 'construction worker: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F478', char: '👸', name: 'princess', category: 'people' },
-  {
-    codes: '1F478 1F3FB',
-    char: '👸🏻',
-    name: 'princess: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F478 1F3FC',
-    char: '👸🏼',
-    name: 'princess: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F478 1F3FD',
-    char: '👸🏽',
-    name: 'princess: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F478 1F3FE',
-    char: '👸🏾',
-    name: 'princess: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F478 1F3FF',
-    char: '👸🏿',
-    name: 'princess: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F473',
-    char: '👳',
-    name: 'person wearing turban',
-    category: 'people',
-  },
-  {
-    codes: '1F473 1F3FB',
-    char: '👳🏻',
-    name: 'person wearing turban: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F473 1F3FC',
-    char: '👳🏼',
-    name: 'person wearing turban: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F473 1F3FD',
-    char: '👳🏽',
-    name: 'person wearing turban: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F473 1F3FE',
-    char: '👳🏾',
-    name: 'person wearing turban: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F473 1F3FF',
-    char: '👳🏿',
-    name: 'person wearing turban: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F472',
-    char: '👲',
-    name: 'man with Chinese cap',
-    category: 'people',
-  },
-  {
-    codes: '1F472 1F3FB',
-    char: '👲🏻',
-    name: 'man with Chinese cap: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F472 1F3FC',
-    char: '👲🏼',
-    name: 'man with Chinese cap: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F472 1F3FD',
-    char: '👲🏽',
-    name: 'man with Chinese cap: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F472 1F3FE',
-    char: '👲🏾',
-    name: 'man with Chinese cap: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F472 1F3FF',
-    char: '👲🏿',
-    name: 'man with Chinese cap: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F470',
-    char: '👰',
-    name: 'bride with veil',
-    category: 'people',
-  },
-  {
-    codes: '1F470 1F3FB',
-    char: '👰🏻',
-    name: 'bride with veil: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F470 1F3FC',
-    char: '👰🏼',
-    name: 'bride with veil: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F470 1F3FD',
-    char: '👰🏽',
-    name: 'bride with veil: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F470 1F3FE',
-    char: '👰🏾',
-    name: 'bride with veil: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F470 1F3FF',
-    char: '👰🏿',
-    name: 'bride with veil: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F47C',
-    char: '👼',
-    name: 'baby angel',
-    category: 'people',
-  },
-  {
-    codes: '1F47C 1F3FB',
-    char: '👼🏻',
-    name: 'baby angel: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F47C 1F3FC',
-    char: '👼🏼',
-    name: 'baby angel: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F47C 1F3FD',
-    char: '👼🏽',
-    name: 'baby angel: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F47C 1F3FE',
-    char: '👼🏾',
-    name: 'baby angel: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F47C 1F3FF',
-    char: '👼🏿',
-    name: 'baby angel: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F385',
-    char: '🎅',
-    name: 'Santa Claus',
-    category: 'people',
-  },
-  {
-    codes: '1F385 1F3FB',
-    char: '🎅🏻',
-    name: 'Santa Claus: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F385 1F3FC',
-    char: '🎅🏼',
-    name: 'Santa Claus: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F385 1F3FD',
-    char: '🎅🏽',
-    name: 'Santa Claus: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F385 1F3FE',
-    char: '🎅🏾',
-    name: 'Santa Claus: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F385 1F3FF',
-    char: '🎅🏿',
-    name: 'Santa Claus: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F486',
-    char: '💆',
-    name: 'person getting massage',
-    category: 'people',
-  },
-  {
-    codes: '1F486 1F3FB',
-    char: '💆🏻',
-    name: 'person getting massage: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F486 1F3FC',
-    char: '💆🏼',
-    name: 'person getting massage: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F486 1F3FD',
-    char: '💆🏽',
-    name: 'person getting massage: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F486 1F3FE',
-    char: '💆🏾',
-    name: 'person getting massage: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F486 1F3FF',
-    char: '💆🏿',
-    name: 'person getting massage: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F487',
-    char: '💇',
-    name: 'person getting haircut',
-    category: 'people',
-  },
-  {
-    codes: '1F487 1F3FB',
-    char: '💇🏻',
-    name: 'person getting haircut: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F487 1F3FC',
-    char: '💇🏼',
-    name: 'person getting haircut: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F487 1F3FD',
-    char: '💇🏽',
-    name: 'person getting haircut: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F487 1F3FE',
-    char: '💇🏾',
-    name: 'person getting haircut: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F487 1F3FF',
-    char: '💇🏿',
-    name: 'person getting haircut: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C3',
-    char: '🏃',
-    name: 'person running',
-    category: 'people',
-  },
-  {
-    codes: '1F3C3 1F3FB',
-    char: '🏃🏻',
-    name: 'person running: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C3 1F3FC',
-    char: '🏃🏼',
-    name: 'person running: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C3 1F3FD',
-    char: '🏃🏽',
-    name: 'person running: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C3 1F3FE',
-    char: '🏃🏾',
-    name: 'person running: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C3 1F3FF',
-    char: '🏃🏿',
-    name: 'person running: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F483',
-    char: '💃',
-    name: 'woman dancing',
-    category: 'people',
-  },
-  {
-    codes: '1F483 1F3FB',
-    char: '💃🏻',
-    name: 'woman dancing: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F483 1F3FC',
-    char: '💃🏼',
-    name: 'woman dancing: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F483 1F3FD',
-    char: '💃🏽',
-    name: 'woman dancing: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F483 1F3FE',
-    char: '💃🏾',
-    name: 'woman dancing: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F483 1F3FF',
-    char: '💃🏿',
-    name: 'woman dancing: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46F',
-    char: '👯',
-    name: 'people with bunny ears',
-    category: 'people',
-  },
-  {
-    codes: '1F3C2',
-    char: '🏂',
-    name: 'snowboarder',
-    category: 'people',
-  },
-  {
-    codes: '1F3C2 1F3FB',
-    char: '🏂🏻',
-    name: 'snowboarder: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C2 1F3FC',
-    char: '🏂🏼',
-    name: 'snowboarder: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C2 1F3FD',
-    char: '🏂🏽',
-    name: 'snowboarder: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C2 1F3FE',
-    char: '🏂🏾',
-    name: 'snowboarder: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C2 1F3FF',
-    char: '🏂🏿',
-    name: 'snowboarder: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C4',
-    char: '🏄',
-    name: 'person surfing',
-    category: 'people',
-  },
-  {
-    codes: '1F3C4 1F3FB',
-    char: '🏄🏻',
-    name: 'person surfing: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C4 1F3FC',
-    char: '🏄🏼',
-    name: 'person surfing: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C4 1F3FD',
-    char: '🏄🏽',
-    name: 'person surfing: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C4 1F3FE',
-    char: '🏄🏾',
-    name: 'person surfing: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3C4 1F3FF',
-    char: '🏄🏿',
-    name: 'person surfing: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3CA',
-    char: '🏊',
-    name: 'person swimming',
-    category: 'people',
-  },
-  {
-    codes: '1F3CA 1F3FB',
-    char: '🏊🏻',
-    name: 'person swimming: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3CA 1F3FC',
-    char: '🏊🏼',
-    name: 'person swimming: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3CA 1F3FD',
-    char: '🏊🏽',
-    name: 'person swimming: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3CA 1F3FE',
-    char: '🏊🏾',
-    name: 'person swimming: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F3CA 1F3FF',
-    char: '🏊🏿',
-    name: 'person swimming: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F6B4',
-    char: '🚴',
-    name: 'person biking',
-    category: 'people',
-  },
-  {
-    codes: '1F6B4 1F3FB',
-    char: '🚴🏻',
-    name: 'person biking: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F6B4 1F3FC',
-    char: '🚴🏼',
-    name: 'person biking: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F6B4 1F3FD',
-    char: '🚴🏽',
-    name: 'person biking: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F6B4 1F3FE',
-    char: '🚴🏾',
-    name: 'person biking: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F6B4 1F3FF',
-    char: '🚴🏿',
-    name: 'person biking: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F6C0',
-    char: '🛀',
-    name: 'person taking bath',
-    category: 'people',
-  },
-  {
-    codes: '1F6C0 1F3FB',
-    char: '🛀🏻',
-    name: 'person taking bath: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F6C0 1F3FC',
-    char: '🛀🏼',
-    name: 'person taking bath: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F6C0 1F3FD',
-    char: '🛀🏽',
-    name: 'person taking bath: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F6C0 1F3FE',
-    char: '🛀🏾',
-    name: 'person taking bath: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F6C0 1F3FF',
-    char: '🛀🏿',
-    name: 'person taking bath: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46D',
-    char: '👭',
-    name: 'women holding hands',
-    category: 'people',
-  },
-  {
-    codes: '1F46D 1F3FB',
-    char: '👭🏻',
-    name: 'women holding hands: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FC 200D 1F91D 200D 1F469 1F3FB',
-    char: '👩🏼‍🤝‍👩🏻',
-    name: 'women holding hands: medium-light-skin-tone, light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46D 1F3FC',
-    char: '👭🏼',
-    name: 'women holding hands: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FD 200D 1F91D 200D 1F469 1F3FB',
-    char: '👩🏽‍🤝‍👩🏻',
-    name: 'women holding hands: medium-skin-tone, light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FD 200D 1F91D 200D 1F469 1F3FC',
-    char: '👩🏽‍🤝‍👩🏼',
-    name: 'women holding hands: medium-skin-tone, medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46D 1F3FD',
-    char: '👭🏽',
-    name: 'women holding hands: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FE 200D 1F91D 200D 1F469 1F3FB',
-    char: '👩🏾‍🤝‍👩🏻',
-    name: 'women holding hands: medium-dark-skin-tone, light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FE 200D 1F91D 200D 1F469 1F3FC',
-    char: '👩🏾‍🤝‍👩🏼',
-    name: 'women holding hands: medium-dark-skin-tone, medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FE 200D 1F91D 200D 1F469 1F3FD',
-    char: '👩🏾‍🤝‍👩🏽',
-    name: 'women holding hands: medium-dark-skin-tone, medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46D 1F3FE',
-    char: '👭🏾',
-    name: 'women holding hands: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FF 200D 1F91D 200D 1F469 1F3FB',
-    char: '👩🏿‍🤝‍👩🏻',
-    name: 'women holding hands: dark-skin-tone, light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FF 200D 1F91D 200D 1F469 1F3FC',
-    char: '👩🏿‍🤝‍👩🏼',
-    name: 'women holding hands: dark-skin-tone, medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FF 200D 1F91D 200D 1F469 1F3FD',
-    char: '👩🏿‍🤝‍👩🏽',
-    name: 'women holding hands: dark-skin-tone, medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FF 200D 1F91D 200D 1F469 1F3FE',
-    char: '👩🏿‍🤝‍👩🏾',
-    name: 'women holding hands: dark-skin-tone, medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46D 1F3FF',
-    char: '👭🏿',
-    name: 'women holding hands: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46B',
-    char: '👫',
-    name: 'woman and man holding hands',
-    category: 'people',
-  },
-  {
-    codes: '1F46B 1F3FB',
-    char: '👫🏻',
-    name: 'woman and man holding hands: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FB 200D 1F91D 200D 1F468 1F3FC',
-    char: '👩🏻‍🤝‍👨🏼',
-    name:
-      'woman and man holding hands: light-skin-tone, medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FB 200D 1F91D 200D 1F468 1F3FD',
-    char: '👩🏻‍🤝‍👨🏽',
-    name: 'woman and man holding hands: light-skin-tone, medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FB 200D 1F91D 200D 1F468 1F3FE',
-    char: '👩🏻‍🤝‍👨🏾',
-    name: 'woman and man holding hands: light-skin-tone, medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FB 200D 1F91D 200D 1F468 1F3FF',
-    char: '👩🏻‍🤝‍👨🏿',
-    name: 'woman and man holding hands: light-skin-tone, dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FC 200D 1F91D 200D 1F468 1F3FB',
-    char: '👩🏼‍🤝‍👨🏻',
-    name:
-      'woman and man holding hands: medium-light-skin-tone, light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46B 1F3FC',
-    char: '👫🏼',
-    name: 'woman and man holding hands: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FC 200D 1F91D 200D 1F468 1F3FD',
-    char: '👩🏼‍🤝‍👨🏽',
-    name:
-      'woman and man holding hands: medium-light-skin-tone, medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FC 200D 1F91D 200D 1F468 1F3FE',
-    char: '👩🏼‍🤝‍👨🏾',
-    name:
-      'woman and man holding hands: medium-light-skin-tone, medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FC 200D 1F91D 200D 1F468 1F3FF',
-    char: '👩🏼‍🤝‍👨🏿',
-    name: 'woman and man holding hands: medium-light-skin-tone, dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FD 200D 1F91D 200D 1F468 1F3FB',
-    char: '👩🏽‍🤝‍👨🏻',
-    name: 'woman and man holding hands: medium-skin-tone, light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FD 200D 1F91D 200D 1F468 1F3FC',
-    char: '👩🏽‍🤝‍👨🏼',
-    name:
-      'woman and man holding hands: medium-skin-tone, medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46B 1F3FD',
-    char: '👫🏽',
-    name: 'woman and man holding hands: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FD 200D 1F91D 200D 1F468 1F3FE',
-    char: '👩🏽‍🤝‍👨🏾',
-    name:
-      'woman and man holding hands: medium-skin-tone, medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FD 200D 1F91D 200D 1F468 1F3FF',
-    char: '👩🏽‍🤝‍👨🏿',
-    name: 'woman and man holding hands: medium-skin-tone, dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FE 200D 1F91D 200D 1F468 1F3FB',
-    char: '👩🏾‍🤝‍👨🏻',
-    name: 'woman and man holding hands: medium-dark-skin-tone, light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FE 200D 1F91D 200D 1F468 1F3FC',
-    char: '👩🏾‍🤝‍👨🏼',
-    name:
-      'woman and man holding hands: medium-dark-skin-tone, medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FE 200D 1F91D 200D 1F468 1F3FD',
-    char: '👩🏾‍🤝‍👨🏽',
-    name:
-      'woman and man holding hands: medium-dark-skin-tone, medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46B 1F3FE',
-    char: '👫🏾',
-    name: 'woman and man holding hands: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FE 200D 1F91D 200D 1F468 1F3FF',
-    char: '👩🏾‍🤝‍👨🏿',
-    name: 'woman and man holding hands: medium-dark-skin-tone, dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FF 200D 1F91D 200D 1F468 1F3FB',
-    char: '👩🏿‍🤝‍👨🏻',
-    name: 'woman and man holding hands: dark-skin-tone, light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FF 200D 1F91D 200D 1F468 1F3FC',
-    char: '👩🏿‍🤝‍👨🏼',
-    name: 'woman and man holding hands: dark-skin-tone, medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FF 200D 1F91D 200D 1F468 1F3FD',
-    char: '👩🏿‍🤝‍👨🏽',
-    name: 'woman and man holding hands: dark-skin-tone, medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F469 1F3FF 200D 1F91D 200D 1F468 1F3FE',
-    char: '👩🏿‍🤝‍👨🏾',
-    name: 'woman and man holding hands: dark-skin-tone, medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46B 1F3FF',
-    char: '👫🏿',
-    name: 'woman and man holding hands: dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46C',
-    char: '👬',
-    name: 'men holding hands',
-    category: 'people',
-  },
-  {
-    codes: '1F46C 1F3FB',
-    char: '👬🏻',
-    name: 'men holding hands: light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FC 200D 1F91D 200D 1F468 1F3FB',
-    char: '👨🏼‍🤝‍👨🏻',
-    name: 'men holding hands: medium-light-skin-tone, light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46C 1F3FC',
-    char: '👬🏼',
-    name: 'men holding hands: medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FD 200D 1F91D 200D 1F468 1F3FB',
-    char: '👨🏽‍🤝‍👨🏻',
-    name: 'men holding hands: medium-skin-tone, light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FD 200D 1F91D 200D 1F468 1F3FC',
-    char: '👨🏽‍🤝‍👨🏼',
-    name: 'men holding hands: medium-skin-tone, medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46C 1F3FD',
-    char: '👬🏽',
-    name: 'men holding hands: medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FE 200D 1F91D 200D 1F468 1F3FB',
-    char: '👨🏾‍🤝‍👨🏻',
-    name: 'men holding hands: medium-dark-skin-tone, light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FE 200D 1F91D 200D 1F468 1F3FC',
-    char: '👨🏾‍🤝‍👨🏼',
-    name: 'men holding hands: medium-dark-skin-tone, medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FE 200D 1F91D 200D 1F468 1F3FD',
-    char: '👨🏾‍🤝‍👨🏽',
-    name: 'men holding hands: medium-dark-skin-tone, medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46C 1F3FE',
-    char: '👬🏾',
-    name: 'men holding hands: medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FF 200D 1F91D 200D 1F468 1F3FB',
-    char: '👨🏿‍🤝‍👨🏻',
-    name: 'men holding hands: dark-skin-tone, light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FF 200D 1F91D 200D 1F468 1F3FC',
-    char: '👨🏿‍🤝‍👨🏼',
-    name: 'men holding hands: dark-skin-tone, medium-light-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FF 200D 1F91D 200D 1F468 1F3FD',
-    char: '👨🏿‍🤝‍👨🏽',
-    name: 'men holding hands: dark-skin-tone, medium-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F468 1F3FF 200D 1F91D 200D 1F468 1F3FE',
-    char: '👨🏿‍🤝‍👨🏾',
-    name: 'men holding hands: dark-skin-tone, medium-dark-skin-tone',
-    category: 'people',
-  },
-  {
-    codes: '1F46C 1F3FF',
-    char: '👬🏿',
-    name: 'men holding hands: dark-skin-tone',
-    category: 'people',
-  },
-  { codes: '1F48F', char: '💏', name: 'kiss', category: 'people' },
-  { codes: '1F46A', char: '👪', name: 'family', category: 'people' },
-  {
-    codes: '1F464',
-    char: '👤',
-    name: 'bust in silhouette',
-    category: 'people',
-  },
-  {
-    codes: '1F465',
-    char: '👥',
-    name: 'busts in silhouette',
-    category: 'people',
-  },
-  {
-    codes: '1F463',
-    char: '👣',
-    name: 'footprints',
-    category: 'people',
-  },
-  {
-    codes: '1F3FB',
-    char: '🏻',
-    name: 'light-skin-tone',
-    category: 'skin-tone',
-  },
-  {
-    codes: '1F3FC',
-    char: '🏼',
-    name: 'medium-light-skin-tone',
-    category: 'skin-tone',
-  },
-  {
-    codes: '1F3FD',
-    char: '🏽',
-    name: 'medium-skin-tone',
-    category: 'skin-tone',
-  },
-  {
-    codes: '1F3FE',
-    char: '🏾',
-    name: 'medium-dark-skin-tone',
-    category: 'skin-tone',
-  },
-  {
-    codes: '1F3FF',
-    char: '🏿',
-    name: 'dark-skin-tone',
-    category: 'skin-tone',
-  },
-  {
-    codes: '1F435',
-    char: '🐵',
-    name: 'monkey face',
-    category: 'nature',
-  },
-  { codes: '1F412', char: '🐒', name: 'monkey', category: 'nature' },
-  { codes: '1F436', char: '🐶', name: 'dog face', category: 'nature' },
-  { codes: '1F415', char: '🐕', name: 'dog', category: 'nature' },
-  { codes: '1F429', char: '🐩', name: 'poodle', category: 'nature' },
-  { codes: '1F43A', char: '🐺', name: 'wolf', category: 'nature' },
-  { codes: '1F431', char: '🐱', name: 'cat face', category: 'nature' },
-  { codes: '1F408', char: '🐈', name: 'cat', category: 'nature' },
-  {
-    codes: '1F42F',
-    char: '🐯',
-    name: 'tiger face',
-    category: 'nature',
-  },
-  { codes: '1F405', char: '🐅', name: 'tiger', category: 'nature' },
-  { codes: '1F406', char: '🐆', name: 'leopard', category: 'nature' },
-  {
-    codes: '1F434',
-    char: '🐴',
-    name: 'horse face',
-    category: 'nature',
-  },
-  { codes: '1F40E', char: '🐎', name: 'horse', category: 'nature' },
-  { codes: '1F42E', char: '🐮', name: 'cow face', category: 'nature' },
-  { codes: '1F402', char: '🐂', name: 'ox', category: 'nature' },
-  {
-    codes: '1F403',
-    char: '🐃',
-    name: 'water buffalo',
-    category: 'nature',
-  },
-  { codes: '1F404', char: '🐄', name: 'cow', category: 'nature' },
-  { codes: '1F437', char: '🐷', name: 'pig face', category: 'nature' },
-  { codes: '1F416', char: '🐖', name: 'pig', category: 'nature' },
-  { codes: '1F417', char: '🐗', name: 'boar', category: 'nature' },
-  { codes: '1F43D', char: '🐽', name: 'pig nose', category: 'nature' },
-  { codes: '1F40F', char: '🐏', name: 'ram', category: 'nature' },
-  { codes: '1F411', char: '🐑', name: 'ewe', category: 'nature' },
-  { codes: '1F410', char: '🐐', name: 'goat', category: 'nature' },
-  { codes: '1F42A', char: '🐪', name: 'camel', category: 'nature' },
-  { codes: '1F418', char: '🐘', name: 'elephant', category: 'nature' },
-  {
-    codes: '1F42D',
-    char: '🐭',
-    name: 'mouse face',
-    category: 'nature',
-  },
-  { codes: '1F401', char: '🐁', name: 'mouse', category: 'nature' },
-  { codes: '1F439', char: '🐹', name: 'hamster', category: 'nature' },
-  {
-    codes: '1F430',
-    char: '🐰',
-    name: 'rabbit face',
-    category: 'nature',
-  },
-  { codes: '1F407', char: '🐇', name: 'rabbit', category: 'nature' },
-  { codes: '1F43B', char: '🐻', name: 'bear', category: 'nature' },
-  { codes: '1F428', char: '🐨', name: 'koala', category: 'nature' },
-  { codes: '1F43C', char: '🐼', name: 'panda', category: 'nature' },
-  {
-    codes: '1F43E',
-    char: '🐾',
-    name: 'paw prints',
-    category: 'nature',
-  },
-  { codes: '1F414', char: '🐔', name: 'chicken', category: 'nature' },
-  { codes: '1F413', char: '🐓', name: 'rooster', category: 'nature' },
-  {
-    codes: '1F423',
-    char: '🐣',
-    name: 'hatching chick',
-    category: 'nature',
-  },
-  {
-    codes: '1F424',
-    char: '🐤',
-    name: 'baby chick',
-    category: 'nature',
-  },
-  {
-    codes: '1F425',
-    char: '🐥',
-    name: 'front-facing baby chick',
-    category: 'nature',
-  },
-  { codes: '1F426', char: '🐦', name: 'bird', category: 'nature' },
-  { codes: '1F427', char: '🐧', name: 'penguin', category: 'nature' },
-  { codes: '1F438', char: '🐸', name: 'frog', category: 'nature' },
-  { codes: '1F40A', char: '🐊', name: 'crocodile', category: 'nature' },
-  { codes: '1F422', char: '🐢', name: 'turtle', category: 'nature' },
-  { codes: '1F40D', char: '🐍', name: 'snake', category: 'nature' },
-  {
-    codes: '1F432',
-    char: '🐲',
-    name: 'dragon face',
-    category: 'nature',
-  },
-  { codes: '1F409', char: '🐉', name: 'dragon', category: 'nature' },
-  {
-    codes: '1F433',
-    char: '🐳',
-    name: 'spouting whale',
-    category: 'nature',
-  },
-  { codes: '1F40B', char: '🐋', name: 'whale', category: 'nature' },
-  { codes: '1F42C', char: '🐬', name: 'dolphin', category: 'nature' },
-  { codes: '1F41F', char: '🐟', name: 'fish', category: 'nature' },
-  {
-    codes: '1F420',
-    char: '🐠',
-    name: 'tropical fish',
-    category: 'nature',
-  },
-  { codes: '1F421', char: '🐡', name: 'blowfish', category: 'nature' },
-  { codes: '1F988', char: '🦈', name: 'shark', category: 'nature' },
-  { codes: '1F419', char: '🐙', name: 'octopus', category: 'nature' },
-  { codes: '1F40C', char: '🐌', name: 'snail', category: 'nature' },
-  { codes: '1F98B', char: '🦋', name: 'butterfly', category: 'nature' },
-  { codes: '1F41B', char: '🐛', name: 'bug', category: 'nature' },
-  { codes: '1F41C', char: '🐜', name: 'ant', category: 'nature' },
-  { codes: '1F41D', char: '🐝', name: 'honeybee', category: 'nature' },
-  {
-    codes: '1F41E',
-    char: '🐞',
-    name: 'lady beetle',
-    category: 'nature',
-  },
-
-  { codes: '1F490', char: '💐', name: 'bouquet', category: 'nature' },
-  {
-    codes: '1F338',
-    char: '🌸',
-    name: 'cherry blossom',
-    category: 'nature',
-  },
-  {
-    codes: '1F4AE',
-    char: '💮',
-    name: 'white flower',
-    category: 'nature',
-  },
-  { codes: '1F339', char: '🌹', name: 'rose', category: 'nature' },
-  { codes: '1F33A', char: '🌺', name: 'hibiscus', category: 'nature' },
-  { codes: '1F33B', char: '🌻', name: 'sunflower', category: 'nature' },
-  { codes: '1F33C', char: '🌼', name: 'blossom', category: 'nature' },
-  { codes: '1F337', char: '🌷', name: 'tulip', category: 'nature' },
-  { codes: '1F331', char: '🌱', name: 'seedling', category: 'nature' },
-  {
-    codes: '1F332',
-    char: '🌲',
-    name: 'evergreen tree',
-    category: 'nature',
-  },
-  {
-    codes: '1F333',
-    char: '🌳',
-    name: 'deciduous tree',
-    category: 'nature',
-  },
-  { codes: '1F334', char: '🌴', name: 'palm tree', category: 'nature' },
-  { codes: '1F335', char: '🌵', name: 'cactus', category: 'nature' },
-  {
-    codes: '1F33E',
-    char: '🌾',
-    name: 'sheaf of rice',
-    category: 'nature',
-  },
-  { codes: '1F33F', char: '🌿', name: 'herb', category: 'nature' },
-  {
-    codes: '2618 FE0F',
-    char: '☘️',
-    name: 'shamrock',
-    category: 'nature',
-  },
-  {
-    codes: '1F340',
-    char: '🍀',
-    name: 'four leaf clover',
-    category: 'nature',
-  },
-  {
-    codes: '1F341',
-    char: '🍁',
-    name: 'maple leaf',
-    category: 'nature',
-  },
-  {
-    codes: '1F342',
-    char: '🍂',
-    name: 'fallen leaf',
-    category: 'nature',
-  },
-  {
-    codes: '1F343',
-    char: '🍃',
-    name: 'leaf fluttering in wind',
-    category: 'nature',
-  },
-  { codes: '1F311', char: '🌑', name: 'new moon', category: 'nature' },
-  {
-    codes: '1F312',
-    char: '🌒',
-    name: 'waxing crescent moon',
-    category: 'nature',
-  },
-  {
-    codes: '1F313',
-    char: '🌓',
-    name: 'first quarter moon',
-    category: 'nature',
-  },
-  {
-    codes: '1F314',
-    char: '🌔',
-    name: 'waxing gibbous moon',
-    category: 'nature',
-  },
-  { codes: '1F315', char: '🌕', name: 'full moon', category: 'nature' },
-  {
-    codes: '1F316',
-    char: '🌖',
-    name: 'waning gibbous moon',
-    category: 'nature',
-  },
-  {
-    codes: '1F317',
-    char: '🌗',
-    name: 'last quarter moon',
-    category: 'nature',
-  },
-  {
-    codes: '1F318',
-    char: '🌘',
-    name: 'waning crescent moon',
-    category: 'nature',
-  },
-  {
-    codes: '1F319',
-    char: '🌙',
-    name: 'crescent moon',
-    category: 'nature',
-  },
-  {
-    codes: '1F31A',
-    char: '🌚',
-    name: 'new moon face',
-    category: 'nature',
-  },
-  {
-    codes: '1F31B',
-    char: '🌛',
-    name: 'first quarter moon face',
-    category: 'nature',
-  },
-  {
-    codes: '1F31C',
-    char: '🌜',
-    name: 'last quarter moon face',
-    category: 'nature',
-  },
-  {
-    codes: '1F31D',
-    char: '🌝',
-    name: 'full moon face',
-    category: 'nature',
-  },
-  {
-    codes: '1F31E',
-    char: '🌞',
-    name: 'sun with face',
-    category: 'nature',
-  },
-  { codes: '2600 FE0F', char: '☀️', name: 'sun', category: 'nature' },
-  { codes: '2B50', char: '⭐', name: 'star', category: 'nature' },
-  {
-    codes: '1F31F',
-    char: '🌟',
-    name: 'glowing star',
-    category: 'nature',
-  },
-  {
-    codes: '1F320',
-    char: '🌠',
-    name: 'shooting star',
-    category: 'nature',
-  },
-  { codes: '1F30C', char: '🌌', name: 'milky way', category: 'nature' },
-  { codes: '2601 FE0F', char: '☁️', name: 'cloud', category: 'nature' },
-  {
-    codes: '26C5',
-    char: '⛅',
-    name: 'sun behind cloud',
-    category: 'nature',
-  },
-  { codes: '1F300', char: '🌀', name: 'cyclone', category: 'nature' },
-  { codes: '1F308', char: '🌈', name: 'rainbow', category: 'nature' },
-  {
-    codes: '1F302',
-    char: '🌂',
-    name: 'closed umbrella',
-    category: 'nature',
-  },
-  {
-    codes: '2614',
-    char: '☔',
-    name: 'umbrella with rain drops',
-    category: 'nature',
-  },
-  {
-    codes: '26A1',
-    char: '⚡',
-    name: 'high voltage',
-    category: 'nature',
-  },
-  {
-    codes: '2744 FE0F',
-    char: '❄️',
-    name: 'snowflake',
-    category: 'nature',
-  },
-  {
-    codes: '26C4',
-    char: '⛄',
-    name: 'snowman without snow',
-    category: 'nature',
-  },
-  { codes: '1F525', char: '🔥', name: 'fire', category: 'nature' },
-  { codes: '1F4A7', char: '💧', name: 'droplet', category: 'nature' },
-  {
-    codes: '1F30A',
-    char: '🌊',
-    name: 'water wave',
-    category: 'nature',
-  },
-  {
-    codes: '1F347',
-    char: '🍇',
-    name: 'grapes',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F348',
-    char: '🍈',
-    name: 'melon',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F349',
-    char: '🍉',
-    name: 'watermelon',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F34A',
-    char: '🍊',
-    name: 'tangerine',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F34B',
-    char: '🍋',
-    name: 'lemon',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F34C',
-    char: '🍌',
-    name: 'banana',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F34D',
-    char: '🍍',
-    name: 'pineapple',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F34E',
-    char: '🍎',
-    name: 'red apple',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F34F',
-    char: '🍏',
-    name: 'green apple',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F350',
-    char: '🍐',
-    name: 'pear',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F351',
-    char: '🍑',
-    name: 'peach',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F352',
-    char: '🍒',
-    name: 'cherries',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F353',
-    char: '🍓',
-    name: 'strawberry',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F345',
-    char: '🍅',
-    name: 'tomato',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F346',
-    char: '🍆',
-    name: 'eggplant',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F33D',
-    char: '🌽',
-    name: 'ear of corn',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F965',
-    char: '🥥',
-    name: 'coconut',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F951',
-    char: '🥑',
-    name: 'avocado',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F954',
-    char: '🥔',
-    name: 'potato',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F955',
-    char: '🥕',
-    name: 'carrot',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F336 FE0F',
-    char: '🌶️',
-    name: 'hot pepper',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F952',
-    char: '🥒',
-    name: 'cucumber',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F96C',
-    char: '🥬',
-    name: 'leafy green',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F966',
-    char: '🥦',
-    name: 'broccoli',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F9C4',
-    char: '🧄',
-    name: 'garlic',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F9C5',
-    char: '🧅',
-    name: 'onion',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F344',
-    char: '🍄',
-    name: 'mushroom',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F330',
-    char: '🌰',
-    name: 'chestnut',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F35E',
-    char: '🍞',
-    name: 'bread',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F356',
-    char: '🍖',
-    name: 'meat on bone',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F357',
-    char: '🍗',
-    name: 'poultry leg',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F354',
-    char: '🍔',
-    name: 'hamburger',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F35F',
-    char: '🍟',
-    name: 'french fries',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F355',
-    char: '🍕',
-    name: 'pizza',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F373',
-    char: '🍳',
-    name: 'cooking',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F958',
-    char: '🥘',
-    name: 'shallow pan of food',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F372',
-    char: '🍲',
-    name: 'pot of food',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F371',
-    char: '🍱',
-    name: 'bento box',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F358',
-    char: '🍘',
-    name: 'rice cracker',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F359',
-    char: '🍙',
-    name: 'rice ball',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F35A',
-    char: '🍚',
-    name: 'cooked rice',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F35B',
-    char: '🍛',
-    name: 'curry rice',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F35C',
-    char: '🍜',
-    name: 'steaming bowl',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F35D',
-    char: '🍝',
-    name: 'spaghetti',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F360',
-    char: '🍠',
-    name: 'roasted sweet potato',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F362',
-    char: '🍢',
-    name: 'oden',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F363',
-    char: '🍣',
-    name: 'sushi',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F364',
-    char: '🍤',
-    name: 'fried shrimp',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F365',
-    char: '🍥',
-    name: 'fish cake with swirl',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F980',
-    char: '🦀',
-    name: 'crab',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F99E',
-    char: '🦞',
-    name: 'lobster',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F990',
-    char: '🦐',
-    name: 'shrimp',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F991',
-    char: '🦑',
-    name: 'squid',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F9AA',
-    char: '🦪',
-    name: 'oyster',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F361',
-    char: '🍡',
-    name: 'dango',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F366',
-    char: '🍦',
-    name: 'soft ice cream',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F367',
-    char: '🍧',
-    name: 'shaved ice',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F368',
-    char: '🍨',
-    name: 'ice cream',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F9C1',
-    char: '🧁',
-    name: 'cupcake',
-    category: 'foodAndDrink',
-  },
-  { codes: '1F967', char: '🥧', name: 'pie', category: 'foodAndDrink' },
-  {
-    codes: '1F369',
-    char: '🍩',
-    name: 'doughnut',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F36A',
-    char: '🍪',
-    name: 'cookie',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F382',
-    char: '🎂',
-    name: 'birthday cake',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F370',
-    char: '🍰',
-    name: 'shortcake',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F36B',
-    char: '🍫',
-    name: 'chocolate bar',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F36C',
-    char: '🍬',
-    name: 'candy',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F36D',
-    char: '🍭',
-    name: 'lollipop',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F36E',
-    char: '🍮',
-    name: 'custard',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F36F',
-    char: '🍯',
-    name: 'honey pot',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F37C',
-    char: '🍼',
-    name: 'baby bottle',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '2615',
-    char: '☕',
-    name: 'hot beverage',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F375',
-    char: '🍵',
-    name: 'teacup without handle',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F376',
-    char: '🍶',
-    name: 'sake',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F377',
-    char: '🍷',
-    name: 'wine glass',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F378',
-    char: '🍸',
-    name: 'cocktail glass',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F379',
-    char: '🍹',
-    name: 'tropical drink',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F37A',
-    char: '🍺',
-    name: 'beer mug',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F37B',
-    char: '🍻',
-    name: 'clinking beer mugs',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F95B',
-    char: '🥛',
-    name: 'glass of milk',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F37E',
-    char: '🍾',
-    name: 'bottle with popping cork',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F942',
-    char: '🥂',
-    name: 'clinking glasses',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F943',
-    char: '🥃',
-    name: 'tumbler glass',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F964',
-    char: '🥤',
-    name: 'cup with straw',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F9C3',
-    char: '🧃',
-    name: 'beverage box',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F9C9',
-    char: '🧉',
-    name: 'mate',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F9CA',
-    char: '🧊',
-    name: 'ice cube',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F962',
-    char: '🥢',
-    name: 'chopsticks',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F37D FE0F',
-    char: '🍽️',
-    name: 'fork and knife with plate',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F944',
-    char: '🥄',
-    name: 'spoon',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F374',
-    char: '🍴',
-    name: 'fork and knife',
-    category: 'foodAndDrink',
-  },
-  {
-    codes: '1F52A',
-    char: '🔪',
-    name: 'kitchen knife',
-    category: 'foodAndDrink',
-  },
-
-  {
-    codes: '1F383',
-    char: '🎃',
-    name: 'jack-o-lantern',
-    category: 'activity',
-  },
-  {
-    codes: '1F384',
-    char: '🎄',
-    name: 'Christmas tree',
-    category: 'activity',
-  },
-  {
-    codes: '1F386',
-    char: '🎆',
-    name: 'fireworks',
-    category: 'activity',
-  },
-  {
-    codes: '1F387',
-    char: '🎇',
-    name: 'sparkler',
-    category: 'activity',
-  },
-  { codes: '2728', char: '✨', name: 'sparkles', category: 'activity' },
-  { codes: '1F388', char: '🎈', name: 'balloon', category: 'activity' },
-  {
-    codes: '1F389',
-    char: '🎉',
-    name: 'party popper',
-    category: 'activity',
-  },
-  {
-    codes: '1F38A',
-    char: '🎊',
-    name: 'confetti ball',
-    category: 'activity',
-  },
-  {
-    codes: '1F38B',
-    char: '🎋',
-    name: 'tanabata tree',
-    category: 'activity',
-  },
-  {
-    codes: '1F38D',
-    char: '🎍',
-    name: 'pine decoration',
-    category: 'activity',
-  },
-  {
-    codes: '1F38E',
-    char: '🎎',
-    name: 'Japanese dolls',
-    category: 'activity',
-  },
-  {
-    codes: '1F38F',
-    char: '🎏',
-    name: 'carp streamer',
-    category: 'activity',
-  },
-  { codes: '1F380', char: '🎀', name: 'ribbon', category: 'activity' },
-  {
-    codes: '1F381',
-    char: '🎁',
-    name: 'wrapped gift',
-    category: 'activity',
-  },
-  { codes: '1F3C6', char: '🏆', name: 'trophy', category: 'activity' },
-  {
-    codes: '26BD',
-    char: '⚽',
-    name: 'soccer ball',
-    category: 'activity',
-  },
-  { codes: '26BE', char: '⚾', name: 'baseball', category: 'activity' },
-  {
-    codes: '1F3C0',
-    char: '🏀',
-    name: 'basketball',
-    category: 'activity',
-  },
-  {
-    codes: '1F3C8',
-    char: '🏈',
-    name: 'american football',
-    category: 'activity',
-  },
-  {
-    codes: '1F3C9',
-    char: '🏉',
-    name: 'rugby football',
-    category: 'activity',
-  },
-  { codes: '1F3BE', char: '🎾', name: 'tennis', category: 'activity' },
-  { codes: '1F3B3', char: '🎳', name: 'bowling', category: 'activity' },
-  {
-    codes: '26F3',
-    char: '⛳',
-    name: 'flag in hole',
-    category: 'activity',
-  },
-  {
-    codes: '1F3A3',
-    char: '🎣',
-    name: 'fishing pole',
-    category: 'activity',
-  },
-  {
-    codes: '1F3BD',
-    char: '🎽',
-    name: 'running shirt',
-    category: 'activity',
-  },
-  { codes: '1F3BF', char: '🎿', name: 'skis', category: 'activity' },
-
-  {
-    codes: '1F3AF',
-    char: '🎯',
-    name: 'direct hit',
-    category: 'activity',
-  },
-
-  {
-    codes: '1F3B1',
-    char: '🎱',
-    name: 'pool 8 ball',
-    category: 'activity',
-  },
-  {
-    codes: '1F52E',
-    char: '🔮',
-    name: 'crystal ball',
-    category: 'activity',
-  },
-  {
-    codes: '1F3AE',
-    char: '🎮',
-    name: 'video game',
-    category: 'activity',
-  },
-  {
-    codes: '1F3B0',
-    char: '🎰',
-    name: 'slot machine',
-    category: 'activity',
-  },
-  {
-    codes: '1F3B2',
-    char: '🎲',
-    name: 'game die',
-    category: 'activity',
-  },
-  {
-    codes: '2666 FE0F',
-    char: '♦️',
-    name: 'diamond suit',
-    category: 'activity',
-  },
-  {
-    codes: '2663 FE0F',
-    char: '♣️',
-    name: 'club suit',
-    category: 'activity',
-  },
-  {
-    codes: '265F FE0F',
-    char: '♟️',
-    name: 'chess pawn',
-    category: 'activity',
-  },
-  {
-    codes: '1F5BC FE0F',
-    char: '🖼️',
-    name: 'framed picture',
-    category: 'activity',
-  },
-  {
-    codes: '1F9F5',
-    char: '🧵',
-    name: 'thread',
-    category: 'activity',
-  },
-  {
-    codes: '1F9F6',
-    char: '🧶',
-    name: 'yarn',
-    category: 'activity',
-  },
-
-  {
-    codes: '1F004',
-    char: '🀄',
-    name: 'mahjong red dragon',
-    category: 'activity',
-  },
-  {
-    codes: '1F3B4',
-    char: '🎴',
-    name: 'flower playing cards',
-    category: 'activity',
-  },
-  {
-    codes: '1F3AD',
-    char: '🎭',
-    name: 'performing arts',
-    category: 'activity',
-  },
-  {
-    codes: '1F3A8',
-    char: '🎨',
-    name: 'artist palette',
-    category: 'activity',
-  },
-
-  {
-    codes: '1F30F',
-    char: '🌏',
-    name: 'globe showing Asia-Australia',
-    category: 'places',
-  },
-  {
-    codes: '1F310',
-    char: '🌐',
-    name: 'globe with meridians',
-    category: 'places',
-  },
-  { codes: '1F30B', char: '🌋', name: 'volcano', category: 'places' },
-  {
-    codes: '1F5FB',
-    char: '🗻',
-    name: 'mount fuji',
-    category: 'places',
-  },
-  { codes: '1F3E0', char: '🏠', name: 'house', category: 'places' },
-  {
-    codes: '1F3E1',
-    char: '🏡',
-    name: 'house with garden',
-    category: 'places',
-  },
-  {
-    codes: '1F3E2',
-    char: '🏢',
-    name: 'office building',
-    category: 'places',
-  },
-  {
-    codes: '1F3E3',
-    char: '🏣',
-    name: 'Japanese post office',
-    category: 'places',
-  },
-  {
-    codes: '1F3E4',
-    char: '🏤',
-    name: 'post office',
-    category: 'places',
-  },
-  { codes: '1F3E5', char: '🏥', name: 'hospital', category: 'places' },
-  { codes: '1F3E6', char: '🏦', name: 'bank', category: 'places' },
-  { codes: '1F3E8', char: '🏨', name: 'hotel', category: 'places' },
-  {
-    codes: '1F3E9',
-    char: '🏩',
-    name: 'love hotel',
-    category: 'places',
-  },
-  {
-    codes: '1F3EA',
-    char: '🏪',
-    name: 'convenience store',
-    category: 'places',
-  },
-  { codes: '1F3EB', char: '🏫', name: 'school', category: 'places' },
-  {
-    codes: '1F3EC',
-    char: '🏬',
-    name: 'department store',
-    category: 'places',
-  },
-  { codes: '1F3ED', char: '🏭', name: 'factory', category: 'places' },
-  {
-    codes: '1F3EF',
-    char: '🏯',
-    name: 'Japanese castle',
-    category: 'places',
-  },
-  { codes: '1F3F0', char: '🏰', name: 'castle', category: 'places' },
-  { codes: '1F492', char: '💒', name: 'wedding', category: 'places' },
-  {
-    codes: '1F5FD',
-    char: '🗽',
-    name: 'Statue of Liberty',
-    category: 'places',
-  },
-  { codes: '26EA', char: '⛪', name: 'church', category: 'places' },
-  { codes: '26F2', char: '⛲', name: 'fountain', category: 'places' },
-  { codes: '26FA', char: '⛺', name: 'tent', category: 'places' },
-  { codes: '1F301', char: '🌁', name: 'foggy', category: 'places' },
-  {
-    codes: '1F303',
-    char: '🌃',
-    name: 'night with stars',
-    category: 'places',
-  },
-  {
-    codes: '1F304',
-    char: '🌄',
-    name: 'sunrise over mountains',
-    category: 'places',
-  },
-  { codes: '1F305', char: '🌅', name: 'sunrise', category: 'places' },
-  {
-    codes: '1F306',
-    char: '🌆',
-    name: 'cityscape at dusk',
-    category: 'places',
-  },
-  { codes: '1F307', char: '🌇', name: 'sunset', category: 'places' },
-  {
-    codes: '1F309',
-    char: '🌉',
-    name: 'bridge at night',
-    category: 'places',
-  },
-  {
-    codes: '2668 FE0F',
-    char: '♨️',
-    name: 'hot springs',
-    category: 'places',
-  },
-  {
-    codes: '1F3A0',
-    char: '🎠',
-    name: 'carousel horse',
-    category: 'places',
-  },
-  {
-    codes: '1F3A1',
-    char: '🎡',
-    name: 'ferris wheel',
-    category: 'places',
-  },
-  {
-    codes: '1F3A2',
-    char: '🎢',
-    name: 'roller coaster',
-    category: 'places',
-  },
-  {
-    codes: '1F488',
-    char: '💈',
-    name: 'barber pole',
-    category: 'places',
-  },
-  {
-    codes: '1F3AA',
-    char: '🎪',
-    name: 'circus tent',
-    category: 'places',
-  },
-  {
-    codes: '1F682',
-    char: '🚂',
-    name: 'locomotive',
-    category: 'places',
-  },
-  {
-    codes: '1F683',
-    char: '🚃',
-    name: 'railway car',
-    category: 'places',
-  },
-  {
-    codes: '1F684',
-    char: '🚄',
-    name: 'high-speed train',
-    category: 'places',
-  },
-  {
-    codes: '1F685',
-    char: '🚅',
-    name: 'bullet train',
-    category: 'places',
-  },
-  { codes: '1F686', char: '🚆', name: 'train', category: 'places' },
-  { codes: '1F687', char: '🚇', name: 'metro', category: 'places' },
-  {
-    codes: '1F688',
-    char: '🚈',
-    name: 'light rail',
-    category: 'places',
-  },
-  { codes: '1F689', char: '🚉', name: 'station', category: 'places' },
-  { codes: '1F68A', char: '🚊', name: 'tram', category: 'places' },
-  { codes: '1F69D', char: '🚝', name: 'monorail', category: 'places' },
-  {
-    codes: '1F69E',
-    char: '🚞',
-    name: 'mountain railway',
-    category: 'places',
-  },
-  { codes: '1F68B', char: '🚋', name: 'tram car', category: 'places' },
-  { codes: '1F68C', char: '🚌', name: 'bus', category: 'places' },
-  {
-    codes: '1F68D',
-    char: '🚍',
-    name: 'oncoming bus',
-    category: 'places',
-  },
-  {
-    codes: '1F68E',
-    char: '🚎',
-    name: 'trolleybus',
-    category: 'places',
-  },
-  { codes: '1F690', char: '🚐', name: 'minibus', category: 'places' },
-  { codes: '1F691', char: '🚑', name: 'ambulance', category: 'places' },
-  {
-    codes: '1F692',
-    char: '🚒',
-    name: 'fire engine',
-    category: 'places',
-  },
-  {
-    codes: '1F693',
-    char: '🚓',
-    name: 'police car',
-    category: 'places',
-  },
-  { codes: '1F695', char: '🚕', name: 'taxi', category: 'places' },
-  {
-    codes: '1F697',
-    char: '🚗',
-    name: 'automobile',
-    category: 'places',
-  },
-  {
-    codes: '1F698',
-    char: '🚘',
-    name: 'oncoming automobile',
-    category: 'places',
-  },
-  {
-    codes: '1F699',
-    char: '🚙',
-    name: 'sport utility vehicle',
-    category: 'places',
-  },
-  {
-    codes: '1F69A',
-    char: '🚚',
-    name: 'delivery truck',
-    category: 'places',
-  },
-  {
-    codes: '1F69B',
-    char: '🚛',
-    name: 'articulated lorry',
-    category: 'places',
-  },
-  { codes: '1F6B2', char: '🚲', name: 'bicycle', category: 'places' },
-  { codes: '26FD', char: '⛽', name: 'fuel pump', category: 'places' },
-  {
-    codes: '1F6A8',
-    char: '🚨',
-    name: 'police car light',
-    category: 'places',
-  },
-  {
-    codes: '1F6A5',
-    char: '🚥',
-    name: 'horizontal traffic light',
-    category: 'places',
-  },
-  {
-    codes: '1F6A6',
-    char: '🚦',
-    name: 'vertical traffic light',
-    category: 'places',
-  },
-  {
-    codes: '1F6A7',
-    char: '🚧',
-    name: 'construction',
-    category: 'places',
-  },
-  { codes: '2693', char: '⚓', name: 'anchor', category: 'places' },
-  { codes: '26F5', char: '⛵', name: 'sailboat', category: 'places' },
-  { codes: '1F6A4', char: '🚤', name: 'speedboat', category: 'places' },
-  { codes: '1F6A2', char: '🚢', name: 'ship', category: 'places' },
-  {
-    codes: '2708 FE0F',
-    char: '✈️',
-    name: 'airplane',
-    category: 'places',
-  },
-  {
-    codes: '1F681',
-    char: '🚁',
-    name: 'helicopter',
-    category: 'places',
-  },
-  {
-    codes: '1F69F',
-    char: '🚟',
-    name: 'suspension railway',
-    category: 'places',
-  },
-  {
-    codes: '1F6A0',
-    char: '🚠',
-    name: 'mountain cableway',
-    category: 'places',
-  },
-  {
-    codes: '1F6A1',
-    char: '🚡',
-    name: 'aerial tramway',
-    category: 'places',
-  },
-  {
-    codes: '1F6F0 FE0F',
-    char: '🛰️',
-    name: 'satellite',
-    category: 'places',
-  },
-  { codes: '1F680', char: '🚀', name: 'rocket', category: 'places' },
-  {
-    codes: '231B',
-    char: '⌛',
-    name: 'hourglass done',
-    category: 'places',
-  },
-  {
-    codes: '23F3',
-    char: '⏳',
-    name: 'hourglass not done',
-    category: 'places',
-  },
-  { codes: '231A', char: '⌚', name: 'watch', category: 'places' },
-  {
-    codes: '23F0',
-    char: '⏰',
-    name: 'alarm clock',
-    category: 'places',
-  },
-  {
-    codes: '1F55B',
-    char: '🕛',
-    name: 'twelve o’clock',
-    category: 'places',
-  },
-  {
-    codes: '1F567',
-    char: '🕧',
-    name: 'twelve-thirty',
-    category: 'places',
-  },
-  {
-    codes: '1F550',
-    char: '🕐',
-    name: 'one o’clock',
-    category: 'places',
-  },
-  {
-    codes: '1F55C',
-    char: '🕜',
-    name: 'one-thirty',
-    category: 'places',
-  },
-  {
-    codes: '1F551',
-    char: '🕑',
-    name: 'two o’clock',
-    category: 'places',
-  },
-  {
-    codes: '1F55D',
-    char: '🕝',
-    name: 'two-thirty',
-    category: 'places',
-  },
-  {
-    codes: '1F552',
-    char: '🕒',
-    name: 'three o’clock',
-    category: 'places',
-  },
-  {
-    codes: '1F55E',
-    char: '🕞',
-    name: 'three-thirty',
-    category: 'places',
-  },
-  {
-    codes: '1F553',
-    char: '🕓',
-    name: 'four o’clock',
-    category: 'places',
-  },
-  {
-    codes: '1F55F',
-    char: '🕟',
-    name: 'four-thirty',
-    category: 'places',
-  },
-  {
-    codes: '1F554',
-    char: '🕔',
-    name: 'five o’clock',
-    category: 'places',
-  },
-  {
-    codes: '1F560',
-    char: '🕠',
-    name: 'five-thirty',
-    category: 'places',
-  },
-  {
-    codes: '1F555',
-    char: '🕕',
-    name: 'six o’clock',
-    category: 'places',
-  },
-  {
-    codes: '1F561',
-    char: '🕡',
-    name: 'six-thirty',
-    category: 'places',
-  },
-  {
-    codes: '1F556',
-    char: '🕖',
-    name: 'seven o’clock',
-    category: 'places',
-  },
-  {
-    codes: '1F562',
-    char: '🕢',
-    name: 'seven-thirty',
-    category: 'places',
-  },
-  {
-    codes: '1F557',
-    char: '🕗',
-    name: 'eight o’clock',
-    category: 'places',
-  },
-  {
-    codes: '1F563',
-    char: '🕣',
-    name: 'eight-thirty',
-    category: 'places',
-  },
-  {
-    codes: '1F558',
-    char: '🕘',
-    name: 'nine o’clock',
-    category: 'places',
-  },
-  {
-    codes: '1F564',
-    char: '🕤',
-    name: 'nine-thirty',
-    category: 'places',
-  },
-  {
-    codes: '1F559',
-    char: '🕙',
-    name: 'ten o’clock',
-    category: 'places',
-  },
-  {
-    codes: '1F565',
-    char: '🕥',
-    name: 'ten-thirty',
-    category: 'places',
-  },
-  {
-    codes: '1F55A',
-    char: '🕚',
-    name: 'eleven o’clock',
-    category: 'places',
-  },
-  {
-    codes: '1F566',
-    char: '🕦',
-    name: 'eleven-thirty',
-    category: 'places',
-  },
-  { codes: '1F453', char: '👓', name: 'glasses', category: 'objects' },
-  { codes: '1F454', char: '👔', name: 'necktie', category: 'objects' },
-  { codes: '1F455', char: '👕', name: 't-shirt', category: 'objects' },
-  { codes: '1F456', char: '👖', name: 'jeans', category: 'objects' },
-  { codes: '1F457', char: '👗', name: 'dress', category: 'objects' },
-  { codes: '1F459', char: '👙', name: 'bikini', category: 'objects' },
-  {
-    codes: '1F45A',
-    char: '👚',
-    name: 'woman’s clothes',
-    category: 'objects',
-  },
-  { codes: '1F45B', char: '👛', name: 'purse', category: 'objects' },
-  { codes: '1F45C', char: '👜', name: 'handbag', category: 'objects' },
-  {
-    codes: '1F45D',
-    char: '👝',
-    name: 'clutch bag',
-    category: 'objects',
-  },
-  { codes: '1F392', char: '🎒', name: 'backpack', category: 'objects' },
-  {
-    codes: '1F45E',
-    char: '👞',
-    name: 'man’s shoe',
-    category: 'objects',
-  },
-  {
-    codes: '1F45F',
-    char: '👟',
-    name: 'running shoe',
-    category: 'objects',
-  },
-  {
-    codes: '1F460',
-    char: '👠',
-    name: 'high-heeled shoe',
-    category: 'objects',
-  },
-  {
-    codes: '1F461',
-    char: '👡',
-    name: 'woman’s sandal',
-    category: 'objects',
-  },
-  {
-    codes: '1F462',
-    char: '👢',
-    name: 'woman’s boot',
-    category: 'objects',
-  },
-  { codes: '1F451', char: '👑', name: 'crown', category: 'objects' },
-  {
-    codes: '1F452',
-    char: '👒',
-    name: 'woman’s hat',
-    category: 'objects',
-  },
-  { codes: '1F3A9', char: '🎩', name: 'top hat', category: 'objects' },
-  {
-    codes: '1F393',
-    char: '🎓',
-    name: 'graduation cap',
-    category: 'objects',
-  },
-  { codes: '1F484', char: '💄', name: 'lipstick', category: 'objects' },
-  { codes: '1F48D', char: '💍', name: 'ring', category: 'objects' },
-  {
-    codes: '1F48E',
-    char: '💎',
-    name: 'gem stone',
-    category: 'objects',
-  },
-  {
-    codes: '1F507',
-    char: '🔇',
-    name: 'muted speaker',
-    category: 'objects',
-  },
-  {
-    codes: '1F508',
-    char: '🔈',
-    name: 'speaker low volume',
-    category: 'objects',
-  },
-  {
-    codes: '1F509',
-    char: '🔉',
-    name: 'speaker medium volume',
-    category: 'objects',
-  },
-  {
-    codes: '1F50A',
-    char: '🔊',
-    name: 'speaker high volume',
-    category: 'objects',
-  },
-  {
-    codes: '1F4E2',
-    char: '📢',
-    name: 'loudspeaker',
-    category: 'objects',
-  },
-  {
-    codes: '1F4E3',
-    char: '📣',
-    name: 'megaphone',
-    category: 'objects',
-  },
-  {
-    codes: '1F4EF',
-    char: '📯',
-    name: 'postal horn',
-    category: 'objects',
-  },
-  { codes: '1F514', char: '🔔', name: 'bell', category: 'objects' },
-  {
-    codes: '1F515',
-    char: '🔕',
-    name: 'bell with slash',
-    category: 'objects',
-  },
-  {
-    codes: '1F3BC',
-    char: '🎼',
-    name: 'musical score',
-    category: 'objects',
-  },
-  {
-    codes: '1F3B5',
-    char: '🎵',
-    name: 'musical note',
-    category: 'objects',
-  },
-  {
-    codes: '1F3B6',
-    char: '🎶',
-    name: 'musical notes',
-    category: 'objects',
-  },
-  {
-    codes: '1F3A4',
-    char: '🎤',
-    name: 'microphone',
-    category: 'objects',
-  },
-  {
-    codes: '1F3A7',
-    char: '🎧',
-    name: 'headphone',
-    category: 'objects',
-  },
-  { codes: '1F4FB', char: '📻', name: 'radio', category: 'objects' },
-  {
-    codes: '1F3B7',
-    char: '🎷',
-    name: 'saxophone',
-    category: 'objects',
-  },
-  { codes: '1F3B8', char: '🎸', name: 'guitar', category: 'objects' },
-  {
-    codes: '1F3B9',
-    char: '🎹',
-    name: 'musical keyboard',
-    category: 'objects',
-  },
-  { codes: '1F3BA', char: '🎺', name: 'trumpet', category: 'objects' },
-  { codes: '1F3BB', char: '🎻', name: 'violin', category: 'objects' },
-  {
-    codes: '1F4F1',
-    char: '📱',
-    name: 'mobile phone',
-    category: 'objects',
-  },
-  {
-    codes: '1F4F2',
-    char: '📲',
-    name: 'mobile phone with arrow',
-    category: 'objects',
-  },
-  {
-    codes: '260E FE0F',
-    char: '☎️',
-    name: 'telephone',
-    category: 'objects',
-  },
-  {
-    codes: '1F4DE',
-    char: '📞',
-    name: 'telephone receiver',
-    category: 'objects',
-  },
-  { codes: '1F4DF', char: '📟', name: 'pager', category: 'objects' },
-  {
-    codes: '1F4E0',
-    char: '📠',
-    name: 'fax machine',
-    category: 'objects',
-  },
-  { codes: '1F50B', char: '🔋', name: 'battery', category: 'objects' },
-  {
-    codes: '1F50C',
-    char: '🔌',
-    name: 'electric plug',
-    category: 'objects',
-  },
-  {
-    codes: '1F4BB',
-    char: '💻',
-    name: 'laptop computer',
-    category: 'objects',
-  },
-  {
-    codes: '1F4BD',
-    char: '💽',
-    name: 'computer disk',
-    category: 'objects',
-  },
-  {
-    codes: '1F4BE',
-    char: '💾',
-    name: 'floppy disk',
-    category: 'objects',
-  },
-  {
-    codes: '1F4BF',
-    char: '💿',
-    name: 'optical disk',
-    category: 'objects',
-  },
-  { codes: '1F4C0', char: '📀', name: 'dvd', category: 'objects' },
-  { codes: '1F9EE', char: '🧮', name: 'abacus', category: 'objects' },
-  {
-    codes: '1F3A5',
-    char: '🎥',
-    name: 'movie camera',
-    category: 'objects',
-  },
-  {
-    codes: '1F39E FE0F',
-    char: '🎞️',
-    name: 'film frames',
-    category: 'objects',
-  },
-  {
-    codes: '1F4FD FE0F',
-    char: '📽️',
-    name: 'film projector',
-    category: 'objects',
-  },
-  {
-    codes: '1F3AC',
-    char: '🎬',
-    name: 'clapper board',
-    category: 'objects',
-  },
-  {
-    codes: '1F4FA',
-    char: '📺',
-    name: 'television',
-    category: 'objects',
-  },
-  { codes: '1F4F7', char: '📷', name: 'camera', category: 'objects' },
-  {
-    codes: '1F4F9',
-    char: '📹',
-    name: 'video camera',
-    category: 'objects',
-  },
-  {
-    codes: '1F4FC',
-    char: '📼',
-    name: 'videocassette',
-    category: 'objects',
-  },
-  {
-    codes: '1F50D',
-    char: '🔍',
-    name: 'magnifying glass tilted left',
-    category: 'objects',
-  },
-  {
-    codes: '1F50E',
-    char: '🔎',
-    name: 'magnifying glass tilted right',
-    category: 'objects',
-  },
-  {
-    codes: '1F4A1',
-    char: '💡',
-    name: 'light bulb',
-    category: 'objects',
-  },
-  {
-    codes: '1F526',
-    char: '🔦',
-    name: 'flashlight',
-    category: 'objects',
-  },
-  {
-    codes: '1F3EE',
-    char: '🏮',
-    name: 'red paper lantern',
-    category: 'objects',
-  },
-  {
-    codes: '1F4D4',
-    char: '📔',
-    name: 'notebook with decorative cover',
-    category: 'objects',
-  },
-  {
-    codes: '1F4D5',
-    char: '📕',
-    name: 'closed book',
-    category: 'objects',
-  },
-  {
-    codes: '1F4D6',
-    char: '📖',
-    name: 'open book',
-    category: 'objects',
-  },
-  {
-    codes: '1F4D7',
-    char: '📗',
-    name: 'green book',
-    category: 'objects',
-  },
-  {
-    codes: '1F4D8',
-    char: '📘',
-    name: 'blue book',
-    category: 'objects',
-  },
-  {
-    codes: '1F4D9',
-    char: '📙',
-    name: 'orange book',
-    category: 'objects',
-  },
-  { codes: '1F4DA', char: '📚', name: 'books', category: 'objects' },
-  { codes: '1F4D3', char: '📓', name: 'notebook', category: 'objects' },
-  { codes: '1F4D2', char: '📒', name: 'ledger', category: 'objects' },
-  {
-    codes: '1F4C3',
-    char: '📃',
-    name: 'page with curl',
-    category: 'objects',
-  },
-  { codes: '1F4DC', char: '📜', name: 'scroll', category: 'objects' },
-  {
-    codes: '1F4C4',
-    char: '📄',
-    name: 'page facing up',
-    category: 'objects',
-  },
-  {
-    codes: '1F4F0',
-    char: '📰',
-    name: 'newspaper',
-    category: 'objects',
-  },
-  {
-    codes: '1F4D1',
-    char: '📑',
-    name: 'bookmark tabs',
-    category: 'objects',
-  },
-  { codes: '1F516', char: '🔖', name: 'bookmark', category: 'objects' },
-  {
-    codes: '1F4B0',
-    char: '💰',
-    name: 'money bag',
-    category: 'objects',
-  },
-  {
-    codes: '1F4B4',
-    char: '💴',
-    name: 'yen banknote',
-    category: 'objects',
-  },
-  {
-    codes: '1F4B5',
-    char: '💵',
-    name: 'dollar banknote',
-    category: 'objects',
-  },
-  {
-    codes: '1F4B6',
-    char: '💶',
-    name: 'euro banknote',
-    category: 'objects',
-  },
-  {
-    codes: '1F4B7',
-    char: '💷',
-    name: 'pound banknote',
-    category: 'objects',
-  },
-  {
-    codes: '1F4B8',
-    char: '💸',
-    name: 'money with wings',
-    category: 'objects',
-  },
-  {
-    codes: '1F4B3',
-    char: '💳',
-    name: 'credit card',
-    category: 'objects',
-  },
-  {
-    codes: '1F4B9',
-    char: '💹',
-    name: 'chart increasing with yen',
-    category: 'objects',
-  },
-  {
-    codes: '1F4B1',
-    char: '💱',
-    name: 'currency exchange',
-    category: 'objects',
-  },
-  {
-    codes: '1F4B2',
-    char: '💲',
-    name: 'heavy dollar sign',
-    category: 'objects',
-  },
-  {
-    codes: '2709 FE0F',
-    char: '✉️',
-    name: 'envelope',
-    category: 'objects',
-  },
-  { codes: '1F4E7', char: '📧', name: 'e-mail', category: 'objects' },
-  {
-    codes: '1F4E8',
-    char: '📨',
-    name: 'incoming envelope',
-    category: 'objects',
-  },
-  {
-    codes: '1F4E9',
-    char: '📩',
-    name: 'envelope with arrow',
-    category: 'objects',
-  },
-  {
-    codes: '1F4E4',
-    char: '📤',
-    name: 'outbox tray',
-    category: 'objects',
-  },
-  {
-    codes: '1F4E5',
-    char: '📥',
-    name: 'inbox tray',
-    category: 'objects',
-  },
-  { codes: '1F4E6', char: '📦', name: 'package', category: 'objects' },
-  {
-    codes: '1F4EB',
-    char: '📫',
-    name: 'closed mailbox with raised flag',
-    category: 'objects',
-  },
-  {
-    codes: '1F4EA',
-    char: '📪',
-    name: 'closed mailbox with lowered flag',
-    category: 'objects',
-  },
-  {
-    codes: '1F4EC',
-    char: '📬',
-    name: 'open mailbox with raised flag',
-    category: 'objects',
-  },
-  {
-    codes: '1F4ED',
-    char: '📭',
-    name: 'open mailbox with lowered flag',
-    category: 'objects',
-  },
-  { codes: '1F4EE', char: '📮', name: 'postbox', category: 'objects' },
-  {
-    codes: '270F FE0F',
-    char: '✏️',
-    name: 'pencil',
-    category: 'objects',
-  },
-  {
-    codes: '2712 FE0F',
-    char: '✒️',
-    name: 'black nib',
-    category: 'objects',
-  },
-  { codes: '1F4DD', char: '📝', name: 'memo', category: 'objects' },
-  {
-    codes: '1F4BC',
-    char: '💼',
-    name: 'briefcase',
-    category: 'objects',
-  },
-  {
-    codes: '1F4C1',
-    char: '📁',
-    name: 'file folder',
-    category: 'objects',
-  },
-  {
-    codes: '1F4C2',
-    char: '📂',
-    name: 'open file folder',
-    category: 'objects',
-  },
-  { codes: '1F4C5', char: '📅', name: 'calendar', category: 'objects' },
-  {
-    codes: '1F4C7',
-    char: '📇',
-    name: 'card index',
-    category: 'objects',
-  },
-  {
-    codes: '1F4C8',
-    char: '📈',
-    name: 'chart increasing',
-    category: 'objects',
-  },
-  {
-    codes: '1F4C9',
-    char: '📉',
-    name: 'chart decreasing',
-    category: 'objects',
-  },
-  {
-    codes: '1F4CA',
-    char: '📊',
-    name: 'bar chart',
-    category: 'objects',
-  },
-  {
-    codes: '1F4CB',
-    char: '📋',
-    name: 'clipboard',
-    category: 'objects',
-  },
-  { codes: '1F4CC', char: '📌', name: 'pushpin', category: 'objects' },
-  {
-    codes: '1F4CD',
-    char: '📍',
-    name: 'round pushpin',
-    category: 'objects',
-  },
-  {
-    codes: '1F4CE',
-    char: '📎',
-    name: 'paperclip',
-    category: 'objects',
-  },
-  {
-    codes: '1F4CF',
-    char: '📏',
-    name: 'straight ruler',
-    category: 'objects',
-  },
-  {
-    codes: '1F4D0',
-    char: '📐',
-    name: 'triangular ruler',
-    category: 'objects',
-  },
-  {
-    codes: '2702 FE0F',
-    char: '✂️',
-    name: 'scissors',
-    category: 'objects',
-  },
-  { codes: '1F512', char: '🔒', name: 'locked', category: 'objects' },
-  { codes: '1F513', char: '🔓', name: 'unlocked', category: 'objects' },
-  {
-    codes: '1F50F',
-    char: '🔏',
-    name: 'locked with pen',
-    category: 'objects',
-  },
-  {
-    codes: '1F510',
-    char: '🔐',
-    name: 'locked with key',
-    category: 'objects',
-  },
-  { codes: '1F511', char: '🔑', name: 'key', category: 'objects' },
-  { codes: '1F528', char: '🔨', name: 'hammer', category: 'objects' },
-  { codes: '1F52B', char: '🔫', name: 'pistol', category: 'objects' },
-  { codes: '1F517', char: '🔗', name: 'link', category: 'objects' },
-  {
-    codes: '1F52C',
-    char: '🔬',
-    name: 'microscope',
-    category: 'objects',
-  },
-  {
-    codes: '1F52D',
-    char: '🔭',
-    name: 'telescope',
-    category: 'objects',
-  },
-  {
-    codes: '1F4E1',
-    char: '📡',
-    name: 'satellite antenna',
-    category: 'objects',
-  },
-  { codes: '1F489', char: '💉', name: 'syringe', category: 'objects' },
-  { codes: '1F48A', char: '💊', name: 'pill', category: 'objects' },
-  { codes: '1F6AA', char: '🚪', name: 'door', category: 'objects' },
-  { codes: '1F6BD', char: '🚽', name: 'toilet', category: 'objects' },
-  { codes: '1F6BF', char: '🚿', name: 'shower', category: 'objects' },
-  { codes: '1F6C1', char: '🛁', name: 'bathtub', category: 'objects' },
-  {
-    codes: '1F6AC',
-    char: '🚬',
-    name: 'cigarette',
-    category: 'objects',
-  },
-  { codes: '1F5FF', char: '🗿', name: 'moai', category: 'objects' },
-  {
-    codes: '2764 FE0F',
-    char: '❤️',
-    name: 'red heart',
-    category: 'symbols',
-  },
-  {
-    codes: '1F49B',
-    char: '💛',
-    name: 'yellow heart',
-    category: 'symbols',
-  },
-  {
-    codes: '1F49A',
-    char: '💚',
-    name: 'green heart',
-    category: 'symbols',
-  },
-  {
-    codes: '1F499',
-    char: '💙',
-    name: 'blue heart',
-    category: 'symbols',
-  },
-  {
-    codes: '1F49C',
-    char: '💜',
-    name: 'purple heart',
-    category: 'symbols',
-  },
-  {
-    codes: '1F48B',
-    char: '💋',
-    name: 'kiss mark',
-    category: 'symbols',
-  },
-  {
-    codes: '1F48C',
-    char: '💌',
-    name: 'love letter',
-    category: 'symbols',
-  },
-  {
-    codes: '1F498',
-    char: '💘',
-    name: 'heart with arrow',
-    category: 'symbols',
-  },
-  {
-    codes: '1F49D',
-    char: '💝',
-    name: 'heart with ribbon',
-    category: 'symbols',
-  },
-  {
-    codes: '1F496',
-    char: '💖',
-    name: 'sparkling heart',
-    category: 'symbols',
-  },
-  {
-    codes: '1F497',
-    char: '💗',
-    name: 'growing heart',
-    category: 'symbols',
-  },
-  {
-    codes: '1F493',
-    char: '💓',
-    name: 'beating heart',
-    category: 'symbols',
-  },
-  {
-    codes: '1F49E',
-    char: '💞',
-    name: 'revolving hearts',
-    category: 'symbols',
-  },
-  {
-    codes: '1F495',
-    char: '💕',
-    name: 'two hearts',
-    category: 'symbols',
-  },
-  {
-    codes: '1F49F',
-    char: '💟',
-    name: 'heart decoration',
-    category: 'symbols',
-  },
-  {
-    codes: '1F494',
-    char: '💔',
-    name: 'broken heart',
-    category: 'symbols',
-  },
-  {
-    codes: '1F4AF',
-    char: '💯',
-    name: 'hundred points',
-    category: 'symbols',
-  },
-  {
-    codes: '1F4A2',
-    char: '💢',
-    name: 'anger symbol',
-    category: 'symbols',
-  },
-  {
-    codes: '1F4A5',
-    char: '💥',
-    name: 'collision',
-    category: 'symbols',
-  },
-  { codes: '1F4AB', char: '💫', name: 'dizzy', category: 'symbols' },
-  {
-    codes: '1F4A6',
-    char: '💦',
-    name: 'sweat droplets',
-    category: 'symbols',
-  },
-  {
-    codes: '1F4A8',
-    char: '💨',
-    name: 'dashing away',
-    category: 'symbols',
-  },
-  { codes: '1F4A3', char: '💣', name: 'bomb', category: 'symbols' },
-  {
-    codes: '1F4AC',
-    char: '💬',
-    name: 'speech balloon',
-    category: 'symbols',
-  },
-  {
-    codes: '1F4AD',
-    char: '💭',
-    name: 'thought balloon',
-    category: 'symbols',
-  },
-  { codes: '1F4A4', char: '💤', name: 'zzz', category: 'symbols' },
-  {
-    codes: '1F6AE',
-    char: '🚮',
-    name: 'litter in bin sign',
-    category: 'symbols',
-  },
-  {
-    codes: '1F6B0',
-    char: '🚰',
-    name: 'potable water',
-    category: 'symbols',
-  },
-  {
-    codes: '267F',
-    char: '♿',
-    name: 'wheelchair symbol',
-    category: 'symbols',
-  },
-  {
-    codes: '1F6B9',
-    char: '🚹',
-    name: 'men’s room',
-    category: 'symbols',
-  },
-  {
-    codes: '1F6BA',
-    char: '🚺',
-    name: 'women’s room',
-    category: 'symbols',
-  },
-  { codes: '1F6BB', char: '🚻', name: 'restroom', category: 'symbols' },
-  {
-    codes: '1F6BC',
-    char: '🚼',
-    name: 'baby symbol',
-    category: 'symbols',
-  },
-  {
-    codes: '1F6BE',
-    char: '🚾',
-    name: 'water closet',
-    category: 'symbols',
-  },
-  {
-    codes: '1F6C2',
-    char: '🛂',
-    name: 'passport control',
-    category: 'symbols',
-  },
-  { codes: '1F6C3', char: '🛃', name: 'customs', category: 'symbols' },
-  {
-    codes: '1F6C4',
-    char: '🛄',
-    name: 'baggage claim',
-    category: 'symbols',
-  },
-  {
-    codes: '1F6C5',
-    char: '🛅',
-    name: 'left luggage',
-    category: 'symbols',
-  },
-  {
-    codes: '26A0 FE0F',
-    char: '⚠️',
-    name: 'warning',
-    category: 'symbols',
-  },
-  {
-    codes: '1F6B8',
-    char: '🚸',
-    name: 'children crossing',
-    category: 'symbols',
-  },
-  { codes: '26D4', char: '⛔', name: 'no entry', category: 'symbols' },
-  {
-    codes: '1F6AB',
-    char: '🚫',
-    name: 'prohibited',
-    category: 'symbols',
-  },
-  {
-    codes: '1F6B3',
-    char: '🚳',
-    name: 'no bicycles',
-    category: 'symbols',
-  },
-  {
-    codes: '1F6AD',
-    char: '🚭',
-    name: 'no smoking',
-    category: 'symbols',
-  },
-  {
-    codes: '1F6AF',
-    char: '🚯',
-    name: 'no littering',
-    category: 'symbols',
-  },
-  {
-    codes: '1F6B1',
-    char: '🚱',
-    name: 'non-potable water',
-    category: 'symbols',
-  },
-  {
-    codes: '1F6B7',
-    char: '🚷',
-    name: 'no pedestrians',
-    category: 'symbols',
-  },
-  {
-    codes: '1F4F5',
-    char: '📵',
-    name: 'no mobile phones',
-    category: 'symbols',
-  },
-  {
-    codes: '1F51E',
-    char: '🔞',
-    name: 'no one under eighteen',
-    category: 'symbols',
-  },
-  {
-    codes: '2622 FE0F',
-    char: '☢️',
-    name: 'radioactive',
-    category: 'symbols',
-  },
-  {
-    codes: '2623 FE0F',
-    char: '☣️',
-    name: 'biohazard',
-    category: 'symbols',
-  },
-  {
-    codes: '2B06 FE0F',
-    char: '⬆️',
-    name: 'up arrow',
-    category: 'symbols',
-  },
-  {
-    codes: '27A1 FE0F',
-    char: '➡️',
-    name: 'right arrow',
-    category: 'symbols',
-  },
-  {
-    codes: '2B07 FE0F',
-    char: '⬇️',
-    name: 'down arrow',
-    category: 'symbols',
-  },
-  {
-    codes: '2B05 FE0F',
-    char: '⬅️',
-    name: 'left arrow',
-    category: 'symbols',
-  },
-  {
-    codes: '21A9 FE0F',
-    char: '↩️',
-    name: 'right arrow curving left',
-    category: 'symbols',
-  },
-  {
-    codes: '21AA FE0F',
-    char: '↪️',
-    name: 'left arrow curving right',
-    category: 'symbols',
-  },
-  {
-    codes: '2934 FE0F',
-    char: '⤴️',
-    name: 'right arrow curving up',
-    category: 'symbols',
-  },
-  {
-    codes: '2935 FE0F',
-    char: '⤵️',
-    name: 'right arrow curving down',
-    category: 'symbols',
-  },
-  {
-    codes: '1F503',
-    char: '🔃',
-    name: 'clockwise vertical arrows',
-    category: 'symbols',
-  },
-  {
-    codes: '1F504',
-    char: '🔄',
-    name: 'counterclockwise arrows button',
-    category: 'symbols',
-  },
-  {
-    codes: '1F519',
-    char: '🔙',
-    name: 'BACK arrow',
-    category: 'symbols',
-  },
-  {
-    codes: '1F51A',
-    char: '🔚',
-    name: 'END arrow',
-    category: 'symbols',
-  },
-  {
-    codes: '1F51B',
-    char: '🔛',
-    name: 'ON! arrow',
-    category: 'symbols',
-  },
-  {
-    codes: '1F51C',
-    char: '🔜',
-    name: 'SOON arrow',
-    category: 'symbols',
-  },
-  {
-    codes: '1F51D',
-    char: '🔝',
-    name: 'TOP arrow',
-    category: 'symbols',
-  },
-  { codes: '2648', char: '♈', name: 'Aries', category: 'symbols' },
-  { codes: '2649', char: '♉', name: 'Taurus', category: 'symbols' },
-  { codes: '264A', char: '♊', name: 'Gemini', category: 'symbols' },
-  { codes: '264B', char: '♋', name: 'Cancer', category: 'symbols' },
-  { codes: '264C', char: '♌', name: 'Leo', category: 'symbols' },
-  { codes: '264D', char: '♍', name: 'Virgo', category: 'symbols' },
-  { codes: '264E', char: '♎', name: 'Libra', category: 'symbols' },
-  { codes: '264F', char: '♏', name: 'Scorpio', category: 'symbols' },
-  {
-    codes: '2650',
-    char: '♐',
-    name: 'Sagittarius',
-    category: 'symbols',
-  },
-  { codes: '2651', char: '♑', name: 'Capricorn', category: 'symbols' },
-  { codes: '2652', char: '♒', name: 'Aquarius', category: 'symbols' },
-  { codes: '2653', char: '♓', name: 'Pisces', category: 'symbols' },
-  { codes: '26CE', char: '⛎', name: 'Ophiuchus', category: 'symbols' },
-  {
-    codes: '1F500',
-    char: '🔀',
-    name: 'shuffle tracks button',
-    category: 'symbols',
-  },
-  {
-    codes: '1F501',
-    char: '🔁',
-    name: 'repeat button',
-    category: 'symbols',
-  },
-  {
-    codes: '1F502',
-    char: '🔂',
-    name: 'repeat single button',
-    category: 'symbols',
-  },
-  {
-    codes: '23E9',
-    char: '⏩',
-    name: 'fast-forward button',
-    category: 'symbols',
-  },
-  {
-    codes: '23EA',
-    char: '⏪',
-    name: 'fast reverse button',
-    category: 'symbols',
-  },
-  {
-    codes: '1F53C',
-    char: '🔼',
-    name: 'upwards button',
-    category: 'symbols',
-  },
-  {
-    codes: '23EB',
-    char: '⏫',
-    name: 'fast up button',
-    category: 'symbols',
-  },
-  {
-    codes: '1F53D',
-    char: '🔽',
-    name: 'downwards button',
-    category: 'symbols',
-  },
-  {
-    codes: '23EC',
-    char: '⏬',
-    name: 'fast down button',
-    category: 'symbols',
-  },
-  { codes: '1F3A6', char: '🎦', name: 'cinema', category: 'symbols' },
-  {
-    codes: '1F505',
-    char: '🔅',
-    name: 'dim button',
-    category: 'symbols',
-  },
-  {
-    codes: '1F506',
-    char: '🔆',
-    name: 'bright button',
-    category: 'symbols',
-  },
-  {
-    codes: '1F4F6',
-    char: '📶',
-    name: 'antenna bars',
-    category: 'symbols',
-  },
-  {
-    codes: '1F4F3',
-    char: '📳',
-    name: 'vibration mode',
-    category: 'symbols',
-  },
-  {
-    codes: '1F4F4',
-    char: '📴',
-    name: 'mobile phone off',
-    category: 'symbols',
-  },
-  {
-    codes: '267B FE0F',
-    char: '♻️',
-    name: 'recycling symbol',
-    category: 'symbols',
-  },
-  {
-    codes: '1F531',
-    char: '🔱',
-    name: 'trident emblem',
-    category: 'symbols',
-  },
-  {
-    codes: '1F4DB',
-    char: '📛',
-    name: 'name badge',
-    category: 'symbols',
-  },
-  {
-    codes: '1F530',
-    char: '🔰',
-    name: 'Japanese symbol for beginner',
-    category: 'symbols',
-  },
-  {
-    codes: '2B55',
-    char: '⭕',
-    name: 'hollow red circle',
-    category: 'symbols',
-  },
-  {
-    codes: '2705',
-    char: '✅',
-    name: 'check mark button',
-    category: 'symbols',
-  },
-  {
-    codes: '2611 FE0F',
-    char: '☑️',
-    name: 'check box with check',
-    category: 'symbols',
-  },
-  {
-    codes: '2714 FE0F',
-    char: '✔️',
-    name: 'check mark',
-    category: 'symbols',
-  },
-  {
-    codes: '274C',
-    char: '❌',
-    name: 'cross mark',
-    category: 'symbols',
-  },
-  {
-    codes: '274E',
-    char: '❎',
-    name: 'cross mark button',
-    category: 'symbols',
-  },
-  {
-    codes: '27B0',
-    char: '➰',
-    name: 'curly loop',
-    category: 'symbols',
-  },
-  {
-    codes: '27BF',
-    char: '➿',
-    name: 'double curly loop',
-    category: 'symbols',
-  },
-  {
-    codes: '303D FE0F',
-    char: '〽️',
-    name: 'part alternation mark',
-    category: 'symbols',
-  },
-  {
-    codes: '303D',
-    char: '〽',
-    name: 'part alternation mark',
-    category: 'symbols',
-  },
-  {
-    codes: '2733 FE0F',
-    char: '✳️',
-    name: 'eight-spoked asterisk',
-    category: 'symbols',
-  },
-  {
-    codes: '2734 FE0F',
-    char: '✴️',
-    name: 'eight-pointed star',
-    category: 'symbols',
-  },
-  {
-    codes: '2747 FE0F',
-    char: '❇️',
-    name: 'sparkle',
-    category: 'symbols',
-  },
-  {
-    codes: '2753',
-    char: '❓',
-    name: 'question mark',
-    category: 'symbols',
-  },
-  {
-    codes: '2757',
-    char: '❗',
-    name: 'exclamation mark',
-    category: 'symbols',
-  },
-  {
-    codes: '1F534',
-    char: '🔴',
-    name: 'red circle',
-    category: 'symbols',
-  },
-  {
-    codes: '1F535',
-    char: '🔵',
-    name: 'blue circle',
-    category: 'symbols',
-  },
-  {
-    codes: '26AB',
-    char: '⚫',
-    name: 'black circle',
-    category: 'symbols',
-  },
-  {
-    codes: '26AA',
-    char: '⚪',
-    name: 'white circle',
-    category: 'symbols',
-  },
-  {
-    codes: '2B1B',
-    char: '⬛',
-    name: 'black large square',
-    category: 'symbols',
-  },
-  {
-    codes: '2B1C',
-    char: '⬜',
-    name: 'white large square',
-    category: 'symbols',
-  },
-  {
-    codes: '25FC FE0F',
-    char: '◼️',
-    name: 'black medium square',
-    category: 'symbols',
-  },
-  {
-    codes: '25FB FE0F',
-    char: '◻️',
-    name: 'white medium square',
-    category: 'symbols',
-  },
-  {
-    codes: '25FE',
-    char: '◾',
-    name: 'black medium-small square',
-    category: 'symbols',
-  },
-  {
-    codes: '25FD',
-    char: '◽',
-    name: 'white medium-small square',
-    category: 'symbols',
-  },
-  {
-    codes: '1F536',
-    char: '🔶',
-    name: 'large orange diamond',
-    category: 'symbols',
-  },
-  {
-    codes: '1F537',
-    char: '🔷',
-    name: 'large blue diamond',
-    category: 'symbols',
-  },
-  {
-    codes: '1F538',
-    char: '🔸',
-    name: 'small orange diamond',
-    category: 'symbols',
-  },
-  {
-    codes: '1F539',
-    char: '🔹',
-    name: 'small blue diamond',
-    category: 'symbols',
-  },
-  {
-    codes: '1F53A',
-    char: '🔺',
-    name: 'red triangle pointed up',
-    category: 'symbols',
-  },
-  {
-    codes: '1F53B',
-    char: '🔻',
-    name: 'red triangle pointed down',
-    category: 'symbols',
-  },
-  {
-    codes: '1F4A0',
-    char: '💠',
-    name: 'diamond with a dot',
-    category: 'symbols',
-  },
-  {
-    codes: '1F518',
-    char: '🔘',
-    name: 'radio button',
-    category: 'symbols',
-  },
-  {
-    codes: '1F533',
-    char: '🔳',
-    name: 'white square button',
-    category: 'symbols',
-  },
-  {
-    codes: '1F532',
-    char: '🔲',
-    name: 'black square button',
-    category: 'symbols',
-  },
-  {
-    codes: '1F3C1',
-    char: '🏁',
-    name: 'chequered flag',
-    category: 'flags',
-  },
-  {
-    codes: '1F6A9',
-    char: '🚩',
-    name: 'triangular flag',
-    category: 'flags',
-  },
-  {
-    codes: '1F38C',
-    char: '🎌',
-    name: 'crossed flags',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1E8',
-    char: '🇦🇨',
-    name: 'flag: Ascension Island',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1E9',
-    char: '🇦🇩',
-    name: 'flag: Andorra',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1EA',
-    char: '🇦🇪',
-    name: 'flag: United Arab Emirates',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1EB',
-    char: '🇦🇫',
-    name: 'flag: Afghanistan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1EC',
-    char: '🇦🇬',
-    name: 'flag: Antigua & Barbuda',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1EE',
-    char: '🇦🇮',
-    name: 'flag: Anguilla',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1F1',
-    char: '🇦🇱',
-    name: 'flag: Albania',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1F2',
-    char: '🇦🇲',
-    name: 'flag: Armenia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1F4',
-    char: '🇦🇴',
-    name: 'flag: Angola',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1F6',
-    char: '🇦🇶',
-    name: 'flag: Antarctica',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1F7',
-    char: '🇦🇷',
-    name: 'flag: Argentina',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1F8',
-    char: '🇦🇸',
-    name: 'flag: American Samoa',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1F9',
-    char: '🇦🇹',
-    name: 'flag: Austria',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1FA',
-    char: '🇦🇺',
-    name: 'flag: Australia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1FC',
-    char: '🇦🇼',
-    name: 'flag: Aruba',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1FD',
-    char: '🇦🇽',
-    name: 'flag: Åland Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E6 1F1FF',
-    char: '🇦🇿',
-    name: 'flag: Azerbaijan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1E6',
-    char: '🇧🇦',
-    name: 'flag: Bosnia & Herzegovina',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1E7',
-    char: '🇧🇧',
-    name: 'flag: Barbados',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1E9',
-    char: '🇧🇩',
-    name: 'flag: Bangladesh',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1EA',
-    char: '🇧🇪',
-    name: 'flag: Belgium',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1EB',
-    char: '🇧🇫',
-    name: 'flag: Burkina Faso',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1EC',
-    char: '🇧🇬',
-    name: 'flag: Bulgaria',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1ED',
-    char: '🇧🇭',
-    name: 'flag: Bahrain',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1EE',
-    char: '🇧🇮',
-    name: 'flag: Burundi',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1EF',
-    char: '🇧🇯',
-    name: 'flag: Benin',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1F1',
-    char: '🇧🇱',
-    name: 'flag: St. Barthélemy',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1F2',
-    char: '🇧🇲',
-    name: 'flag: Bermuda',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1F3',
-    char: '🇧🇳',
-    name: 'flag: Brunei',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1F4',
-    char: '🇧🇴',
-    name: 'flag: Bolivia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1F6',
-    char: '🇧🇶',
-    name: 'flag: Caribbean Netherlands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1F7',
-    char: '🇧🇷',
-    name: 'flag: Brazil',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1F8',
-    char: '🇧🇸',
-    name: 'flag: Bahamas',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1F9',
-    char: '🇧🇹',
-    name: 'flag: Bhutan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1FB',
-    char: '🇧🇻',
-    name: 'flag: Bouvet Island',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1FC',
-    char: '🇧🇼',
-    name: 'flag: Botswana',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1FE',
-    char: '🇧🇾',
-    name: 'flag: Belarus',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E7 1F1FF',
-    char: '🇧🇿',
-    name: 'flag: Belize',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1E6',
-    char: '🇨🇦',
-    name: 'flag: Canada',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1E8',
-    char: '🇨🇨',
-    name: 'flag: Cocos (Keeling) Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1E9',
-    char: '🇨🇩',
-    name: 'flag: Congo - Kinshasa',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1EB',
-    char: '🇨🇫',
-    name: 'flag: Central African Republic',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1EC',
-    char: '🇨🇬',
-    name: 'flag: Congo - Brazzaville',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1ED',
-    char: '🇨🇭',
-    name: 'flag: Switzerland',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1EE',
-    char: '🇨🇮',
-    name: 'flag: Côte d’Ivoire',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1F0',
-    char: '🇨🇰',
-    name: 'flag: Cook Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1F1',
-    char: '🇨🇱',
-    name: 'flag: Chile',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1F2',
-    char: '🇨🇲',
-    name: 'flag: Cameroon',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1F3',
-    char: '🇨🇳',
-    name: 'flag: China',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1F4',
-    char: '🇨🇴',
-    name: 'flag: Colombia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1F5',
-    char: '🇨🇵',
-    name: 'flag: Clipperton Island',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1F7',
-    char: '🇨🇷',
-    name: 'flag: Costa Rica',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1FA',
-    char: '🇨🇺',
-    name: 'flag: Cuba',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1FB',
-    char: '🇨🇻',
-    name: 'flag: Cape Verde',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1FC',
-    char: '🇨🇼',
-    name: 'flag: Curaçao',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1FD',
-    char: '🇨🇽',
-    name: 'flag: Christmas Island',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1FE',
-    char: '🇨🇾',
-    name: 'flag: Cyprus',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E8 1F1FF',
-    char: '🇨🇿',
-    name: 'flag: Czechia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E9 1F1EA',
-    char: '🇩🇪',
-    name: 'flag: Germany',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E9 1F1EC',
-    char: '🇩🇬',
-    name: 'flag: Diego Garcia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E9 1F1EF',
-    char: '🇩🇯',
-    name: 'flag: Djibouti',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E9 1F1F0',
-    char: '🇩🇰',
-    name: 'flag: Denmark',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E9 1F1F2',
-    char: '🇩🇲',
-    name: 'flag: Dominica',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E9 1F1F4',
-    char: '🇩🇴',
-    name: 'flag: Dominican Republic',
-    category: 'flags',
-  },
-  {
-    codes: '1F1E9 1F1FF',
-    char: '🇩🇿',
-    name: 'flag: Algeria',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EA 1F1E6',
-    char: '🇪🇦',
-    name: 'flag: Ceuta & Melilla',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EA 1F1E8',
-    char: '🇪🇨',
-    name: 'flag: Ecuador',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EA 1F1EA',
-    char: '🇪🇪',
-    name: 'flag: Estonia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EA 1F1EC',
-    char: '🇪🇬',
-    name: 'flag: Egypt',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EA 1F1ED',
-    char: '🇪🇭',
-    name: 'flag: Western Sahara',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EA 1F1F7',
-    char: '🇪🇷',
-    name: 'flag: Eritrea',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EA 1F1F8',
-    char: '🇪🇸',
-    name: 'flag: Spain',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EA 1F1F9',
-    char: '🇪🇹',
-    name: 'flag: Ethiopia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EA 1F1FA',
-    char: '🇪🇺',
-    name: 'flag: European Union',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EB 1F1EE',
-    char: '🇫🇮',
-    name: 'flag: Finland',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EB 1F1EF',
-    char: '🇫🇯',
-    name: 'flag: Fiji',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EB 1F1F0',
-    char: '🇫🇰',
-    name: 'flag: Falkland Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EB 1F1F2',
-    char: '🇫🇲',
-    name: 'flag: Micronesia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EB 1F1F4',
-    char: '🇫🇴',
-    name: 'flag: Faroe Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EB 1F1F7',
-    char: '🇫🇷',
-    name: 'flag: France',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1E6',
-    char: '🇬🇦',
-    name: 'flag: Gabon',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1E7',
-    char: '🇬🇧',
-    name: 'flag: United Kingdom',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1E9',
-    char: '🇬🇩',
-    name: 'flag: Grenada',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1EA',
-    char: '🇬🇪',
-    name: 'flag: Georgia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1EB',
-    char: '🇬🇫',
-    name: 'flag: French Guiana',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1EC',
-    char: '🇬🇬',
-    name: 'flag: Guernsey',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1ED',
-    char: '🇬🇭',
-    name: 'flag: Ghana',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1EE',
-    char: '🇬🇮',
-    name: 'flag: Gibraltar',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1F1',
-    char: '🇬🇱',
-    name: 'flag: Greenland',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1F2',
-    char: '🇬🇲',
-    name: 'flag: Gambia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1F3',
-    char: '🇬🇳',
-    name: 'flag: Guinea',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1F5',
-    char: '🇬🇵',
-    name: 'flag: Guadeloupe',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1F6',
-    char: '🇬🇶',
-    name: 'flag: Equatorial Guinea',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1F7',
-    char: '🇬🇷',
-    name: 'flag: Greece',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1F8',
-    char: '🇬🇸',
-    name: 'flag: South Georgia & South Sandwich Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1F9',
-    char: '🇬🇹',
-    name: 'flag: Guatemala',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1FA',
-    char: '🇬🇺',
-    name: 'flag: Guam',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1FC',
-    char: '🇬🇼',
-    name: 'flag: Guinea-Bissau',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EC 1F1FE',
-    char: '🇬🇾',
-    name: 'flag: Guyana',
-    category: 'flags',
-  },
-  {
-    codes: '1F1ED 1F1F0',
-    char: '🇭🇰',
-    name: 'flag: Hong Kong SAR China',
-    category: 'flags',
-  },
-  {
-    codes: '1F1ED 1F1F2',
-    char: '🇭🇲',
-    name: 'flag: Heard & McDonald Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1ED 1F1F3',
-    char: '🇭🇳',
-    name: 'flag: Honduras',
-    category: 'flags',
-  },
-  {
-    codes: '1F1ED 1F1F7',
-    char: '🇭🇷',
-    name: 'flag: Croatia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1ED 1F1F9',
-    char: '🇭🇹',
-    name: 'flag: Haiti',
-    category: 'flags',
-  },
-  {
-    codes: '1F1ED 1F1FA',
-    char: '🇭🇺',
-    name: 'flag: Hungary',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EE 1F1E8',
-    char: '🇮🇨',
-    name: 'flag: Canary Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EE 1F1E9',
-    char: '🇮🇩',
-    name: 'flag: Indonesia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EE 1F1EA',
-    char: '🇮🇪',
-    name: 'flag: Ireland',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EE 1F1F1',
-    char: '🇮🇱',
-    name: 'flag: Israel',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EE 1F1F2',
-    char: '🇮🇲',
-    name: 'flag: Isle of Man',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EE 1F1F3',
-    char: '🇮🇳',
-    name: 'flag: India',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EE 1F1F4',
-    char: '🇮🇴',
-    name: 'flag: British Indian Ocean Territory',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EE 1F1F6',
-    char: '🇮🇶',
-    name: 'flag: Iraq',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EE 1F1F7',
-    char: '🇮🇷',
-    name: 'flag: Iran',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EE 1F1F8',
-    char: '🇮🇸',
-    name: 'flag: Iceland',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EE 1F1F9',
-    char: '🇮🇹',
-    name: 'flag: Italy',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EF 1F1EA',
-    char: '🇯🇪',
-    name: 'flag: Jersey',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EF 1F1F2',
-    char: '🇯🇲',
-    name: 'flag: Jamaica',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EF 1F1F4',
-    char: '🇯🇴',
-    name: 'flag: Jordan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1EF 1F1F5',
-    char: '🇯🇵',
-    name: 'flag: Japan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F0 1F1EA',
-    char: '🇰🇪',
-    name: 'flag: Kenya',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F0 1F1EC',
-    char: '🇰🇬',
-    name: 'flag: Kyrgyzstan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F0 1F1ED',
-    char: '🇰🇭',
-    name: 'flag: Cambodia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F0 1F1EE',
-    char: '🇰🇮',
-    name: 'flag: Kiribati',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F0 1F1F2',
-    char: '🇰🇲',
-    name: 'flag: Comoros',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F0 1F1F3',
-    char: '🇰🇳',
-    name: 'flag: St. Kitts & Nevis',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F0 1F1F5',
-    char: '🇰🇵',
-    name: 'flag: North Korea',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F0 1F1F7',
-    char: '🇰🇷',
-    name: 'flag: South Korea',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F0 1F1FC',
-    char: '🇰🇼',
-    name: 'flag: Kuwait',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F0 1F1FE',
-    char: '🇰🇾',
-    name: 'flag: Cayman Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F0 1F1FF',
-    char: '🇰🇿',
-    name: 'flag: Kazakhstan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F1 1F1E6',
-    char: '🇱🇦',
-    name: 'flag: Laos',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F1 1F1E7',
-    char: '🇱🇧',
-    name: 'flag: Lebanon',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F1 1F1E8',
-    char: '🇱🇨',
-    name: 'flag: St. Lucia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F1 1F1EE',
-    char: '🇱🇮',
-    name: 'flag: Liechtenstein',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F1 1F1F0',
-    char: '🇱🇰',
-    name: 'flag: Sri Lanka',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F1 1F1F7',
-    char: '🇱🇷',
-    name: 'flag: Liberia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F1 1F1F8',
-    char: '🇱🇸',
-    name: 'flag: Lesotho',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F1 1F1F9',
-    char: '🇱🇹',
-    name: 'flag: Lithuania',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F1 1F1FA',
-    char: '🇱🇺',
-    name: 'flag: Luxembourg',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F1 1F1FB',
-    char: '🇱🇻',
-    name: 'flag: Latvia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F1 1F1FE',
-    char: '🇱🇾',
-    name: 'flag: Libya',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1E6',
-    char: '🇲🇦',
-    name: 'flag: Morocco',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1E8',
-    char: '🇲🇨',
-    name: 'flag: Monaco',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1E9',
-    char: '🇲🇩',
-    name: 'flag: Moldova',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1EA',
-    char: '🇲🇪',
-    name: 'flag: Montenegro',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1EB',
-    char: '🇲🇫',
-    name: 'flag: St. Martin',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1EC',
-    char: '🇲🇬',
-    name: 'flag: Madagascar',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1ED',
-    char: '🇲🇭',
-    name: 'flag: Marshall Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1F0',
-    char: '🇲🇰',
-    name: 'flag: Macedonia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1F1',
-    char: '🇲🇱',
-    name: 'flag: Mali',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1F2',
-    char: '🇲🇲',
-    name: 'flag: Myanmar (Burma)',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1F3',
-    char: '🇲🇳',
-    name: 'flag: Mongolia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1F4',
-    char: '🇲🇴',
-    name: 'flag: Macao SAR China',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1F5',
-    char: '🇲🇵',
-    name: 'flag: Northern Mariana Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1F6',
-    char: '🇲🇶',
-    name: 'flag: Martinique',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1F7',
-    char: '🇲🇷',
-    name: 'flag: Mauritania',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1F8',
-    char: '🇲🇸',
-    name: 'flag: Montserrat',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1F9',
-    char: '🇲🇹',
-    name: 'flag: Malta',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1FA',
-    char: '🇲🇺',
-    name: 'flag: Mauritius',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1FB',
-    char: '🇲🇻',
-    name: 'flag: Maldives',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1FC',
-    char: '🇲🇼',
-    name: 'flag: Malawi',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1FD',
-    char: '🇲🇽',
-    name: 'flag: Mexico',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1FE',
-    char: '🇲🇾',
-    name: 'flag: Malaysia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F2 1F1FF',
-    char: '🇲🇿',
-    name: 'flag: Mozambique',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F3 1F1E6',
-    char: '🇳🇦',
-    name: 'flag: Namibia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F3 1F1E8',
-    char: '🇳🇨',
-    name: 'flag: New Caledonia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F3 1F1EA',
-    char: '🇳🇪',
-    name: 'flag: Niger',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F3 1F1EB',
-    char: '🇳🇫',
-    name: 'flag: Norfolk Island',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F3 1F1EC',
-    char: '🇳🇬',
-    name: 'flag: Nigeria',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F3 1F1EE',
-    char: '🇳🇮',
-    name: 'flag: Nicaragua',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F3 1F1F1',
-    char: '🇳🇱',
-    name: 'flag: Netherlands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F3 1F1F4',
-    char: '🇳🇴',
-    name: 'flag: Norway',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F3 1F1F5',
-    char: '🇳🇵',
-    name: 'flag: Nepal',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F3 1F1F7',
-    char: '🇳🇷',
-    name: 'flag: Nauru',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F3 1F1FA',
-    char: '🇳🇺',
-    name: 'flag: Niue',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F3 1F1FF',
-    char: '🇳🇿',
-    name: 'flag: New Zealand',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F4 1F1F2',
-    char: '🇴🇲',
-    name: 'flag: Oman',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1E6',
-    char: '🇵🇦',
-    name: 'flag: Panama',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1EA',
-    char: '🇵🇪',
-    name: 'flag: Peru',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1EB',
-    char: '🇵🇫',
-    name: 'flag: French Polynesia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1EC',
-    char: '🇵🇬',
-    name: 'flag: Papua New Guinea',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1ED',
-    char: '🇵🇭',
-    name: 'flag: Philippines',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1F0',
-    char: '🇵🇰',
-    name: 'flag: Pakistan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1F1',
-    char: '🇵🇱',
-    name: 'flag: Poland',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1F2',
-    char: '🇵🇲',
-    name: 'flag: St. Pierre & Miquelon',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1F3',
-    char: '🇵🇳',
-    name: 'flag: Pitcairn Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1F7',
-    char: '🇵🇷',
-    name: 'flag: Puerto Rico',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1F8',
-    char: '🇵🇸',
-    name: 'flag: Palestinian Territories',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1F9',
-    char: '🇵🇹',
-    name: 'flag: Portugal',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1FC',
-    char: '🇵🇼',
-    name: 'flag: Palau',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F5 1F1FE',
-    char: '🇵🇾',
-    name: 'flag: Paraguay',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F6 1F1E6',
-    char: '🇶🇦',
-    name: 'flag: Qatar',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F7 1F1EA',
-    char: '🇷🇪',
-    name: 'flag: Réunion',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F7 1F1F4',
-    char: '🇷🇴',
-    name: 'flag: Romania',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F7 1F1F8',
-    char: '🇷🇸',
-    name: 'flag: Serbia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F7 1F1FA',
-    char: '🇷🇺',
-    name: 'flag: Russia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F7 1F1FC',
-    char: '🇷🇼',
-    name: 'flag: Rwanda',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1E6',
-    char: '🇸🇦',
-    name: 'flag: Saudi Arabia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1E7',
-    char: '🇸🇧',
-    name: 'flag: Solomon Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1E8',
-    char: '🇸🇨',
-    name: 'flag: Seychelles',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1E9',
-    char: '🇸🇩',
-    name: 'flag: Sudan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1EA',
-    char: '🇸🇪',
-    name: 'flag: Sweden',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1EC',
-    char: '🇸🇬',
-    name: 'flag: Singapore',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1ED',
-    char: '🇸🇭',
-    name: 'flag: St. Helena',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1EE',
-    char: '🇸🇮',
-    name: 'flag: Slovenia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1EF',
-    char: '🇸🇯',
-    name: 'flag: Svalbard & Jan Mayen',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1F0',
-    char: '🇸🇰',
-    name: 'flag: Slovakia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1F1',
-    char: '🇸🇱',
-    name: 'flag: Sierra Leone',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1F2',
-    char: '🇸🇲',
-    name: 'flag: San Marino',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1F3',
-    char: '🇸🇳',
-    name: 'flag: Senegal',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1F4',
-    char: '🇸🇴',
-    name: 'flag: Somalia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1F7',
-    char: '🇸🇷',
-    name: 'flag: Suriname',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1F8',
-    char: '🇸🇸',
-    name: 'flag: South Sudan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1F9',
-    char: '🇸🇹',
-    name: 'flag: São Tomé & Príncipe',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1FB',
-    char: '🇸🇻',
-    name: 'flag: El Salvador',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1FD',
-    char: '🇸🇽',
-    name: 'flag: Sint Maarten',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1FE',
-    char: '🇸🇾',
-    name: 'flag: Syria',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F8 1F1FF',
-    char: '🇸🇿',
-    name: 'flag: Eswatini',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1E6',
-    char: '🇹🇦',
-    name: 'flag: Tristan da Cunha',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1E8',
-    char: '🇹🇨',
-    name: 'flag: Turks & Caicos Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1E9',
-    char: '🇹🇩',
-    name: 'flag: Chad',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1EB',
-    char: '🇹🇫',
-    name: 'flag: French Southern Territories',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1EC',
-    char: '🇹🇬',
-    name: 'flag: Togo',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1ED',
-    char: '🇹🇭',
-    name: 'flag: Thailand',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1EF',
-    char: '🇹🇯',
-    name: 'flag: Tajikistan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1F0',
-    char: '🇹🇰',
-    name: 'flag: Tokelau',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1F1',
-    char: '🇹🇱',
-    name: 'flag: Timor-Leste',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1F2',
-    char: '🇹🇲',
-    name: 'flag: Turkmenistan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1F3',
-    char: '🇹🇳',
-    name: 'flag: Tunisia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1F4',
-    char: '🇹🇴',
-    name: 'flag: Tonga',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1F7',
-    char: '🇹🇷',
-    name: 'flag: Turkey',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1F9',
-    char: '🇹🇹',
-    name: 'flag: Trinidad & Tobago',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1FB',
-    char: '🇹🇻',
-    name: 'flag: Tuvalu',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1FC',
-    char: '🇹🇼',
-    name: 'flag: Taiwan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1F9 1F1FF',
-    char: '🇹🇿',
-    name: 'flag: Tanzania',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FA 1F1E6',
-    char: '🇺🇦',
-    name: 'flag: Ukraine',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FA 1F1EC',
-    char: '🇺🇬',
-    name: 'flag: Uganda',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FA 1F1F2',
-    char: '🇺🇲',
-    name: 'flag: U.S. Outlying Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FA 1F1F3',
-    char: '🇺🇳',
-    name: 'flag: United Nations',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FA 1F1F8',
-    char: '🇺🇸',
-    name: 'flag: United States',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FA 1F1FE',
-    char: '🇺🇾',
-    name: 'flag: Uruguay',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FA 1F1FF',
-    char: '🇺🇿',
-    name: 'flag: Uzbekistan',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FB 1F1E6',
-    char: '🇻🇦',
-    name: 'flag: Vatican City',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FB 1F1E8',
-    char: '🇻🇨',
-    name: 'flag: St. Vincent & Grenadines',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FB 1F1EA',
-    char: '🇻🇪',
-    name: 'flag: Venezuela',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FB 1F1EC',
-    char: '🇻🇬',
-    name: 'flag: British Virgin Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FB 1F1EE',
-    char: '🇻🇮',
-    name: 'flag: U.S. Virgin Islands',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FB 1F1F3',
-    char: '🇻🇳',
-    name: 'flag: Vietnam',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FB 1F1FA',
-    char: '🇻🇺',
-    name: 'flag: Vanuatu',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FC 1F1EB',
-    char: '🇼🇫',
-    name: 'flag: Wallis & Futuna',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FC 1F1F8',
-    char: '🇼🇸',
-    name: 'flag: Samoa',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FD 1F1F0',
-    char: '🇽🇰',
-    name: 'flag: Kosovo',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FE 1F1EA',
-    char: '🇾🇪',
-    name: 'flag: Yemen',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FE 1F1F9',
-    char: '🇾🇹',
-    name: 'flag: Mayotte',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FF 1F1E6',
-    char: '🇿🇦',
-    name: 'flag: South Africa',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FF 1F1F2',
-    char: '🇿🇲',
-    name: 'flag: Zambia',
-    category: 'flags',
-  },
-  {
-    codes: '1F1FF 1F1FC',
-    char: '🇿🇼',
-    name: 'flag: Zimbabwe',
-    category: 'flags',
-  },
-];
