+++ emojis/flags.js
@@ -1,0 +1,1568 @@
+export default [
+  {
+    codes: '1F3C1',
+    char: '🏁',
+    name: 'chequered flag',
+    category: 'flags',
+  },
+  {
+    codes: '1F6A9',
+    char: '🚩',
+    name: 'triangular flag',
+    category: 'flags',
+  },
+  {
+    codes: '1F38C',
+    char: '🎌',
+    name: 'crossed flags',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1E8',
+    char: '🇦🇨',
+    name: 'flag: Ascension Island',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1E9',
+    char: '🇦🇩',
+    name: 'flag: Andorra',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1EA',
+    char: '🇦🇪',
+    name: 'flag: United Arab Emirates',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1EB',
+    char: '🇦🇫',
+    name: 'flag: Afghanistan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1EC',
+    char: '🇦🇬',
+    name: 'flag: Antigua & Barbuda',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1EE',
+    char: '🇦🇮',
+    name: 'flag: Anguilla',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1F1',
+    char: '🇦🇱',
+    name: 'flag: Albania',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1F2',
+    char: '🇦🇲',
+    name: 'flag: Armenia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1F4',
+    char: '🇦🇴',
+    name: 'flag: Angola',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1F6',
+    char: '🇦🇶',
+    name: 'flag: Antarctica',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1F7',
+    char: '🇦🇷',
+    name: 'flag: Argentina',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1F8',
+    char: '🇦🇸',
+    name: 'flag: American Samoa',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1F9',
+    char: '🇦🇹',
+    name: 'flag: Austria',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1FA',
+    char: '🇦🇺',
+    name: 'flag: Australia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1FC',
+    char: '🇦🇼',
+    name: 'flag: Aruba',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1FD',
+    char: '🇦🇽',
+    name: 'flag: Åland Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E6 1F1FF',
+    char: '🇦🇿',
+    name: 'flag: Azerbaijan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1E6',
+    char: '🇧🇦',
+    name: 'flag: Bosnia & Herzegovina',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1E7',
+    char: '🇧🇧',
+    name: 'flag: Barbados',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1E9',
+    char: '🇧🇩',
+    name: 'flag: Bangladesh',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1EA',
+    char: '🇧🇪',
+    name: 'flag: Belgium',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1EB',
+    char: '🇧🇫',
+    name: 'flag: Burkina Faso',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1EC',
+    char: '🇧🇬',
+    name: 'flag: Bulgaria',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1ED',
+    char: '🇧🇭',
+    name: 'flag: Bahrain',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1EE',
+    char: '🇧🇮',
+    name: 'flag: Burundi',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1EF',
+    char: '🇧🇯',
+    name: 'flag: Benin',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1F1',
+    char: '🇧🇱',
+    name: 'flag: St. Barthélemy',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1F2',
+    char: '🇧🇲',
+    name: 'flag: Bermuda',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1F3',
+    char: '🇧🇳',
+    name: 'flag: Brunei',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1F4',
+    char: '🇧🇴',
+    name: 'flag: Bolivia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1F6',
+    char: '🇧🇶',
+    name: 'flag: Caribbean Netherlands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1F7',
+    char: '🇧🇷',
+    name: 'flag: Brazil',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1F8',
+    char: '🇧🇸',
+    name: 'flag: Bahamas',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1F9',
+    char: '🇧🇹',
+    name: 'flag: Bhutan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1FB',
+    char: '🇧🇻',
+    name: 'flag: Bouvet Island',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1FC',
+    char: '🇧🇼',
+    name: 'flag: Botswana',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1FE',
+    char: '🇧🇾',
+    name: 'flag: Belarus',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E7 1F1FF',
+    char: '🇧🇿',
+    name: 'flag: Belize',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1E6',
+    char: '🇨🇦',
+    name: 'flag: Canada',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1E8',
+    char: '🇨🇨',
+    name: 'flag: Cocos (Keeling) Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1E9',
+    char: '🇨🇩',
+    name: 'flag: Congo - Kinshasa',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1EB',
+    char: '🇨🇫',
+    name: 'flag: Central African Republic',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1EC',
+    char: '🇨🇬',
+    name: 'flag: Congo - Brazzaville',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1ED',
+    char: '🇨🇭',
+    name: 'flag: Switzerland',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1EE',
+    char: '🇨🇮',
+    name: 'flag: Côte d’Ivoire',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1F0',
+    char: '🇨🇰',
+    name: 'flag: Cook Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1F1',
+    char: '🇨🇱',
+    name: 'flag: Chile',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1F2',
+    char: '🇨🇲',
+    name: 'flag: Cameroon',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1F3',
+    char: '🇨🇳',
+    name: 'flag: China',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1F4',
+    char: '🇨🇴',
+    name: 'flag: Colombia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1F5',
+    char: '🇨🇵',
+    name: 'flag: Clipperton Island',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1F7',
+    char: '🇨🇷',
+    name: 'flag: Costa Rica',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1FA',
+    char: '🇨🇺',
+    name: 'flag: Cuba',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1FB',
+    char: '🇨🇻',
+    name: 'flag: Cape Verde',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1FC',
+    char: '🇨🇼',
+    name: 'flag: Curaçao',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1FD',
+    char: '🇨🇽',
+    name: 'flag: Christmas Island',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1FE',
+    char: '🇨🇾',
+    name: 'flag: Cyprus',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E8 1F1FF',
+    char: '🇨🇿',
+    name: 'flag: Czechia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E9 1F1EA',
+    char: '🇩🇪',
+    name: 'flag: Germany',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E9 1F1EC',
+    char: '🇩🇬',
+    name: 'flag: Diego Garcia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E9 1F1EF',
+    char: '🇩🇯',
+    name: 'flag: Djibouti',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E9 1F1F0',
+    char: '🇩🇰',
+    name: 'flag: Denmark',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E9 1F1F2',
+    char: '🇩🇲',
+    name: 'flag: Dominica',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E9 1F1F4',
+    char: '🇩🇴',
+    name: 'flag: Dominican Republic',
+    category: 'flags',
+  },
+  {
+    codes: '1F1E9 1F1FF',
+    char: '🇩🇿',
+    name: 'flag: Algeria',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EA 1F1E6',
+    char: '🇪🇦',
+    name: 'flag: Ceuta & Melilla',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EA 1F1E8',
+    char: '🇪🇨',
+    name: 'flag: Ecuador',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EA 1F1EA',
+    char: '🇪🇪',
+    name: 'flag: Estonia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EA 1F1EC',
+    char: '🇪🇬',
+    name: 'flag: Egypt',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EA 1F1ED',
+    char: '🇪🇭',
+    name: 'flag: Western Sahara',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EA 1F1F7',
+    char: '🇪🇷',
+    name: 'flag: Eritrea',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EA 1F1F8',
+    char: '🇪🇸',
+    name: 'flag: Spain',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EA 1F1F9',
+    char: '🇪🇹',
+    name: 'flag: Ethiopia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EA 1F1FA',
+    char: '🇪🇺',
+    name: 'flag: European Union',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EB 1F1EE',
+    char: '🇫🇮',
+    name: 'flag: Finland',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EB 1F1EF',
+    char: '🇫🇯',
+    name: 'flag: Fiji',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EB 1F1F0',
+    char: '🇫🇰',
+    name: 'flag: Falkland Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EB 1F1F2',
+    char: '🇫🇲',
+    name: 'flag: Micronesia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EB 1F1F4',
+    char: '🇫🇴',
+    name: 'flag: Faroe Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EB 1F1F7',
+    char: '🇫🇷',
+    name: 'flag: France',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1E6',
+    char: '🇬🇦',
+    name: 'flag: Gabon',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1E7',
+    char: '🇬🇧',
+    name: 'flag: United Kingdom',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1E9',
+    char: '🇬🇩',
+    name: 'flag: Grenada',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1EA',
+    char: '🇬🇪',
+    name: 'flag: Georgia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1EB',
+    char: '🇬🇫',
+    name: 'flag: French Guiana',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1EC',
+    char: '🇬🇬',
+    name: 'flag: Guernsey',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1ED',
+    char: '🇬🇭',
+    name: 'flag: Ghana',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1EE',
+    char: '🇬🇮',
+    name: 'flag: Gibraltar',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1F1',
+    char: '🇬🇱',
+    name: 'flag: Greenland',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1F2',
+    char: '🇬🇲',
+    name: 'flag: Gambia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1F3',
+    char: '🇬🇳',
+    name: 'flag: Guinea',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1F5',
+    char: '🇬🇵',
+    name: 'flag: Guadeloupe',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1F6',
+    char: '🇬🇶',
+    name: 'flag: Equatorial Guinea',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1F7',
+    char: '🇬🇷',
+    name: 'flag: Greece',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1F8',
+    char: '🇬🇸',
+    name: 'flag: South Georgia & South Sandwich Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1F9',
+    char: '🇬🇹',
+    name: 'flag: Guatemala',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1FA',
+    char: '🇬🇺',
+    name: 'flag: Guam',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1FC',
+    char: '🇬🇼',
+    name: 'flag: Guinea-Bissau',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EC 1F1FE',
+    char: '🇬🇾',
+    name: 'flag: Guyana',
+    category: 'flags',
+  },
+  {
+    codes: '1F1ED 1F1F0',
+    char: '🇭🇰',
+    name: 'flag: Hong Kong SAR China',
+    category: 'flags',
+  },
+  {
+    codes: '1F1ED 1F1F2',
+    char: '🇭🇲',
+    name: 'flag: Heard & McDonald Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1ED 1F1F3',
+    char: '🇭🇳',
+    name: 'flag: Honduras',
+    category: 'flags',
+  },
+  {
+    codes: '1F1ED 1F1F7',
+    char: '🇭🇷',
+    name: 'flag: Croatia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1ED 1F1F9',
+    char: '🇭🇹',
+    name: 'flag: Haiti',
+    category: 'flags',
+  },
+  {
+    codes: '1F1ED 1F1FA',
+    char: '🇭🇺',
+    name: 'flag: Hungary',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EE 1F1E8',
+    char: '🇮🇨',
+    name: 'flag: Canary Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EE 1F1E9',
+    char: '🇮🇩',
+    name: 'flag: Indonesia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EE 1F1EA',
+    char: '🇮🇪',
+    name: 'flag: Ireland',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EE 1F1F1',
+    char: '🇮🇱',
+    name: 'flag: Israel',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EE 1F1F2',
+    char: '🇮🇲',
+    name: 'flag: Isle of Man',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EE 1F1F3',
+    char: '🇮🇳',
+    name: 'flag: India',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EE 1F1F4',
+    char: '🇮🇴',
+    name: 'flag: British Indian Ocean Territory',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EE 1F1F6',
+    char: '🇮🇶',
+    name: 'flag: Iraq',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EE 1F1F7',
+    char: '🇮🇷',
+    name: 'flag: Iran',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EE 1F1F8',
+    char: '🇮🇸',
+    name: 'flag: Iceland',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EE 1F1F9',
+    char: '🇮🇹',
+    name: 'flag: Italy',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EF 1F1EA',
+    char: '🇯🇪',
+    name: 'flag: Jersey',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EF 1F1F2',
+    char: '🇯🇲',
+    name: 'flag: Jamaica',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EF 1F1F4',
+    char: '🇯🇴',
+    name: 'flag: Jordan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1EF 1F1F5',
+    char: '🇯🇵',
+    name: 'flag: Japan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F0 1F1EA',
+    char: '🇰🇪',
+    name: 'flag: Kenya',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F0 1F1EC',
+    char: '🇰🇬',
+    name: 'flag: Kyrgyzstan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F0 1F1ED',
+    char: '🇰🇭',
+    name: 'flag: Cambodia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F0 1F1EE',
+    char: '🇰🇮',
+    name: 'flag: Kiribati',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F0 1F1F2',
+    char: '🇰🇲',
+    name: 'flag: Comoros',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F0 1F1F3',
+    char: '🇰🇳',
+    name: 'flag: St. Kitts & Nevis',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F0 1F1F5',
+    char: '🇰🇵',
+    name: 'flag: North Korea',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F0 1F1F7',
+    char: '🇰🇷',
+    name: 'flag: South Korea',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F0 1F1FC',
+    char: '🇰🇼',
+    name: 'flag: Kuwait',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F0 1F1FE',
+    char: '🇰🇾',
+    name: 'flag: Cayman Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F0 1F1FF',
+    char: '🇰🇿',
+    name: 'flag: Kazakhstan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F1 1F1E6',
+    char: '🇱🇦',
+    name: 'flag: Laos',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F1 1F1E7',
+    char: '🇱🇧',
+    name: 'flag: Lebanon',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F1 1F1E8',
+    char: '🇱🇨',
+    name: 'flag: St. Lucia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F1 1F1EE',
+    char: '🇱🇮',
+    name: 'flag: Liechtenstein',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F1 1F1F0',
+    char: '🇱🇰',
+    name: 'flag: Sri Lanka',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F1 1F1F7',
+    char: '🇱🇷',
+    name: 'flag: Liberia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F1 1F1F8',
+    char: '🇱🇸',
+    name: 'flag: Lesotho',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F1 1F1F9',
+    char: '🇱🇹',
+    name: 'flag: Lithuania',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F1 1F1FA',
+    char: '🇱🇺',
+    name: 'flag: Luxembourg',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F1 1F1FB',
+    char: '🇱🇻',
+    name: 'flag: Latvia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F1 1F1FE',
+    char: '🇱🇾',
+    name: 'flag: Libya',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1E6',
+    char: '🇲🇦',
+    name: 'flag: Morocco',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1E8',
+    char: '🇲🇨',
+    name: 'flag: Monaco',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1E9',
+    char: '🇲🇩',
+    name: 'flag: Moldova',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1EA',
+    char: '🇲🇪',
+    name: 'flag: Montenegro',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1EB',
+    char: '🇲🇫',
+    name: 'flag: St. Martin',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1EC',
+    char: '🇲🇬',
+    name: 'flag: Madagascar',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1ED',
+    char: '🇲🇭',
+    name: 'flag: Marshall Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1F0',
+    char: '🇲🇰',
+    name: 'flag: Macedonia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1F1',
+    char: '🇲🇱',
+    name: 'flag: Mali',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1F2',
+    char: '🇲🇲',
+    name: 'flag: Myanmar (Burma)',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1F3',
+    char: '🇲🇳',
+    name: 'flag: Mongolia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1F4',
+    char: '🇲🇴',
+    name: 'flag: Macao SAR China',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1F5',
+    char: '🇲🇵',
+    name: 'flag: Northern Mariana Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1F6',
+    char: '🇲🇶',
+    name: 'flag: Martinique',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1F7',
+    char: '🇲🇷',
+    name: 'flag: Mauritania',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1F8',
+    char: '🇲🇸',
+    name: 'flag: Montserrat',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1F9',
+    char: '🇲🇹',
+    name: 'flag: Malta',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1FA',
+    char: '🇲🇺',
+    name: 'flag: Mauritius',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1FB',
+    char: '🇲🇻',
+    name: 'flag: Maldives',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1FC',
+    char: '🇲🇼',
+    name: 'flag: Malawi',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1FD',
+    char: '🇲🇽',
+    name: 'flag: Mexico',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1FE',
+    char: '🇲🇾',
+    name: 'flag: Malaysia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F2 1F1FF',
+    char: '🇲🇿',
+    name: 'flag: Mozambique',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F3 1F1E6',
+    char: '🇳🇦',
+    name: 'flag: Namibia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F3 1F1E8',
+    char: '🇳🇨',
+    name: 'flag: New Caledonia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F3 1F1EA',
+    char: '🇳🇪',
+    name: 'flag: Niger',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F3 1F1EB',
+    char: '🇳🇫',
+    name: 'flag: Norfolk Island',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F3 1F1EC',
+    char: '🇳🇬',
+    name: 'flag: Nigeria',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F3 1F1EE',
+    char: '🇳🇮',
+    name: 'flag: Nicaragua',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F3 1F1F1',
+    char: '🇳🇱',
+    name: 'flag: Netherlands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F3 1F1F4',
+    char: '🇳🇴',
+    name: 'flag: Norway',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F3 1F1F5',
+    char: '🇳🇵',
+    name: 'flag: Nepal',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F3 1F1F7',
+    char: '🇳🇷',
+    name: 'flag: Nauru',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F3 1F1FA',
+    char: '🇳🇺',
+    name: 'flag: Niue',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F3 1F1FF',
+    char: '🇳🇿',
+    name: 'flag: New Zealand',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F4 1F1F2',
+    char: '🇴🇲',
+    name: 'flag: Oman',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1E6',
+    char: '🇵🇦',
+    name: 'flag: Panama',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1EA',
+    char: '🇵🇪',
+    name: 'flag: Peru',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1EB',
+    char: '🇵🇫',
+    name: 'flag: French Polynesia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1EC',
+    char: '🇵🇬',
+    name: 'flag: Papua New Guinea',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1ED',
+    char: '🇵🇭',
+    name: 'flag: Philippines',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1F0',
+    char: '🇵🇰',
+    name: 'flag: Pakistan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1F1',
+    char: '🇵🇱',
+    name: 'flag: Poland',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1F2',
+    char: '🇵🇲',
+    name: 'flag: St. Pierre & Miquelon',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1F3',
+    char: '🇵🇳',
+    name: 'flag: Pitcairn Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1F7',
+    char: '🇵🇷',
+    name: 'flag: Puerto Rico',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1F8',
+    char: '🇵🇸',
+    name: 'flag: Palestinian Territories',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1F9',
+    char: '🇵🇹',
+    name: 'flag: Portugal',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1FC',
+    char: '🇵🇼',
+    name: 'flag: Palau',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F5 1F1FE',
+    char: '🇵🇾',
+    name: 'flag: Paraguay',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F6 1F1E6',
+    char: '🇶🇦',
+    name: 'flag: Qatar',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F7 1F1EA',
+    char: '🇷🇪',
+    name: 'flag: Réunion',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F7 1F1F4',
+    char: '🇷🇴',
+    name: 'flag: Romania',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F7 1F1F8',
+    char: '🇷🇸',
+    name: 'flag: Serbia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F7 1F1FA',
+    char: '🇷🇺',
+    name: 'flag: Russia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F7 1F1FC',
+    char: '🇷🇼',
+    name: 'flag: Rwanda',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1E6',
+    char: '🇸🇦',
+    name: 'flag: Saudi Arabia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1E7',
+    char: '🇸🇧',
+    name: 'flag: Solomon Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1E8',
+    char: '🇸🇨',
+    name: 'flag: Seychelles',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1E9',
+    char: '🇸🇩',
+    name: 'flag: Sudan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1EA',
+    char: '🇸🇪',
+    name: 'flag: Sweden',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1EC',
+    char: '🇸🇬',
+    name: 'flag: Singapore',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1ED',
+    char: '🇸🇭',
+    name: 'flag: St. Helena',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1EE',
+    char: '🇸🇮',
+    name: 'flag: Slovenia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1EF',
+    char: '🇸🇯',
+    name: 'flag: Svalbard & Jan Mayen',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1F0',
+    char: '🇸🇰',
+    name: 'flag: Slovakia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1F1',
+    char: '🇸🇱',
+    name: 'flag: Sierra Leone',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1F2',
+    char: '🇸🇲',
+    name: 'flag: San Marino',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1F3',
+    char: '🇸🇳',
+    name: 'flag: Senegal',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1F4',
+    char: '🇸🇴',
+    name: 'flag: Somalia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1F7',
+    char: '🇸🇷',
+    name: 'flag: Suriname',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1F8',
+    char: '🇸🇸',
+    name: 'flag: South Sudan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1F9',
+    char: '🇸🇹',
+    name: 'flag: São Tomé & Príncipe',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1FB',
+    char: '🇸🇻',
+    name: 'flag: El Salvador',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1FD',
+    char: '🇸🇽',
+    name: 'flag: Sint Maarten',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1FE',
+    char: '🇸🇾',
+    name: 'flag: Syria',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F8 1F1FF',
+    char: '🇸🇿',
+    name: 'flag: Eswatini',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1E6',
+    char: '🇹🇦',
+    name: 'flag: Tristan da Cunha',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1E8',
+    char: '🇹🇨',
+    name: 'flag: Turks & Caicos Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1E9',
+    char: '🇹🇩',
+    name: 'flag: Chad',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1EB',
+    char: '🇹🇫',
+    name: 'flag: French Southern Territories',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1EC',
+    char: '🇹🇬',
+    name: 'flag: Togo',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1ED',
+    char: '🇹🇭',
+    name: 'flag: Thailand',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1EF',
+    char: '🇹🇯',
+    name: 'flag: Tajikistan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1F0',
+    char: '🇹🇰',
+    name: 'flag: Tokelau',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1F1',
+    char: '🇹🇱',
+    name: 'flag: Timor-Leste',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1F2',
+    char: '🇹🇲',
+    name: 'flag: Turkmenistan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1F3',
+    char: '🇹🇳',
+    name: 'flag: Tunisia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1F4',
+    char: '🇹🇴',
+    name: 'flag: Tonga',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1F7',
+    char: '🇹🇷',
+    name: 'flag: Turkey',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1F9',
+    char: '🇹🇹',
+    name: 'flag: Trinidad & Tobago',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1FB',
+    char: '🇹🇻',
+    name: 'flag: Tuvalu',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1FC',
+    char: '🇹🇼',
+    name: 'flag: Taiwan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1F9 1F1FF',
+    char: '🇹🇿',
+    name: 'flag: Tanzania',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FA 1F1E6',
+    char: '🇺🇦',
+    name: 'flag: Ukraine',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FA 1F1EC',
+    char: '🇺🇬',
+    name: 'flag: Uganda',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FA 1F1F2',
+    char: '🇺🇲',
+    name: 'flag: U.S. Outlying Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FA 1F1F3',
+    char: '🇺🇳',
+    name: 'flag: United Nations',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FA 1F1F8',
+    char: '🇺🇸',
+    name: 'flag: United States',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FA 1F1FE',
+    char: '🇺🇾',
+    name: 'flag: Uruguay',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FA 1F1FF',
+    char: '🇺🇿',
+    name: 'flag: Uzbekistan',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FB 1F1E6',
+    char: '🇻🇦',
+    name: 'flag: Vatican City',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FB 1F1E8',
+    char: '🇻🇨',
+    name: 'flag: St. Vincent & Grenadines',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FB 1F1EA',
+    char: '🇻🇪',
+    name: 'flag: Venezuela',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FB 1F1EC',
+    char: '🇻🇬',
+    name: 'flag: British Virgin Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FB 1F1EE',
+    char: '🇻🇮',
+    name: 'flag: U.S. Virgin Islands',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FB 1F1F3',
+    char: '🇻🇳',
+    name: 'flag: Vietnam',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FB 1F1FA',
+    char: '🇻🇺',
+    name: 'flag: Vanuatu',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FC 1F1EB',
+    char: '🇼🇫',
+    name: 'flag: Wallis & Futuna',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FC 1F1F8',
+    char: '🇼🇸',
+    name: 'flag: Samoa',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FD 1F1F0',
+    char: '🇽🇰',
+    name: 'flag: Kosovo',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FE 1F1EA',
+    char: '🇾🇪',
+    name: 'flag: Yemen',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FE 1F1F9',
+    char: '🇾🇹',
+    name: 'flag: Mayotte',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FF 1F1E6',
+    char: '🇿🇦',
+    name: 'flag: South Africa',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FF 1F1F2',
+    char: '🇿🇲',
+    name: 'flag: Zambia',
+    category: 'flags',
+  },
+  {
+    codes: '1F1FF 1F1FC',
+    char: '🇿🇼',
+    name: 'flag: Zimbabwe',
+    category: 'flags',
+  },
+];
