+++ emojis/foodAndDrink.js
@@ -1,0 +1,543 @@
+export default [
+  {
+    codes: '1F347',
+    char: '🍇',
+    name: 'grapes',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F348',
+    char: '🍈',
+    name: 'melon',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F349',
+    char: '🍉',
+    name: 'watermelon',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F34A',
+    char: '🍊',
+    name: 'tangerine',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F34B',
+    char: '🍋',
+    name: 'lemon',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F34C',
+    char: '🍌',
+    name: 'banana',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F34D',
+    char: '🍍',
+    name: 'pineapple',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F34E',
+    char: '🍎',
+    name: 'red apple',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F34F',
+    char: '🍏',
+    name: 'green apple',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F350',
+    char: '🍐',
+    name: 'pear',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F351',
+    char: '🍑',
+    name: 'peach',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F352',
+    char: '🍒',
+    name: 'cherries',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F353',
+    char: '🍓',
+    name: 'strawberry',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F345',
+    char: '🍅',
+    name: 'tomato',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F346',
+    char: '🍆',
+    name: 'eggplant',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F33D',
+    char: '🌽',
+    name: 'ear of corn',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F965',
+    char: '🥥',
+    name: 'coconut',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F951',
+    char: '🥑',
+    name: 'avocado',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F954',
+    char: '🥔',
+    name: 'potato',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F955',
+    char: '🥕',
+    name: 'carrot',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F336 FE0F',
+    char: '🌶️',
+    name: 'hot pepper',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F952',
+    char: '🥒',
+    name: 'cucumber',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F96C',
+    char: '🥬',
+    name: 'leafy green',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F966',
+    char: '🥦',
+    name: 'broccoli',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F9C4',
+    char: '🧄',
+    name: 'garlic',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F9C5',
+    char: '🧅',
+    name: 'onion',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F344',
+    char: '🍄',
+    name: 'mushroom',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F330',
+    char: '🌰',
+    name: 'chestnut',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F35E',
+    char: '🍞',
+    name: 'bread',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F356',
+    char: '🍖',
+    name: 'meat on bone',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F357',
+    char: '🍗',
+    name: 'poultry leg',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F354',
+    char: '🍔',
+    name: 'hamburger',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F35F',
+    char: '🍟',
+    name: 'french fries',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F355',
+    char: '🍕',
+    name: 'pizza',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F373',
+    char: '🍳',
+    name: 'cooking',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F958',
+    char: '🥘',
+    name: 'shallow pan of food',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F372',
+    char: '🍲',
+    name: 'pot of food',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F371',
+    char: '🍱',
+    name: 'bento box',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F358',
+    char: '🍘',
+    name: 'rice cracker',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F359',
+    char: '🍙',
+    name: 'rice ball',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F35A',
+    char: '🍚',
+    name: 'cooked rice',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F35B',
+    char: '🍛',
+    name: 'curry rice',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F35C',
+    char: '🍜',
+    name: 'steaming bowl',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F35D',
+    char: '🍝',
+    name: 'spaghetti',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F360',
+    char: '🍠',
+    name: 'roasted sweet potato',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F362',
+    char: '🍢',
+    name: 'oden',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F363',
+    char: '🍣',
+    name: 'sushi',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F364',
+    char: '🍤',
+    name: 'fried shrimp',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F365',
+    char: '🍥',
+    name: 'fish cake with swirl',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F980',
+    char: '🦀',
+    name: 'crab',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F99E',
+    char: '🦞',
+    name: 'lobster',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F990',
+    char: '🦐',
+    name: 'shrimp',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F991',
+    char: '🦑',
+    name: 'squid',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F9AA',
+    char: '🦪',
+    name: 'oyster',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F361',
+    char: '🍡',
+    name: 'dango',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F366',
+    char: '🍦',
+    name: 'soft ice cream',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F367',
+    char: '🍧',
+    name: 'shaved ice',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F368',
+    char: '🍨',
+    name: 'ice cream',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F9C1',
+    char: '🧁',
+    name: 'cupcake',
+    category: 'foodAndDrink',
+  },
+  { codes: '1F967', char: '🥧', name: 'pie', category: 'foodAndDrink' },
+  {
+    codes: '1F369',
+    char: '🍩',
+    name: 'doughnut',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F36A',
+    char: '🍪',
+    name: 'cookie',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F382',
+    char: '🎂',
+    name: 'birthday cake',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F370',
+    char: '🍰',
+    name: 'shortcake',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F36B',
+    char: '🍫',
+    name: 'chocolate bar',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F36C',
+    char: '🍬',
+    name: 'candy',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F36D',
+    char: '🍭',
+    name: 'lollipop',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F36E',
+    char: '🍮',
+    name: 'custard',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F36F',
+    char: '🍯',
+    name: 'honey pot',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F37C',
+    char: '🍼',
+    name: 'baby bottle',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '2615',
+    char: '☕',
+    name: 'hot beverage',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F375',
+    char: '🍵',
+    name: 'teacup without handle',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F376',
+    char: '🍶',
+    name: 'sake',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F377',
+    char: '🍷',
+    name: 'wine glass',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F378',
+    char: '🍸',
+    name: 'cocktail glass',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F379',
+    char: '🍹',
+    name: 'tropical drink',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F37A',
+    char: '🍺',
+    name: 'beer mug',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F37B',
+    char: '🍻',
+    name: 'clinking beer mugs',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F95B',
+    char: '🥛',
+    name: 'glass of milk',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F37E',
+    char: '🍾',
+    name: 'bottle with popping cork',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F942',
+    char: '🥂',
+    name: 'clinking glasses',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F943',
+    char: '🥃',
+    name: 'tumbler glass',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F964',
+    char: '🥤',
+    name: 'cup with straw',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F9C3',
+    char: '🧃',
+    name: 'beverage box',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F9C9',
+    char: '🧉',
+    name: 'mate',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F9CA',
+    char: '🧊',
+    name: 'ice cube',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F962',
+    char: '🥢',
+    name: 'chopsticks',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F37D FE0F',
+    char: '🍽️',
+    name: 'fork and knife with plate',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F944',
+    char: '🥄',
+    name: 'spoon',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F374',
+    char: '🍴',
+    name: 'fork and knife',
+    category: 'foodAndDrink',
+  },
+  {
+    codes: '1F52A',
+    char: '🔪',
+    name: 'kitchen knife',
+    category: 'foodAndDrink',
+  },
+];
