+++ emojis/index.js
@@ -1,0 +1,23 @@
+import skintone from './skintone.js';
+import people from './people.js';
+import nature from './nature.js';
+import foodAndDrink from './foodAndDrink.js';
+import activity from './activity.js';
+import places from './places.js';
+import objects from './objects.js';
+import symbols from './symbols.js';
+import flags from './flags.js';
+
+const emojis = [
+  skintone,
+  people,
+  nature,
+  foodAndDrink,
+  activity,
+  places,
+  objects,
+  symbols,
+  flags,
+];
+
+export default emojis;
