+++ emojis/nature.js
@@ -1,0 +1,341 @@
+export default [
+  {
+    codes: '1F435',
+    char: '🐵',
+    name: 'monkey face',
+    category: 'nature',
+  },
+  { codes: '1F412', char: '🐒', name: 'monkey', category: 'nature' },
+  { codes: '1F436', char: '🐶', name: 'dog face', category: 'nature' },
+  { codes: '1F415', char: '🐕', name: 'dog', category: 'nature' },
+  { codes: '1F429', char: '🐩', name: 'poodle', category: 'nature' },
+  { codes: '1F43A', char: '🐺', name: 'wolf', category: 'nature' },
+  { codes: '1F431', char: '🐱', name: 'cat face', category: 'nature' },
+  { codes: '1F408', char: '🐈', name: 'cat', category: 'nature' },
+  {
+    codes: '1F42F',
+    char: '🐯',
+    name: 'tiger face',
+    category: 'nature',
+  },
+  { codes: '1F405', char: '🐅', name: 'tiger', category: 'nature' },
+  { codes: '1F406', char: '🐆', name: 'leopard', category: 'nature' },
+  {
+    codes: '1F434',
+    char: '🐴',
+    name: 'horse face',
+    category: 'nature',
+  },
+  { codes: '1F40E', char: '🐎', name: 'horse', category: 'nature' },
+  { codes: '1F42E', char: '🐮', name: 'cow face', category: 'nature' },
+  { codes: '1F402', char: '🐂', name: 'ox', category: 'nature' },
+  {
+    codes: '1F403',
+    char: '🐃',
+    name: 'water buffalo',
+    category: 'nature',
+  },
+  { codes: '1F404', char: '🐄', name: 'cow', category: 'nature' },
+  { codes: '1F437', char: '🐷', name: 'pig face', category: 'nature' },
+  { codes: '1F416', char: '🐖', name: 'pig', category: 'nature' },
+  { codes: '1F417', char: '🐗', name: 'boar', category: 'nature' },
+  { codes: '1F43D', char: '🐽', name: 'pig nose', category: 'nature' },
+  { codes: '1F40F', char: '🐏', name: 'ram', category: 'nature' },
+  { codes: '1F411', char: '🐑', name: 'ewe', category: 'nature' },
+  { codes: '1F410', char: '🐐', name: 'goat', category: 'nature' },
+  { codes: '1F42A', char: '🐪', name: 'camel', category: 'nature' },
+  { codes: '1F418', char: '🐘', name: 'elephant', category: 'nature' },
+  {
+    codes: '1F42D',
+    char: '🐭',
+    name: 'mouse face',
+    category: 'nature',
+  },
+  { codes: '1F401', char: '🐁', name: 'mouse', category: 'nature' },
+  { codes: '1F439', char: '🐹', name: 'hamster', category: 'nature' },
+  {
+    codes: '1F430',
+    char: '🐰',
+    name: 'rabbit face',
+    category: 'nature',
+  },
+  { codes: '1F407', char: '🐇', name: 'rabbit', category: 'nature' },
+  { codes: '1F43B', char: '🐻', name: 'bear', category: 'nature' },
+  { codes: '1F428', char: '🐨', name: 'koala', category: 'nature' },
+  { codes: '1F43C', char: '🐼', name: 'panda', category: 'nature' },
+  {
+    codes: '1F43E',
+    char: '🐾',
+    name: 'paw prints',
+    category: 'nature',
+  },
+  { codes: '1F414', char: '🐔', name: 'chicken', category: 'nature' },
+  { codes: '1F413', char: '🐓', name: 'rooster', category: 'nature' },
+  {
+    codes: '1F423',
+    char: '🐣',
+    name: 'hatching chick',
+    category: 'nature',
+  },
+  {
+    codes: '1F424',
+    char: '🐤',
+    name: 'baby chick',
+    category: 'nature',
+  },
+  {
+    codes: '1F425',
+    char: '🐥',
+    name: 'front-facing baby chick',
+    category: 'nature',
+  },
+  { codes: '1F426', char: '🐦', name: 'bird', category: 'nature' },
+  { codes: '1F427', char: '🐧', name: 'penguin', category: 'nature' },
+  { codes: '1F438', char: '🐸', name: 'frog', category: 'nature' },
+  { codes: '1F40A', char: '🐊', name: 'crocodile', category: 'nature' },
+  { codes: '1F422', char: '🐢', name: 'turtle', category: 'nature' },
+  { codes: '1F40D', char: '🐍', name: 'snake', category: 'nature' },
+  {
+    codes: '1F432',
+    char: '🐲',
+    name: 'dragon face',
+    category: 'nature',
+  },
+  { codes: '1F409', char: '🐉', name: 'dragon', category: 'nature' },
+  {
+    codes: '1F433',
+    char: '🐳',
+    name: 'spouting whale',
+    category: 'nature',
+  },
+  { codes: '1F40B', char: '🐋', name: 'whale', category: 'nature' },
+  { codes: '1F42C', char: '🐬', name: 'dolphin', category: 'nature' },
+  { codes: '1F41F', char: '🐟', name: 'fish', category: 'nature' },
+  {
+    codes: '1F420',
+    char: '🐠',
+    name: 'tropical fish',
+    category: 'nature',
+  },
+  { codes: '1F421', char: '🐡', name: 'blowfish', category: 'nature' },
+  { codes: '1F988', char: '🦈', name: 'shark', category: 'nature' },
+  { codes: '1F419', char: '🐙', name: 'octopus', category: 'nature' },
+  { codes: '1F40C', char: '🐌', name: 'snail', category: 'nature' },
+  { codes: '1F98B', char: '🦋', name: 'butterfly', category: 'nature' },
+  { codes: '1F41B', char: '🐛', name: 'bug', category: 'nature' },
+  { codes: '1F41C', char: '🐜', name: 'ant', category: 'nature' },
+  { codes: '1F41D', char: '🐝', name: 'honeybee', category: 'nature' },
+  {
+    codes: '1F41E',
+    char: '🐞',
+    name: 'lady beetle',
+    category: 'nature',
+  },
+
+  { codes: '1F490', char: '💐', name: 'bouquet', category: 'nature' },
+  {
+    codes: '1F338',
+    char: '🌸',
+    name: 'cherry blossom',
+    category: 'nature',
+  },
+  {
+    codes: '1F4AE',
+    char: '💮',
+    name: 'white flower',
+    category: 'nature',
+  },
+  { codes: '1F339', char: '🌹', name: 'rose', category: 'nature' },
+  { codes: '1F33A', char: '🌺', name: 'hibiscus', category: 'nature' },
+  { codes: '1F33B', char: '🌻', name: 'sunflower', category: 'nature' },
+  { codes: '1F33C', char: '🌼', name: 'blossom', category: 'nature' },
+  { codes: '1F337', char: '🌷', name: 'tulip', category: 'nature' },
+  { codes: '1F331', char: '🌱', name: 'seedling', category: 'nature' },
+  {
+    codes: '1F332',
+    char: '🌲',
+    name: 'evergreen tree',
+    category: 'nature',
+  },
+  {
+    codes: '1F333',
+    char: '🌳',
+    name: 'deciduous tree',
+    category: 'nature',
+  },
+  { codes: '1F334', char: '🌴', name: 'palm tree', category: 'nature' },
+  { codes: '1F335', char: '🌵', name: 'cactus', category: 'nature' },
+  {
+    codes: '1F33E',
+    char: '🌾',
+    name: 'sheaf of rice',
+    category: 'nature',
+  },
+  { codes: '1F33F', char: '🌿', name: 'herb', category: 'nature' },
+  {
+    codes: '2618 FE0F',
+    char: '☘️',
+    name: 'shamrock',
+    category: 'nature',
+  },
+  {
+    codes: '1F340',
+    char: '🍀',
+    name: 'four leaf clover',
+    category: 'nature',
+  },
+  {
+    codes: '1F341',
+    char: '🍁',
+    name: 'maple leaf',
+    category: 'nature',
+  },
+  {
+    codes: '1F342',
+    char: '🍂',
+    name: 'fallen leaf',
+    category: 'nature',
+  },
+  {
+    codes: '1F343',
+    char: '🍃',
+    name: 'leaf fluttering in wind',
+    category: 'nature',
+  },
+  { codes: '1F311', char: '🌑', name: 'new moon', category: 'nature' },
+  {
+    codes: '1F312',
+    char: '🌒',
+    name: 'waxing crescent moon',
+    category: 'nature',
+  },
+  {
+    codes: '1F313',
+    char: '🌓',
+    name: 'first quarter moon',
+    category: 'nature',
+  },
+  {
+    codes: '1F314',
+    char: '🌔',
+    name: 'waxing gibbous moon',
+    category: 'nature',
+  },
+  { codes: '1F315', char: '🌕', name: 'full moon', category: 'nature' },
+  {
+    codes: '1F316',
+    char: '🌖',
+    name: 'waning gibbous moon',
+    category: 'nature',
+  },
+  {
+    codes: '1F317',
+    char: '🌗',
+    name: 'last quarter moon',
+    category: 'nature',
+  },
+  {
+    codes: '1F318',
+    char: '🌘',
+    name: 'waning crescent moon',
+    category: 'nature',
+  },
+  {
+    codes: '1F319',
+    char: '🌙',
+    name: 'crescent moon',
+    category: 'nature',
+  },
+  {
+    codes: '1F31A',
+    char: '🌚',
+    name: 'new moon face',
+    category: 'nature',
+  },
+  {
+    codes: '1F31B',
+    char: '🌛',
+    name: 'first quarter moon face',
+    category: 'nature',
+  },
+  {
+    codes: '1F31C',
+    char: '🌜',
+    name: 'last quarter moon face',
+    category: 'nature',
+  },
+  {
+    codes: '1F31D',
+    char: '🌝',
+    name: 'full moon face',
+    category: 'nature',
+  },
+  {
+    codes: '1F31E',
+    char: '🌞',
+    name: 'sun with face',
+    category: 'nature',
+  },
+  { codes: '2600 FE0F', char: '☀️', name: 'sun', category: 'nature' },
+  { codes: '2B50', char: '⭐', name: 'star', category: 'nature' },
+  {
+    codes: '1F31F',
+    char: '🌟',
+    name: 'glowing star',
+    category: 'nature',
+  },
+  {
+    codes: '1F320',
+    char: '🌠',
+    name: 'shooting star',
+    category: 'nature',
+  },
+  { codes: '1F30C', char: '🌌', name: 'milky way', category: 'nature' },
+  { codes: '2601 FE0F', char: '☁️', name: 'cloud', category: 'nature' },
+  {
+    codes: '26C5',
+    char: '⛅',
+    name: 'sun behind cloud',
+    category: 'nature',
+  },
+  { codes: '1F300', char: '🌀', name: 'cyclone', category: 'nature' },
+  { codes: '1F308', char: '🌈', name: 'rainbow', category: 'nature' },
+  {
+    codes: '1F302',
+    char: '🌂',
+    name: 'closed umbrella',
+    category: 'nature',
+  },
+  {
+    codes: '2614',
+    char: '☔',
+    name: 'umbrella with rain drops',
+    category: 'nature',
+  },
+  {
+    codes: '26A1',
+    char: '⚡',
+    name: 'high voltage',
+    category: 'nature',
+  },
+  {
+    codes: '2744 FE0F',
+    char: '❄️',
+    name: 'snowflake',
+    category: 'nature',
+  },
+  {
+    codes: '26C4',
+    char: '⛄',
+    name: 'snowman without snow',
+    category: 'nature',
+  },
+  { codes: '1F525', char: '🔥', name: 'fire', category: 'nature' },
+  { codes: '1F4A7', char: '💧', name: 'droplet', category: 'nature' },
+  {
+    codes: '1F30A',
+    char: '🌊',
+    name: 'water wave',
+    category: 'nature',
+  },
+];
