+++ emojis/objects.js
@@ -1,0 +1,631 @@
+export default [
+  { codes: '1F453', char: '👓', name: 'glasses', category: 'objects' },
+  { codes: '1F454', char: '👔', name: 'necktie', category: 'objects' },
+  { codes: '1F455', char: '👕', name: 't-shirt', category: 'objects' },
+  { codes: '1F456', char: '👖', name: 'jeans', category: 'objects' },
+  { codes: '1F457', char: '👗', name: 'dress', category: 'objects' },
+  { codes: '1F459', char: '👙', name: 'bikini', category: 'objects' },
+  {
+    codes: '1F45A',
+    char: '👚',
+    name: 'woman’s clothes',
+    category: 'objects',
+  },
+  { codes: '1F45B', char: '👛', name: 'purse', category: 'objects' },
+  { codes: '1F45C', char: '👜', name: 'handbag', category: 'objects' },
+  {
+    codes: '1F45D',
+    char: '👝',
+    name: 'clutch bag',
+    category: 'objects',
+  },
+  { codes: '1F392', char: '🎒', name: 'backpack', category: 'objects' },
+  {
+    codes: '1F45E',
+    char: '👞',
+    name: 'man’s shoe',
+    category: 'objects',
+  },
+  {
+    codes: '1F45F',
+    char: '👟',
+    name: 'running shoe',
+    category: 'objects',
+  },
+  {
+    codes: '1F460',
+    char: '👠',
+    name: 'high-heeled shoe',
+    category: 'objects',
+  },
+  {
+    codes: '1F461',
+    char: '👡',
+    name: 'woman’s sandal',
+    category: 'objects',
+  },
+  {
+    codes: '1F462',
+    char: '👢',
+    name: 'woman’s boot',
+    category: 'objects',
+  },
+  { codes: '1F451', char: '👑', name: 'crown', category: 'objects' },
+  {
+    codes: '1F452',
+    char: '👒',
+    name: 'woman’s hat',
+    category: 'objects',
+  },
+  { codes: '1F3A9', char: '🎩', name: 'top hat', category: 'objects' },
+  {
+    codes: '1F393',
+    char: '🎓',
+    name: 'graduation cap',
+    category: 'objects',
+  },
+  { codes: '1F484', char: '💄', name: 'lipstick', category: 'objects' },
+  { codes: '1F48D', char: '💍', name: 'ring', category: 'objects' },
+  {
+    codes: '1F48E',
+    char: '💎',
+    name: 'gem stone',
+    category: 'objects',
+  },
+  {
+    codes: '1F507',
+    char: '🔇',
+    name: 'muted speaker',
+    category: 'objects',
+  },
+  {
+    codes: '1F508',
+    char: '🔈',
+    name: 'speaker low volume',
+    category: 'objects',
+  },
+  {
+    codes: '1F509',
+    char: '🔉',
+    name: 'speaker medium volume',
+    category: 'objects',
+  },
+  {
+    codes: '1F50A',
+    char: '🔊',
+    name: 'speaker high volume',
+    category: 'objects',
+  },
+  {
+    codes: '1F4E2',
+    char: '📢',
+    name: 'loudspeaker',
+    category: 'objects',
+  },
+  {
+    codes: '1F4E3',
+    char: '📣',
+    name: 'megaphone',
+    category: 'objects',
+  },
+  {
+    codes: '1F4EF',
+    char: '📯',
+    name: 'postal horn',
+    category: 'objects',
+  },
+  { codes: '1F514', char: '🔔', name: 'bell', category: 'objects' },
+  {
+    codes: '1F515',
+    char: '🔕',
+    name: 'bell with slash',
+    category: 'objects',
+  },
+  {
+    codes: '1F3BC',
+    char: '🎼',
+    name: 'musical score',
+    category: 'objects',
+  },
+  {
+    codes: '1F3B5',
+    char: '🎵',
+    name: 'musical note',
+    category: 'objects',
+  },
+  {
+    codes: '1F3B6',
+    char: '🎶',
+    name: 'musical notes',
+    category: 'objects',
+  },
+  {
+    codes: '1F3A4',
+    char: '🎤',
+    name: 'microphone',
+    category: 'objects',
+  },
+  {
+    codes: '1F3A7',
+    char: '🎧',
+    name: 'headphone',
+    category: 'objects',
+  },
+  { codes: '1F4FB', char: '📻', name: 'radio', category: 'objects' },
+  {
+    codes: '1F3B7',
+    char: '🎷',
+    name: 'saxophone',
+    category: 'objects',
+  },
+  { codes: '1F3B8', char: '🎸', name: 'guitar', category: 'objects' },
+  {
+    codes: '1F3B9',
+    char: '🎹',
+    name: 'musical keyboard',
+    category: 'objects',
+  },
+  { codes: '1F3BA', char: '🎺', name: 'trumpet', category: 'objects' },
+  { codes: '1F3BB', char: '🎻', name: 'violin', category: 'objects' },
+  {
+    codes: '1F4F1',
+    char: '📱',
+    name: 'mobile phone',
+    category: 'objects',
+  },
+  {
+    codes: '1F4F2',
+    char: '📲',
+    name: 'mobile phone with arrow',
+    category: 'objects',
+  },
+  {
+    codes: '260E FE0F',
+    char: '☎️',
+    name: 'telephone',
+    category: 'objects',
+  },
+  {
+    codes: '1F4DE',
+    char: '📞',
+    name: 'telephone receiver',
+    category: 'objects',
+  },
+  { codes: '1F4DF', char: '📟', name: 'pager', category: 'objects' },
+  {
+    codes: '1F4E0',
+    char: '📠',
+    name: 'fax machine',
+    category: 'objects',
+  },
+  { codes: '1F50B', char: '🔋', name: 'battery', category: 'objects' },
+  {
+    codes: '1F50C',
+    char: '🔌',
+    name: 'electric plug',
+    category: 'objects',
+  },
+  {
+    codes: '1F4BB',
+    char: '💻',
+    name: 'laptop computer',
+    category: 'objects',
+  },
+  {
+    codes: '1F4BD',
+    char: '💽',
+    name: 'computer disk',
+    category: 'objects',
+  },
+  {
+    codes: '1F4BE',
+    char: '💾',
+    name: 'floppy disk',
+    category: 'objects',
+  },
+  {
+    codes: '1F4BF',
+    char: '💿',
+    name: 'optical disk',
+    category: 'objects',
+  },
+  { codes: '1F4C0', char: '📀', name: 'dvd', category: 'objects' },
+  { codes: '1F9EE', char: '🧮', name: 'abacus', category: 'objects' },
+  {
+    codes: '1F3A5',
+    char: '🎥',
+    name: 'movie camera',
+    category: 'objects',
+  },
+  {
+    codes: '1F39E FE0F',
+    char: '🎞️',
+    name: 'film frames',
+    category: 'objects',
+  },
+  {
+    codes: '1F4FD FE0F',
+    char: '📽️',
+    name: 'film projector',
+    category: 'objects',
+  },
+  {
+    codes: '1F3AC',
+    char: '🎬',
+    name: 'clapper board',
+    category: 'objects',
+  },
+  {
+    codes: '1F4FA',
+    char: '📺',
+    name: 'television',
+    category: 'objects',
+  },
+  { codes: '1F4F7', char: '📷', name: 'camera', category: 'objects' },
+  {
+    codes: '1F4F9',
+    char: '📹',
+    name: 'video camera',
+    category: 'objects',
+  },
+  {
+    codes: '1F4FC',
+    char: '📼',
+    name: 'videocassette',
+    category: 'objects',
+  },
+  {
+    codes: '1F50D',
+    char: '🔍',
+    name: 'magnifying glass tilted left',
+    category: 'objects',
+  },
+  {
+    codes: '1F50E',
+    char: '🔎',
+    name: 'magnifying glass tilted right',
+    category: 'objects',
+  },
+  {
+    codes: '1F4A1',
+    char: '💡',
+    name: 'light bulb',
+    category: 'objects',
+  },
+  {
+    codes: '1F526',
+    char: '🔦',
+    name: 'flashlight',
+    category: 'objects',
+  },
+  {
+    codes: '1F3EE',
+    char: '🏮',
+    name: 'red paper lantern',
+    category: 'objects',
+  },
+  {
+    codes: '1F4D4',
+    char: '📔',
+    name: 'notebook with decorative cover',
+    category: 'objects',
+  },
+  {
+    codes: '1F4D5',
+    char: '📕',
+    name: 'closed book',
+    category: 'objects',
+  },
+  {
+    codes: '1F4D6',
+    char: '📖',
+    name: 'open book',
+    category: 'objects',
+  },
+  {
+    codes: '1F4D7',
+    char: '📗',
+    name: 'green book',
+    category: 'objects',
+  },
+  {
+    codes: '1F4D8',
+    char: '📘',
+    name: 'blue book',
+    category: 'objects',
+  },
+  {
+    codes: '1F4D9',
+    char: '📙',
+    name: 'orange book',
+    category: 'objects',
+  },
+  { codes: '1F4DA', char: '📚', name: 'books', category: 'objects' },
+  { codes: '1F4D3', char: '📓', name: 'notebook', category: 'objects' },
+  { codes: '1F4D2', char: '📒', name: 'ledger', category: 'objects' },
+  {
+    codes: '1F4C3',
+    char: '📃',
+    name: 'page with curl',
+    category: 'objects',
+  },
+  { codes: '1F4DC', char: '📜', name: 'scroll', category: 'objects' },
+  {
+    codes: '1F4C4',
+    char: '📄',
+    name: 'page facing up',
+    category: 'objects',
+  },
+  {
+    codes: '1F4F0',
+    char: '📰',
+    name: 'newspaper',
+    category: 'objects',
+  },
+  {
+    codes: '1F4D1',
+    char: '📑',
+    name: 'bookmark tabs',
+    category: 'objects',
+  },
+  { codes: '1F516', char: '🔖', name: 'bookmark', category: 'objects' },
+  {
+    codes: '1F4B0',
+    char: '💰',
+    name: 'money bag',
+    category: 'objects',
+  },
+  {
+    codes: '1F4B4',
+    char: '💴',
+    name: 'yen banknote',
+    category: 'objects',
+  },
+  {
+    codes: '1F4B5',
+    char: '💵',
+    name: 'dollar banknote',
+    category: 'objects',
+  },
+  {
+    codes: '1F4B6',
+    char: '💶',
+    name: 'euro banknote',
+    category: 'objects',
+  },
+  {
+    codes: '1F4B7',
+    char: '💷',
+    name: 'pound banknote',
+    category: 'objects',
+  },
+  {
+    codes: '1F4B8',
+    char: '💸',
+    name: 'money with wings',
+    category: 'objects',
+  },
+  {
+    codes: '1F4B3',
+    char: '💳',
+    name: 'credit card',
+    category: 'objects',
+  },
+  {
+    codes: '1F4B9',
+    char: '💹',
+    name: 'chart increasing with yen',
+    category: 'objects',
+  },
+  {
+    codes: '1F4B1',
+    char: '💱',
+    name: 'currency exchange',
+    category: 'objects',
+  },
+  {
+    codes: '1F4B2',
+    char: '💲',
+    name: 'heavy dollar sign',
+    category: 'objects',
+  },
+  {
+    codes: '2709 FE0F',
+    char: '✉️',
+    name: 'envelope',
+    category: 'objects',
+  },
+  { codes: '1F4E7', char: '📧', name: 'e-mail', category: 'objects' },
+  {
+    codes: '1F4E8',
+    char: '📨',
+    name: 'incoming envelope',
+    category: 'objects',
+  },
+  {
+    codes: '1F4E9',
+    char: '📩',
+    name: 'envelope with arrow',
+    category: 'objects',
+  },
+  {
+    codes: '1F4E4',
+    char: '📤',
+    name: 'outbox tray',
+    category: 'objects',
+  },
+  {
+    codes: '1F4E5',
+    char: '📥',
+    name: 'inbox tray',
+    category: 'objects',
+  },
+  { codes: '1F4E6', char: '📦', name: 'package', category: 'objects' },
+  {
+    codes: '1F4EB',
+    char: '📫',
+    name: 'closed mailbox with raised flag',
+    category: 'objects',
+  },
+  {
+    codes: '1F4EA',
+    char: '📪',
+    name: 'closed mailbox with lowered flag',
+    category: 'objects',
+  },
+  {
+    codes: '1F4EC',
+    char: '📬',
+    name: 'open mailbox with raised flag',
+    category: 'objects',
+  },
+  {
+    codes: '1F4ED',
+    char: '📭',
+    name: 'open mailbox with lowered flag',
+    category: 'objects',
+  },
+  { codes: '1F4EE', char: '📮', name: 'postbox', category: 'objects' },
+  {
+    codes: '270F FE0F',
+    char: '✏️',
+    name: 'pencil',
+    category: 'objects',
+  },
+  {
+    codes: '2712 FE0F',
+    char: '✒️',
+    name: 'black nib',
+    category: 'objects',
+  },
+  { codes: '1F4DD', char: '📝', name: 'memo', category: 'objects' },
+  {
+    codes: '1F4BC',
+    char: '💼',
+    name: 'briefcase',
+    category: 'objects',
+  },
+  {
+    codes: '1F4C1',
+    char: '📁',
+    name: 'file folder',
+    category: 'objects',
+  },
+  {
+    codes: '1F4C2',
+    char: '📂',
+    name: 'open file folder',
+    category: 'objects',
+  },
+  { codes: '1F4C5', char: '📅', name: 'calendar', category: 'objects' },
+  {
+    codes: '1F4C7',
+    char: '📇',
+    name: 'card index',
+    category: 'objects',
+  },
+  {
+    codes: '1F4C8',
+    char: '📈',
+    name: 'chart increasing',
+    category: 'objects',
+  },
+  {
+    codes: '1F4C9',
+    char: '📉',
+    name: 'chart decreasing',
+    category: 'objects',
+  },
+  {
+    codes: '1F4CA',
+    char: '📊',
+    name: 'bar chart',
+    category: 'objects',
+  },
+  {
+    codes: '1F4CB',
+    char: '📋',
+    name: 'clipboard',
+    category: 'objects',
+  },
+  { codes: '1F4CC', char: '📌', name: 'pushpin', category: 'objects' },
+  {
+    codes: '1F4CD',
+    char: '📍',
+    name: 'round pushpin',
+    category: 'objects',
+  },
+  {
+    codes: '1F4CE',
+    char: '📎',
+    name: 'paperclip',
+    category: 'objects',
+  },
+  {
+    codes: '1F4CF',
+    char: '📏',
+    name: 'straight ruler',
+    category: 'objects',
+  },
+  {
+    codes: '1F4D0',
+    char: '📐',
+    name: 'triangular ruler',
+    category: 'objects',
+  },
+  {
+    codes: '2702 FE0F',
+    char: '✂️',
+    name: 'scissors',
+    category: 'objects',
+  },
+  { codes: '1F512', char: '🔒', name: 'locked', category: 'objects' },
+  { codes: '1F513', char: '🔓', name: 'unlocked', category: 'objects' },
+  {
+    codes: '1F50F',
+    char: '🔏',
+    name: 'locked with pen',
+    category: 'objects',
+  },
+  {
+    codes: '1F510',
+    char: '🔐',
+    name: 'locked with key',
+    category: 'objects',
+  },
+  { codes: '1F511', char: '🔑', name: 'key', category: 'objects' },
+  { codes: '1F528', char: '🔨', name: 'hammer', category: 'objects' },
+  { codes: '1F52B', char: '🔫', name: 'pistol', category: 'objects' },
+  { codes: '1F517', char: '🔗', name: 'link', category: 'objects' },
+  {
+    codes: '1F52C',
+    char: '🔬',
+    name: 'microscope',
+    category: 'objects',
+  },
+  {
+    codes: '1F52D',
+    char: '🔭',
+    name: 'telescope',
+    category: 'objects',
+  },
+  {
+    codes: '1F4E1',
+    char: '📡',
+    name: 'satellite antenna',
+    category: 'objects',
+  },
+  { codes: '1F489', char: '💉', name: 'syringe', category: 'objects' },
+  { codes: '1F48A', char: '💊', name: 'pill', category: 'objects' },
+  { codes: '1F6AA', char: '🚪', name: 'door', category: 'objects' },
+  { codes: '1F6BD', char: '🚽', name: 'toilet', category: 'objects' },
+  { codes: '1F6BF', char: '🚿', name: 'shower', category: 'objects' },
+  { codes: '1F6C1', char: '🛁', name: 'bathtub', category: 'objects' },
+  {
+    codes: '1F6AC',
+    char: '🚬',
+    name: 'cigarette',
+    category: 'objects',
+  },
+  { codes: '1F5FF', char: '🗿', name: 'moai', category: 'objects' },
+];
