+++ emojis/people.js
@@ -1,0 +1,2763 @@
+export default [
+  {
+    codes: '1F600',
+    char: '😀',
+    name: 'grinning face',
+    category: 'people',
+  },
+  {
+    codes: '1F604',
+    char: '😄',
+    name: 'grinning face with smiling eyes',
+    category: 'people',
+  },
+  {
+    codes: '1F60A',
+    char: '😊',
+    name: 'smiling face with smiling eyes',
+    category: 'people',
+  },
+  {
+    codes: '1F601',
+    char: '😁',
+    name: 'beaming face with smiling eyes',
+    category: 'people',
+  },
+  {
+    codes: '1F606',
+    char: '😆',
+    name: 'grinning squinting face',
+    category: 'people',
+  },
+  {
+    codes: '1F605',
+    char: '😅',
+    name: 'grinning face with sweat',
+    category: 'people',
+  },
+  {
+    codes: '1F602',
+    char: '😂',
+    name: 'face with tears of joy',
+    category: 'people',
+  },
+  {
+    codes: '1F642',
+    char: '🙂',
+    name: 'slightly smiling face',
+    category: 'people',
+  },
+  {
+    codes: '1F609',
+    char: '😉',
+    name: 'winking face',
+    category: 'people',
+  },
+  {
+    codes: '1F607',
+    char: '😇',
+    name: 'smiling face with halo',
+    category: 'people',
+  },
+  {
+    codes: '1F60D',
+    char: '😍',
+    name: 'smiling face with heart-eyes',
+    category: 'people',
+  },
+  {
+    codes: '1F618',
+    char: '😘',
+    name: 'face blowing a kiss',
+    category: 'people',
+  },
+  {
+    codes: '1F617',
+    char: '😗',
+    name: 'kissing face',
+    category: 'people',
+  },
+  {
+    codes: '1F61A',
+    char: '😚',
+    name: 'kissing face with closed eyes',
+    category: 'people',
+  },
+  {
+    codes: '1F60B',
+    char: '😋',
+    name: 'face savoring food',
+    category: 'people',
+  },
+  {
+    codes: '1F61B',
+    char: '😛',
+    name: 'face with tongue',
+    category: 'people',
+  },
+  {
+    codes: '1F61C',
+    char: '😜',
+    name: 'winking face with tongue',
+    category: 'people',
+  },
+  {
+    codes: '1F61D',
+    char: '😝',
+    name: 'squinting face with tongue',
+    category: 'people',
+  },
+  {
+    codes: '1F911',
+    char: '🤑',
+    name: 'money-mouth face',
+    category: 'people',
+  },
+  {
+    codes: '1F917',
+    char: '🤗',
+    name: 'hugging face',
+    category: 'people',
+  },
+  {
+    codes: '1F92D',
+    char: '🤭',
+    name: 'face with hand over mouth',
+    category: 'people',
+  },
+  {
+    codes: '1F92B',
+    char: '🤫',
+    name: 'shushing face',
+    category: 'people',
+  },
+  {
+    codes: '1F914',
+    char: '🤔',
+    name: 'thinking face',
+    category: 'people',
+  },
+  {
+    codes: '1F910',
+    char: '🤐',
+    name: 'zipper-mouth face',
+    category: 'people',
+  },
+  {
+    codes: '1F928',
+    char: '🤨',
+    name: 'face with raised eyebrow',
+    category: 'people',
+  },
+  {
+    codes: '1F610',
+    char: '😐',
+    name: 'neutral face',
+    category: 'people',
+  },
+  {
+    codes: '1F611',
+    char: '😑',
+    name: 'expressionless face',
+    category: 'people',
+  },
+  {
+    codes: '1F60F',
+    char: '😏',
+    name: 'smirking face',
+    category: 'people',
+  },
+  {
+    codes: '1F612',
+    char: '😒',
+    name: 'unamused face',
+    category: 'people',
+  },
+  {
+    codes: '1F62C',
+    char: '😬',
+    name: 'grimacing face',
+    category: 'people',
+  },
+  {
+    codes: '1F60C',
+    char: '😌',
+    name: 'relieved face',
+    category: 'people',
+  },
+  {
+    codes: '1F614',
+    char: '😔',
+    name: 'pensive face',
+    category: 'people',
+  },
+  {
+    codes: '1F62A',
+    char: '😪',
+    name: 'sleepy face',
+    category: 'people',
+  },
+  {
+    codes: '1F634',
+    char: '😴',
+    name: 'sleeping face',
+    category: 'people',
+  },
+  {
+    codes: '1F637',
+    char: '😷',
+    name: 'face with medical mask',
+    category: 'people',
+  },
+  {
+    codes: '1F635',
+    char: '😵',
+    name: 'dizzy face',
+    category: 'people',
+  },
+  {
+    codes: '1F920',
+    char: '🤠',
+    name: 'cowboy hat face',
+    category: 'people)',
+  },
+  {
+    codes: '1F973',
+    char: '🥳',
+    name: 'partying face',
+    category: 'people)',
+  },
+  {
+    codes: '1F60E',
+    char: '😎',
+    name: 'smiling face with sunglasses',
+    category: 'people',
+  },
+  {
+    codes: '1F615',
+    char: '😕',
+    name: 'confused face',
+    category: 'people',
+  },
+  {
+    codes: '1F61F',
+    char: '😟',
+    name: 'worried face',
+    category: 'people',
+  },
+  {
+    codes: '1F641',
+    char: '🙁',
+    name: 'slightly frowning face',
+    category: 'people',
+  },
+  {
+    codes: '1F62E',
+    char: '😮',
+    name: 'face with open mouth',
+    category: 'people',
+  },
+  {
+    codes: '1F632',
+    char: '😲',
+    name: 'astonished face',
+    category: 'people',
+  },
+  {
+    codes: '1F633',
+    char: '😳',
+    name: 'flushed face',
+    category: 'people',
+  },
+  {
+    codes: '1F626',
+    char: '😦',
+    name: 'frowning face with open mouth',
+    category: 'people',
+  },
+  {
+    codes: '1F628',
+    char: '😨',
+    name: 'fearful face',
+    category: 'people',
+  },
+  {
+    codes: '1F630',
+    char: '😰',
+    name: 'anxious face with sweat',
+    category: 'people',
+  },
+  {
+    codes: '1F625',
+    char: '😥',
+    name: 'sad but relieved face',
+    category: 'people',
+  },
+  {
+    codes: '1F62D',
+    char: '😭',
+    name: 'loudly crying face',
+    category: 'people',
+  },
+  {
+    codes: '1F631',
+    char: '😱',
+    name: 'face screaming in fear',
+    category: 'people',
+  },
+  {
+    codes: '1F616',
+    char: '😖',
+    name: 'confounded face',
+    category: 'people',
+  },
+  {
+    codes: '1F61E',
+    char: '😞',
+    name: 'disappointed face',
+    category: 'people',
+  },
+  {
+    codes: '1F613',
+    char: '😓',
+    name: 'downcast face with sweat',
+    category: 'people',
+  },
+  {
+    codes: '1F629',
+    char: '😩',
+    name: 'weary face',
+    category: 'people',
+  },
+  {
+    codes: '1F62B',
+    char: '😫',
+    name: 'tired face',
+    category: 'people',
+  },
+  {
+    codes: '1F624',
+    char: '😤',
+    name: 'face with steam from nose',
+    category: 'people',
+  },
+  {
+    codes: '1F621',
+    char: '😡',
+    name: 'pouting face',
+    category: 'people',
+  },
+  {
+    codes: '1F620',
+    char: '😠',
+    name: 'angry face',
+    category: 'people',
+  },
+  {
+    codes: '1F608',
+    char: '😈',
+    name: 'smiling face with horns',
+    category: 'people',
+  },
+  { codes: '1F480', char: '💀', name: 'skull', category: 'people' },
+  {
+    codes: '1F4A9',
+    char: '💩',
+    name: 'pile of poo',
+    category: 'people',
+  },
+  { codes: '1F479', char: '👹', name: 'ogre', category: 'people' },
+  { codes: '1F47A', char: '👺', name: 'goblin', category: 'people' },
+  { codes: '1F47B', char: '👻', name: 'ghost', category: 'people' },
+  { codes: '1F47D', char: '👽', name: 'alien', category: 'people' },
+  {
+    codes: '1F47E',
+    char: '👾',
+    name: 'alien monster',
+    category: 'people',
+  },
+  {
+    codes: '1F648',
+    char: '🙈',
+    name: 'see-no-evil monkey',
+    category: 'people',
+  },
+  {
+    codes: '1F649',
+    char: '🙉',
+    name: 'hear-no-evil monkey',
+    category: 'people',
+  },
+  {
+    codes: '1F64A',
+    char: '🙊',
+    name: 'speak-no-evil monkey',
+    category: 'people',
+  },
+  {
+    codes: '1F44B',
+    char: '👋',
+    name: 'waving hand',
+    category: 'people',
+  },
+  {
+    codes: '1F44B 1F3FB',
+    char: '👋🏻',
+    name: 'waving hand: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44B 1F3FC',
+    char: '👋🏼',
+    name: 'waving hand: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44B 1F3FD',
+    char: '👋🏽',
+    name: 'waving hand: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44B 1F3FE',
+    char: '👋🏾',
+    name: 'waving hand: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44B 1F3FF',
+    char: '👋🏿',
+    name: 'waving hand: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270B',
+    char: '✋',
+    name: 'raised hand',
+    category: 'people',
+  },
+  {
+    codes: '270B 1F3FB',
+    char: '✋🏻',
+    name: 'raised hand: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270B 1F3FC',
+    char: '✋🏼',
+    name: 'raised hand: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270B 1F3FD',
+    char: '✋🏽',
+    name: 'raised hand: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270B 1F3FE',
+    char: '✋🏾',
+    name: 'raised hand: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270B 1F3FF',
+    char: '✋🏿',
+    name: 'raised hand: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F44C', char: '👌', name: 'OK hand', category: 'people' },
+  {
+    codes: '1F44C 1F3FB',
+    char: '👌🏻',
+    name: 'OK hand: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44C 1F3FC',
+    char: '👌🏼',
+    name: 'OK hand: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44C 1F3FD',
+    char: '👌🏽',
+    name: 'OK hand: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44C 1F3FE',
+    char: '👌🏾',
+    name: 'OK hand: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44C 1F3FF',
+    char: '👌🏿',
+    name: 'OK hand: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270C FE0F',
+    char: '✌️',
+    name: 'victory hand',
+    category: 'people',
+  },
+  {
+    codes: '270C 1F3FB',
+    char: '✌🏻',
+    name: 'victory hand: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270C 1F3FC',
+    char: '✌🏼',
+    name: 'victory hand: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270C 1F3FD',
+    char: '✌🏽',
+    name: 'victory hand: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270C 1F3FE',
+    char: '✌🏾',
+    name: 'victory hand: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270C 1F3FF',
+    char: '✌🏿',
+    name: 'victory hand: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F448',
+    char: '👈',
+    name: 'backhand index pointing left',
+    category: 'people',
+  },
+  {
+    codes: '1F448 1F3FB',
+    char: '👈🏻',
+    name: 'backhand index pointing left: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F448 1F3FC',
+    char: '👈🏼',
+    name: 'backhand index pointing left: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F448 1F3FD',
+    char: '👈🏽',
+    name: 'backhand index pointing left: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F448 1F3FE',
+    char: '👈🏾',
+    name: 'backhand index pointing left: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F448 1F3FF',
+    char: '👈🏿',
+    name: 'backhand index pointing left: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F449',
+    char: '👉',
+    name: 'backhand index pointing right',
+    category: 'people',
+  },
+  {
+    codes: '1F449 1F3FB',
+    char: '👉🏻',
+    name: 'backhand index pointing right: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F449 1F3FC',
+    char: '👉🏼',
+    name: 'backhand index pointing right: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F449 1F3FD',
+    char: '👉🏽',
+    name: 'backhand index pointing right: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F449 1F3FE',
+    char: '👉🏾',
+    name: 'backhand index pointing right: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F449 1F3FF',
+    char: '👉🏿',
+    name: 'backhand index pointing right: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F446',
+    char: '👆',
+    name: 'backhand index pointing up',
+    category: 'people',
+  },
+  {
+    codes: '1F446 1F3FB',
+    char: '👆🏻',
+    name: 'backhand index pointing up: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F446 1F3FC',
+    char: '👆🏼',
+    name: 'backhand index pointing up: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F446 1F3FD',
+    char: '👆🏽',
+    name: 'backhand index pointing up: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F446 1F3FE',
+    char: '👆🏾',
+    name: 'backhand index pointing up: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F446 1F3FF',
+    char: '👆🏿',
+    name: 'backhand index pointing up: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F447',
+    char: '👇',
+    name: 'backhand index pointing down',
+    category: 'people',
+  },
+  {
+    codes: '1F447 1F3FB',
+    char: '👇🏻',
+    name: 'backhand index pointing down: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F447 1F3FC',
+    char: '👇🏼',
+    name: 'backhand index pointing down: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F447 1F3FD',
+    char: '👇🏽',
+    name: 'backhand index pointing down: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F447 1F3FE',
+    char: '👇🏾',
+    name: 'backhand index pointing down: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F447 1F3FF',
+    char: '👇🏿',
+    name: 'backhand index pointing down: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '261D FE0F',
+    char: '☝️',
+    name: 'index pointing up',
+    category: 'people',
+  },
+  {
+    codes: '261D',
+    char: '☝',
+    name: 'index pointing up',
+    category: 'people',
+  },
+  {
+    codes: '261D 1F3FB',
+    char: '☝🏻',
+    name: 'index pointing up: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '261D 1F3FC',
+    char: '☝🏼',
+    name: 'index pointing up: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '261D 1F3FD',
+    char: '☝🏽',
+    name: 'index pointing up: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '261D 1F3FE',
+    char: '☝🏾',
+    name: 'index pointing up: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '261D 1F3FF',
+    char: '☝🏿',
+    name: 'index pointing up: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F44D', char: '👍', name: 'thumbs up', category: 'people' },
+  {
+    codes: '1F44D 1F3FB',
+    char: '👍🏻',
+    name: 'thumbs up: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44D 1F3FC',
+    char: '👍🏼',
+    name: 'thumbs up: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44D 1F3FD',
+    char: '👍🏽',
+    name: 'thumbs up: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44D 1F3FE',
+    char: '👍🏾',
+    name: 'thumbs up: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44D 1F3FF',
+    char: '👍🏿',
+    name: 'thumbs up: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44E',
+    char: '👎',
+    name: 'thumbs down',
+    category: 'people',
+  },
+  {
+    codes: '1F44E 1F3FB',
+    char: '👎🏻',
+    name: 'thumbs down: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44E 1F3FC',
+    char: '👎🏼',
+    name: 'thumbs down: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44E 1F3FD',
+    char: '👎🏽',
+    name: 'thumbs down: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44E 1F3FE',
+    char: '👎🏾',
+    name: 'thumbs down: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44E 1F3FF',
+    char: '👎🏿',
+    name: 'thumbs down: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270A',
+    char: '✊',
+    name: 'raised fist',
+    category: 'people',
+  },
+  {
+    codes: '270A 1F3FB',
+    char: '✊🏻',
+    name: 'raised fist: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270A 1F3FC',
+    char: '✊🏼',
+    name: 'raised fist: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270A 1F3FD',
+    char: '✊🏽',
+    name: 'raised fist: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270A 1F3FE',
+    char: '✊🏾',
+    name: 'raised fist: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '270A 1F3FF',
+    char: '✊🏿',
+    name: 'raised fist: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44A',
+    char: '👊',
+    name: 'oncoming fist',
+    category: 'people',
+  },
+  {
+    codes: '1F44A 1F3FB',
+    char: '👊🏻',
+    name: 'oncoming fist: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44A 1F3FC',
+    char: '👊🏼',
+    name: 'oncoming fist: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44A 1F3FD',
+    char: '👊🏽',
+    name: 'oncoming fist: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44A 1F3FE',
+    char: '👊🏾',
+    name: 'oncoming fist: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44A 1F3FF',
+    char: '👊🏿',
+    name: 'oncoming fist: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44F',
+    char: '👏',
+    name: 'clapping hands',
+    category: 'people',
+  },
+  {
+    codes: '1F44F 1F3FB',
+    char: '👏🏻',
+    name: 'clapping hands: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44F 1F3FC',
+    char: '👏🏼',
+    name: 'clapping hands: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44F 1F3FD',
+    char: '👏🏽',
+    name: 'clapping hands: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44F 1F3FE',
+    char: '👏🏾',
+    name: 'clapping hands: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F44F 1F3FF',
+    char: '👏🏿',
+    name: 'clapping hands: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64C',
+    char: '🙌',
+    name: 'raising hands',
+    category: 'people',
+  },
+  {
+    codes: '1F64C 1F3FB',
+    char: '🙌🏻',
+    name: 'raising hands: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64C 1F3FC',
+    char: '🙌🏼',
+    name: 'raising hands: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64C 1F3FD',
+    char: '🙌🏽',
+    name: 'raising hands: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64C 1F3FE',
+    char: '🙌🏾',
+    name: 'raising hands: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64C 1F3FF',
+    char: '🙌🏿',
+    name: 'raising hands: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F450',
+    char: '👐',
+    name: 'open hands',
+    category: 'people',
+  },
+  {
+    codes: '1F450 1F3FB',
+    char: '👐🏻',
+    name: 'open hands: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F450 1F3FC',
+    char: '👐🏼',
+    name: 'open hands: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F450 1F3FD',
+    char: '👐🏽',
+    name: 'open hands: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F450 1F3FE',
+    char: '👐🏾',
+    name: 'open hands: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F450 1F3FF',
+    char: '👐🏿',
+    name: 'open hands: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64F',
+    char: '🙏',
+    name: 'folded hands',
+    category: 'people',
+  },
+  {
+    codes: '1F64F 1F3FB',
+    char: '🙏🏻',
+    name: 'folded hands: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64F 1F3FC',
+    char: '🙏🏼',
+    name: 'folded hands: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64F 1F3FD',
+    char: '🙏🏽',
+    name: 'folded hands: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64F 1F3FE',
+    char: '🙏🏾',
+    name: 'folded hands: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64F 1F3FF',
+    char: '🙏🏿',
+    name: 'folded hands: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F485',
+    char: '💅',
+    name: 'nail polish',
+    category: 'people',
+  },
+  {
+    codes: '1F485 1F3FB',
+    char: '💅🏻',
+    name: 'nail polish: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F485 1F3FC',
+    char: '💅🏼',
+    name: 'nail polish: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F485 1F3FD',
+    char: '💅🏽',
+    name: 'nail polish: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F485 1F3FE',
+    char: '💅🏾',
+    name: 'nail polish: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F485 1F3FF',
+    char: '💅🏿',
+    name: 'nail polish: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F4AA',
+    char: '💪',
+    name: 'flexed biceps',
+    category: 'people',
+  },
+  {
+    codes: '1F4AA 1F3FB',
+    char: '💪🏻',
+    name: 'flexed biceps: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F4AA 1F3FC',
+    char: '💪🏼',
+    name: 'flexed biceps: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F4AA 1F3FD',
+    char: '💪🏽',
+    name: 'flexed biceps: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F4AA 1F3FE',
+    char: '💪🏾',
+    name: 'flexed biceps: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F4AA 1F3FF',
+    char: '💪🏿',
+    name: 'flexed biceps: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F442', char: '👂', name: 'ear', category: 'people' },
+  {
+    codes: '1F442 1F3FB',
+    char: '👂🏻',
+    name: 'ear: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F442 1F3FC',
+    char: '👂🏼',
+    name: 'ear: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F442 1F3FD',
+    char: '👂🏽',
+    name: 'ear: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F442 1F3FE',
+    char: '👂🏾',
+    name: 'ear: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F442 1F3FF',
+    char: '👂🏿',
+    name: 'ear: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F443', char: '👃', name: 'nose', category: 'people' },
+  {
+    codes: '1F443 1F3FB',
+    char: '👃🏻',
+    name: 'nose: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F443 1F3FC',
+    char: '👃🏼',
+    name: 'nose: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F443 1F3FD',
+    char: '👃🏽',
+    name: 'nose: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F443 1F3FE',
+    char: '👃🏾',
+    name: 'nose: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F443 1F3FF',
+    char: '👃🏿',
+    name: 'nose: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F440', char: '👀', name: 'eyes', category: 'people' },
+  { codes: '1F445', char: '👅', name: 'tongue', category: 'people' },
+  { codes: '1F444', char: '👄', name: 'mouth', category: 'people' },
+  { codes: '1F476', char: '👶', name: 'baby', category: 'people' },
+  {
+    codes: '1F476 1F3FB',
+    char: '👶🏻',
+    name: 'baby: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F476 1F3FC',
+    char: '👶🏼',
+    name: 'baby: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F476 1F3FD',
+    char: '👶🏽',
+    name: 'baby: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F476 1F3FE',
+    char: '👶🏾',
+    name: 'baby: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F476 1F3FF',
+    char: '👶🏿',
+    name: 'baby: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F9D2 1F3FB',
+    char: '🧒🏻',
+    name: 'child: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F9D2 1F3FC',
+    char: '🧒🏼',
+    name: 'child: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F9D2 1F3FD',
+    char: '🧒🏽',
+    name: 'child: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F9D2 1F3FE',
+    char: '🧒🏾',
+    name: 'child: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F9D2 1F3FF',
+    char: '🧒🏿',
+    name: 'child: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F466', char: '👦', name: 'boy', category: 'people' },
+  {
+    codes: '1F466 1F3FB',
+    char: '👦🏻',
+    name: 'boy: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F466 1F3FC',
+    char: '👦🏼',
+    name: 'boy: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F466 1F3FD',
+    char: '👦🏽',
+    name: 'boy: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F466 1F3FE',
+    char: '👦🏾',
+    name: 'boy: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F466 1F3FF',
+    char: '👦🏿',
+    name: 'boy: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F467', char: '👧', name: 'girl', category: 'people' },
+  {
+    codes: '1F467 1F3FB',
+    char: '👧🏻',
+    name: 'girl: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F467 1F3FC',
+    char: '👧🏼',
+    name: 'girl: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F467 1F3FD',
+    char: '👧🏽',
+    name: 'girl: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F467 1F3FE',
+    char: '👧🏾',
+    name: 'girl: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F467 1F3FF',
+    char: '👧🏿',
+    name: 'girl: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F471 1F3FB',
+    char: '👱🏻',
+    name: 'person: light-skin-tone, blond hair',
+    category: 'people',
+  },
+  {
+    codes: '1F471 1F3FC',
+    char: '👱🏼',
+    name: 'person: medium-light-skin-tone, blond hair',
+    category: 'people',
+  },
+  {
+    codes: '1F471 1F3FD',
+    char: '👱🏽',
+    name: 'person: medium-skin-tone, blond hair',
+    category: 'people',
+  },
+  {
+    codes: '1F471 1F3FE',
+    char: '👱🏾',
+    name: 'person: medium-dark-skin-tone, blond hair',
+    category: 'people',
+  },
+  {
+    codes: '1F471 1F3FF',
+    char: '👱🏿',
+    name: 'person: dark-skin-tone, blond hair',
+    category: 'people',
+  },
+  { codes: '1F468', char: '👨', name: 'man', category: 'people' },
+  {
+    codes: '1F468 1F3FB',
+    char: '👨🏻',
+    name: 'man: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FC',
+    char: '👨🏼',
+    name: 'man: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FD',
+    char: '👨🏽',
+    name: 'man: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FE',
+    char: '👨🏾',
+    name: 'man: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FF',
+    char: '👨🏿',
+    name: 'man: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F469', char: '👩', name: 'woman', category: 'people' },
+  {
+    codes: '1F469 1F3FB',
+    char: '👩🏻',
+    name: 'woman: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FC',
+    char: '👩🏼',
+    name: 'woman: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FD',
+    char: '👩🏽',
+    name: 'woman: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FE',
+    char: '👩🏾',
+    name: 'woman: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FF',
+    char: '👩🏿',
+    name: 'woman: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F474', char: '👴', name: 'old man', category: 'people' },
+  {
+    codes: '1F474 1F3FB',
+    char: '👴🏻',
+    name: 'old man: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F474 1F3FC',
+    char: '👴🏼',
+    name: 'old man: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F474 1F3FD',
+    char: '👴🏽',
+    name: 'old man: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F474 1F3FE',
+    char: '👴🏾',
+    name: 'old man: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F474 1F3FF',
+    char: '👴🏿',
+    name: 'old man: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F475', char: '👵', name: 'old woman', category: 'people' },
+  {
+    codes: '1F475 1F3FB',
+    char: '👵🏻',
+    name: 'old woman: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F475 1F3FC',
+    char: '👵🏼',
+    name: 'old woman: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F475 1F3FD',
+    char: '👵🏽',
+    name: 'old woman: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F475 1F3FE',
+    char: '👵🏾',
+    name: 'old woman: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F475 1F3FF',
+    char: '👵🏿',
+    name: 'old woman: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64D',
+    char: '🙍',
+    name: 'person frowning',
+    category: 'people',
+  },
+  {
+    codes: '1F64D 1F3FB',
+    char: '🙍🏻',
+    name: 'person frowning: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64D 1F3FC',
+    char: '🙍🏼',
+    name: 'person frowning: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64D 1F3FD',
+    char: '🙍🏽',
+    name: 'person frowning: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64D 1F3FE',
+    char: '🙍🏾',
+    name: 'person frowning: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64D 1F3FF',
+    char: '🙍🏿',
+    name: 'person frowning: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64E',
+    char: '🙎',
+    name: 'person pouting',
+    category: 'people',
+  },
+  {
+    codes: '1F64E 1F3FB',
+    char: '🙎🏻',
+    name: 'person pouting: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64E 1F3FC',
+    char: '🙎🏼',
+    name: 'person pouting: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64E 1F3FD',
+    char: '🙎🏽',
+    name: 'person pouting: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64E 1F3FE',
+    char: '🙎🏾',
+    name: 'person pouting: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64E 1F3FF',
+    char: '🙎🏿',
+    name: 'person pouting: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F645',
+    char: '🙅',
+    name: 'person gesturing NO',
+    category: 'people',
+  },
+  {
+    codes: '1F645 1F3FB',
+    char: '🙅🏻',
+    name: 'person gesturing NO: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F645 1F3FC',
+    char: '🙅🏼',
+    name: 'person gesturing NO: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F645 1F3FD',
+    char: '🙅🏽',
+    name: 'person gesturing NO: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F645 1F3FE',
+    char: '🙅🏾',
+    name: 'person gesturing NO: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F645 1F3FF',
+    char: '🙅🏿',
+    name: 'person gesturing NO: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F646',
+    char: '🙆',
+    name: 'person gesturing OK',
+    category: 'people',
+  },
+  {
+    codes: '1F646 1F3FB',
+    char: '🙆🏻',
+    name: 'person gesturing OK: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F646 1F3FC',
+    char: '🙆🏼',
+    name: 'person gesturing OK: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F646 1F3FD',
+    char: '🙆🏽',
+    name: 'person gesturing OK: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F646 1F3FE',
+    char: '🙆🏾',
+    name: 'person gesturing OK: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F646 1F3FF',
+    char: '🙆🏿',
+    name: 'person gesturing OK: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F481',
+    char: '💁',
+    name: 'person tipping hand',
+    category: 'people',
+  },
+  {
+    codes: '1F481 1F3FB',
+    char: '💁🏻',
+    name: 'person tipping hand: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F481 1F3FC',
+    char: '💁🏼',
+    name: 'person tipping hand: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F481 1F3FD',
+    char: '💁🏽',
+    name: 'person tipping hand: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F481 1F3FE',
+    char: '💁🏾',
+    name: 'person tipping hand: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F481 1F3FF',
+    char: '💁🏿',
+    name: 'person tipping hand: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64B',
+    char: '🙋',
+    name: 'person raising hand',
+    category: 'people',
+  },
+  {
+    codes: '1F64B 1F3FB',
+    char: '🙋🏻',
+    name: 'person raising hand: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64B 1F3FC',
+    char: '🙋🏼',
+    name: 'person raising hand: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64B 1F3FD',
+    char: '🙋🏽',
+    name: 'person raising hand: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64B 1F3FE',
+    char: '🙋🏾',
+    name: 'person raising hand: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F64B 1F3FF',
+    char: '🙋🏿',
+    name: 'person raising hand: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F647',
+    char: '🙇',
+    name: 'person bowing',
+    category: 'people',
+  },
+  {
+    codes: '1F647 1F3FB',
+    char: '🙇🏻',
+    name: 'person bowing: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F647 1F3FC',
+    char: '🙇🏼',
+    name: 'person bowing: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F647 1F3FD',
+    char: '🙇🏽',
+    name: 'person bowing: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F647 1F3FE',
+    char: '🙇🏾',
+    name: 'person bowing: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F647 1F3FF',
+    char: '🙇🏿',
+    name: 'person bowing: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46E',
+    char: '👮',
+    name: 'police officer',
+    category: 'people',
+  },
+  {
+    codes: '1F46E 1F3FB',
+    char: '👮🏻',
+    name: 'police officer: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46E 1F3FC',
+    char: '👮🏼',
+    name: 'police officer: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46E 1F3FD',
+    char: '👮🏽',
+    name: 'police officer: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46E 1F3FE',
+    char: '👮🏾',
+    name: 'police officer: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46E 1F3FF',
+    char: '👮🏿',
+    name: 'police officer: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 200D 1F4BB',
+    char: '👨‍💻',
+    name: 'man technologist',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FB 200D 1F4BB',
+    char: '👨🏻‍💻',
+    name: 'man technologist: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FC 200D 1F4BB',
+    char: '👨🏼‍💻',
+    name: 'man technologist: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FD 200D 1F4BB',
+    char: '👨🏽‍💻',
+    name: 'man technologist: medium skin tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FE 200D 1F4BB',
+    char: '👨🏾‍💻',
+    name: 'man technologist: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FF 200D 1F4BB',
+    char: '👨🏿‍💻',
+    name: 'man technologist: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 200D 1F4BB',
+    char: '👩‍💻',
+    name: 'woman technologist',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FB 200D 1F4BB',
+    char: '👩🏻‍💻',
+    name: 'woman technologist: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FC 200D 1F4BB',
+    char: '👩🏼‍💻',
+    name: 'woman technologist: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FD 200D 1F4BB',
+    char: '👩🏽‍💻',
+    name: 'woman technologist: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FE 200D 1F4BB',
+    char: '👩🏾‍💻',
+    name: 'woman technologist: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FF 200D 1F4BB',
+    char: '👩🏿‍💻',
+    name: 'woman technologist: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F482', char: '💂', name: 'guard', category: 'people' },
+  {
+    codes: '1F482 1F3FB',
+    char: '💂🏻',
+    name: 'guard: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F482 1F3FC',
+    char: '💂🏼',
+    name: 'guard: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F482 1F3FD',
+    char: '💂🏽',
+    name: 'guard: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F482 1F3FE',
+    char: '💂🏾',
+    name: 'guard: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F482 1F3FF',
+    char: '💂🏿',
+    name: 'guard: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F477',
+    char: '👷',
+    name: 'construction worker',
+    category: 'people',
+  },
+  {
+    codes: '1F477 1F3FB',
+    char: '👷🏻',
+    name: 'construction worker: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F477 1F3FC',
+    char: '👷🏼',
+    name: 'construction worker: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F477 1F3FD',
+    char: '👷🏽',
+    name: 'construction worker: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F477 1F3FE',
+    char: '👷🏾',
+    name: 'construction worker: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F477 1F3FF',
+    char: '👷🏿',
+    name: 'construction worker: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F478', char: '👸', name: 'princess', category: 'people' },
+  {
+    codes: '1F478 1F3FB',
+    char: '👸🏻',
+    name: 'princess: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F478 1F3FC',
+    char: '👸🏼',
+    name: 'princess: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F478 1F3FD',
+    char: '👸🏽',
+    name: 'princess: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F478 1F3FE',
+    char: '👸🏾',
+    name: 'princess: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F478 1F3FF',
+    char: '👸🏿',
+    name: 'princess: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F473',
+    char: '👳',
+    name: 'person wearing turban',
+    category: 'people',
+  },
+  {
+    codes: '1F473 1F3FB',
+    char: '👳🏻',
+    name: 'person wearing turban: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F473 1F3FC',
+    char: '👳🏼',
+    name: 'person wearing turban: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F473 1F3FD',
+    char: '👳🏽',
+    name: 'person wearing turban: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F473 1F3FE',
+    char: '👳🏾',
+    name: 'person wearing turban: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F473 1F3FF',
+    char: '👳🏿',
+    name: 'person wearing turban: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F472',
+    char: '👲',
+    name: 'man with Chinese cap',
+    category: 'people',
+  },
+  {
+    codes: '1F472 1F3FB',
+    char: '👲🏻',
+    name: 'man with Chinese cap: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F472 1F3FC',
+    char: '👲🏼',
+    name: 'man with Chinese cap: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F472 1F3FD',
+    char: '👲🏽',
+    name: 'man with Chinese cap: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F472 1F3FE',
+    char: '👲🏾',
+    name: 'man with Chinese cap: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F472 1F3FF',
+    char: '👲🏿',
+    name: 'man with Chinese cap: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F470',
+    char: '👰',
+    name: 'bride with veil',
+    category: 'people',
+  },
+  {
+    codes: '1F470 1F3FB',
+    char: '👰🏻',
+    name: 'bride with veil: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F470 1F3FC',
+    char: '👰🏼',
+    name: 'bride with veil: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F470 1F3FD',
+    char: '👰🏽',
+    name: 'bride with veil: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F470 1F3FE',
+    char: '👰🏾',
+    name: 'bride with veil: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F470 1F3FF',
+    char: '👰🏿',
+    name: 'bride with veil: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F47C',
+    char: '👼',
+    name: 'baby angel',
+    category: 'people',
+  },
+  {
+    codes: '1F47C 1F3FB',
+    char: '👼🏻',
+    name: 'baby angel: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F47C 1F3FC',
+    char: '👼🏼',
+    name: 'baby angel: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F47C 1F3FD',
+    char: '👼🏽',
+    name: 'baby angel: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F47C 1F3FE',
+    char: '👼🏾',
+    name: 'baby angel: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F47C 1F3FF',
+    char: '👼🏿',
+    name: 'baby angel: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F385',
+    char: '🎅',
+    name: 'Santa Claus',
+    category: 'people',
+  },
+  {
+    codes: '1F385 1F3FB',
+    char: '🎅🏻',
+    name: 'Santa Claus: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F385 1F3FC',
+    char: '🎅🏼',
+    name: 'Santa Claus: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F385 1F3FD',
+    char: '🎅🏽',
+    name: 'Santa Claus: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F385 1F3FE',
+    char: '🎅🏾',
+    name: 'Santa Claus: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F385 1F3FF',
+    char: '🎅🏿',
+    name: 'Santa Claus: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F486',
+    char: '💆',
+    name: 'person getting massage',
+    category: 'people',
+  },
+  {
+    codes: '1F486 1F3FB',
+    char: '💆🏻',
+    name: 'person getting massage: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F486 1F3FC',
+    char: '💆🏼',
+    name: 'person getting massage: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F486 1F3FD',
+    char: '💆🏽',
+    name: 'person getting massage: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F486 1F3FE',
+    char: '💆🏾',
+    name: 'person getting massage: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F486 1F3FF',
+    char: '💆🏿',
+    name: 'person getting massage: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F487',
+    char: '💇',
+    name: 'person getting haircut',
+    category: 'people',
+  },
+  {
+    codes: '1F487 1F3FB',
+    char: '💇🏻',
+    name: 'person getting haircut: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F487 1F3FC',
+    char: '💇🏼',
+    name: 'person getting haircut: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F487 1F3FD',
+    char: '💇🏽',
+    name: 'person getting haircut: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F487 1F3FE',
+    char: '💇🏾',
+    name: 'person getting haircut: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F487 1F3FF',
+    char: '💇🏿',
+    name: 'person getting haircut: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C3',
+    char: '🏃',
+    name: 'person running',
+    category: 'people',
+  },
+  {
+    codes: '1F3C3 1F3FB',
+    char: '🏃🏻',
+    name: 'person running: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C3 1F3FC',
+    char: '🏃🏼',
+    name: 'person running: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C3 1F3FD',
+    char: '🏃🏽',
+    name: 'person running: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C3 1F3FE',
+    char: '🏃🏾',
+    name: 'person running: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C3 1F3FF',
+    char: '🏃🏿',
+    name: 'person running: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F483',
+    char: '💃',
+    name: 'woman dancing',
+    category: 'people',
+  },
+  {
+    codes: '1F483 1F3FB',
+    char: '💃🏻',
+    name: 'woman dancing: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F483 1F3FC',
+    char: '💃🏼',
+    name: 'woman dancing: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F483 1F3FD',
+    char: '💃🏽',
+    name: 'woman dancing: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F483 1F3FE',
+    char: '💃🏾',
+    name: 'woman dancing: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F483 1F3FF',
+    char: '💃🏿',
+    name: 'woman dancing: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46F',
+    char: '👯',
+    name: 'people with bunny ears',
+    category: 'people',
+  },
+  {
+    codes: '1F3C2',
+    char: '🏂',
+    name: 'snowboarder',
+    category: 'people',
+  },
+  {
+    codes: '1F3C2 1F3FB',
+    char: '🏂🏻',
+    name: 'snowboarder: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C2 1F3FC',
+    char: '🏂🏼',
+    name: 'snowboarder: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C2 1F3FD',
+    char: '🏂🏽',
+    name: 'snowboarder: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C2 1F3FE',
+    char: '🏂🏾',
+    name: 'snowboarder: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C2 1F3FF',
+    char: '🏂🏿',
+    name: 'snowboarder: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C4',
+    char: '🏄',
+    name: 'person surfing',
+    category: 'people',
+  },
+  {
+    codes: '1F3C4 1F3FB',
+    char: '🏄🏻',
+    name: 'person surfing: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C4 1F3FC',
+    char: '🏄🏼',
+    name: 'person surfing: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C4 1F3FD',
+    char: '🏄🏽',
+    name: 'person surfing: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C4 1F3FE',
+    char: '🏄🏾',
+    name: 'person surfing: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3C4 1F3FF',
+    char: '🏄🏿',
+    name: 'person surfing: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3CA',
+    char: '🏊',
+    name: 'person swimming',
+    category: 'people',
+  },
+  {
+    codes: '1F3CA 1F3FB',
+    char: '🏊🏻',
+    name: 'person swimming: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3CA 1F3FC',
+    char: '🏊🏼',
+    name: 'person swimming: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3CA 1F3FD',
+    char: '🏊🏽',
+    name: 'person swimming: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3CA 1F3FE',
+    char: '🏊🏾',
+    name: 'person swimming: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F3CA 1F3FF',
+    char: '🏊🏿',
+    name: 'person swimming: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F6B4',
+    char: '🚴',
+    name: 'person biking',
+    category: 'people',
+  },
+  {
+    codes: '1F6B4 1F3FB',
+    char: '🚴🏻',
+    name: 'person biking: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F6B4 1F3FC',
+    char: '🚴🏼',
+    name: 'person biking: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F6B4 1F3FD',
+    char: '🚴🏽',
+    name: 'person biking: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F6B4 1F3FE',
+    char: '🚴🏾',
+    name: 'person biking: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F6B4 1F3FF',
+    char: '🚴🏿',
+    name: 'person biking: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F6C0',
+    char: '🛀',
+    name: 'person taking bath',
+    category: 'people',
+  },
+  {
+    codes: '1F6C0 1F3FB',
+    char: '🛀🏻',
+    name: 'person taking bath: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F6C0 1F3FC',
+    char: '🛀🏼',
+    name: 'person taking bath: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F6C0 1F3FD',
+    char: '🛀🏽',
+    name: 'person taking bath: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F6C0 1F3FE',
+    char: '🛀🏾',
+    name: 'person taking bath: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F6C0 1F3FF',
+    char: '🛀🏿',
+    name: 'person taking bath: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46D',
+    char: '👭',
+    name: 'women holding hands',
+    category: 'people',
+  },
+  {
+    codes: '1F46D 1F3FB',
+    char: '👭🏻',
+    name: 'women holding hands: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FC 200D 1F91D 200D 1F469 1F3FB',
+    char: '👩🏼‍🤝‍👩🏻',
+    name: 'women holding hands: medium-light-skin-tone, light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46D 1F3FC',
+    char: '👭🏼',
+    name: 'women holding hands: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FD 200D 1F91D 200D 1F469 1F3FB',
+    char: '👩🏽‍🤝‍👩🏻',
+    name: 'women holding hands: medium-skin-tone, light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FD 200D 1F91D 200D 1F469 1F3FC',
+    char: '👩🏽‍🤝‍👩🏼',
+    name: 'women holding hands: medium-skin-tone, medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46D 1F3FD',
+    char: '👭🏽',
+    name: 'women holding hands: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FE 200D 1F91D 200D 1F469 1F3FB',
+    char: '👩🏾‍🤝‍👩🏻',
+    name: 'women holding hands: medium-dark-skin-tone, light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FE 200D 1F91D 200D 1F469 1F3FC',
+    char: '👩🏾‍🤝‍👩🏼',
+    name: 'women holding hands: medium-dark-skin-tone, medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FE 200D 1F91D 200D 1F469 1F3FD',
+    char: '👩🏾‍🤝‍👩🏽',
+    name: 'women holding hands: medium-dark-skin-tone, medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46D 1F3FE',
+    char: '👭🏾',
+    name: 'women holding hands: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FF 200D 1F91D 200D 1F469 1F3FB',
+    char: '👩🏿‍🤝‍👩🏻',
+    name: 'women holding hands: dark-skin-tone, light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FF 200D 1F91D 200D 1F469 1F3FC',
+    char: '👩🏿‍🤝‍👩🏼',
+    name: 'women holding hands: dark-skin-tone, medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FF 200D 1F91D 200D 1F469 1F3FD',
+    char: '👩🏿‍🤝‍👩🏽',
+    name: 'women holding hands: dark-skin-tone, medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FF 200D 1F91D 200D 1F469 1F3FE',
+    char: '👩🏿‍🤝‍👩🏾',
+    name: 'women holding hands: dark-skin-tone, medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46D 1F3FF',
+    char: '👭🏿',
+    name: 'women holding hands: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46B',
+    char: '👫',
+    name: 'woman and man holding hands',
+    category: 'people',
+  },
+  {
+    codes: '1F46B 1F3FB',
+    char: '👫🏻',
+    name: 'woman and man holding hands: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FB 200D 1F91D 200D 1F468 1F3FC',
+    char: '👩🏻‍🤝‍👨🏼',
+    name:
+      'woman and man holding hands: light-skin-tone, medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FB 200D 1F91D 200D 1F468 1F3FD',
+    char: '👩🏻‍🤝‍👨🏽',
+    name: 'woman and man holding hands: light-skin-tone, medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FB 200D 1F91D 200D 1F468 1F3FE',
+    char: '👩🏻‍🤝‍👨🏾',
+    name: 'woman and man holding hands: light-skin-tone, medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FB 200D 1F91D 200D 1F468 1F3FF',
+    char: '👩🏻‍🤝‍👨🏿',
+    name: 'woman and man holding hands: light-skin-tone, dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FC 200D 1F91D 200D 1F468 1F3FB',
+    char: '👩🏼‍🤝‍👨🏻',
+    name:
+      'woman and man holding hands: medium-light-skin-tone, light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46B 1F3FC',
+    char: '👫🏼',
+    name: 'woman and man holding hands: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FC 200D 1F91D 200D 1F468 1F3FD',
+    char: '👩🏼‍🤝‍👨🏽',
+    name:
+      'woman and man holding hands: medium-light-skin-tone, medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FC 200D 1F91D 200D 1F468 1F3FE',
+    char: '👩🏼‍🤝‍👨🏾',
+    name:
+      'woman and man holding hands: medium-light-skin-tone, medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FC 200D 1F91D 200D 1F468 1F3FF',
+    char: '👩🏼‍🤝‍👨🏿',
+    name: 'woman and man holding hands: medium-light-skin-tone, dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FD 200D 1F91D 200D 1F468 1F3FB',
+    char: '👩🏽‍🤝‍👨🏻',
+    name: 'woman and man holding hands: medium-skin-tone, light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FD 200D 1F91D 200D 1F468 1F3FC',
+    char: '👩🏽‍🤝‍👨🏼',
+    name:
+      'woman and man holding hands: medium-skin-tone, medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46B 1F3FD',
+    char: '👫🏽',
+    name: 'woman and man holding hands: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FD 200D 1F91D 200D 1F468 1F3FE',
+    char: '👩🏽‍🤝‍👨🏾',
+    name:
+      'woman and man holding hands: medium-skin-tone, medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FD 200D 1F91D 200D 1F468 1F3FF',
+    char: '👩🏽‍🤝‍👨🏿',
+    name: 'woman and man holding hands: medium-skin-tone, dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FE 200D 1F91D 200D 1F468 1F3FB',
+    char: '👩🏾‍🤝‍👨🏻',
+    name: 'woman and man holding hands: medium-dark-skin-tone, light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FE 200D 1F91D 200D 1F468 1F3FC',
+    char: '👩🏾‍🤝‍👨🏼',
+    name:
+      'woman and man holding hands: medium-dark-skin-tone, medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FE 200D 1F91D 200D 1F468 1F3FD',
+    char: '👩🏾‍🤝‍👨🏽',
+    name:
+      'woman and man holding hands: medium-dark-skin-tone, medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46B 1F3FE',
+    char: '👫🏾',
+    name: 'woman and man holding hands: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FE 200D 1F91D 200D 1F468 1F3FF',
+    char: '👩🏾‍🤝‍👨🏿',
+    name: 'woman and man holding hands: medium-dark-skin-tone, dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FF 200D 1F91D 200D 1F468 1F3FB',
+    char: '👩🏿‍🤝‍👨🏻',
+    name: 'woman and man holding hands: dark-skin-tone, light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FF 200D 1F91D 200D 1F468 1F3FC',
+    char: '👩🏿‍🤝‍👨🏼',
+    name: 'woman and man holding hands: dark-skin-tone, medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FF 200D 1F91D 200D 1F468 1F3FD',
+    char: '👩🏿‍🤝‍👨🏽',
+    name: 'woman and man holding hands: dark-skin-tone, medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F469 1F3FF 200D 1F91D 200D 1F468 1F3FE',
+    char: '👩🏿‍🤝‍👨🏾',
+    name: 'woman and man holding hands: dark-skin-tone, medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46B 1F3FF',
+    char: '👫🏿',
+    name: 'woman and man holding hands: dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46C',
+    char: '👬',
+    name: 'men holding hands',
+    category: 'people',
+  },
+  {
+    codes: '1F46C 1F3FB',
+    char: '👬🏻',
+    name: 'men holding hands: light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FC 200D 1F91D 200D 1F468 1F3FB',
+    char: '👨🏼‍🤝‍👨🏻',
+    name: 'men holding hands: medium-light-skin-tone, light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46C 1F3FC',
+    char: '👬🏼',
+    name: 'men holding hands: medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FD 200D 1F91D 200D 1F468 1F3FB',
+    char: '👨🏽‍🤝‍👨🏻',
+    name: 'men holding hands: medium-skin-tone, light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FD 200D 1F91D 200D 1F468 1F3FC',
+    char: '👨🏽‍🤝‍👨🏼',
+    name: 'men holding hands: medium-skin-tone, medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46C 1F3FD',
+    char: '👬🏽',
+    name: 'men holding hands: medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FE 200D 1F91D 200D 1F468 1F3FB',
+    char: '👨🏾‍🤝‍👨🏻',
+    name: 'men holding hands: medium-dark-skin-tone, light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FE 200D 1F91D 200D 1F468 1F3FC',
+    char: '👨🏾‍🤝‍👨🏼',
+    name: 'men holding hands: medium-dark-skin-tone, medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FE 200D 1F91D 200D 1F468 1F3FD',
+    char: '👨🏾‍🤝‍👨🏽',
+    name: 'men holding hands: medium-dark-skin-tone, medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46C 1F3FE',
+    char: '👬🏾',
+    name: 'men holding hands: medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FF 200D 1F91D 200D 1F468 1F3FB',
+    char: '👨🏿‍🤝‍👨🏻',
+    name: 'men holding hands: dark-skin-tone, light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FF 200D 1F91D 200D 1F468 1F3FC',
+    char: '👨🏿‍🤝‍👨🏼',
+    name: 'men holding hands: dark-skin-tone, medium-light-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FF 200D 1F91D 200D 1F468 1F3FD',
+    char: '👨🏿‍🤝‍👨🏽',
+    name: 'men holding hands: dark-skin-tone, medium-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F468 1F3FF 200D 1F91D 200D 1F468 1F3FE',
+    char: '👨🏿‍🤝‍👨🏾',
+    name: 'men holding hands: dark-skin-tone, medium-dark-skin-tone',
+    category: 'people',
+  },
+  {
+    codes: '1F46C 1F3FF',
+    char: '👬🏿',
+    name: 'men holding hands: dark-skin-tone',
+    category: 'people',
+  },
+  { codes: '1F48F', char: '💏', name: 'kiss', category: 'people' },
+  { codes: '1F46A', char: '👪', name: 'family', category: 'people' },
+  {
+    codes: '1F464',
+    char: '👤',
+    name: 'bust in silhouette',
+    category: 'people',
+  },
+  {
+    codes: '1F465',
+    char: '👥',
+    name: 'busts in silhouette',
+    category: 'people',
+  },
+  {
+    codes: '1F463',
+    char: '👣',
+    name: 'footprints',
+    category: 'people',
+  },
+];
