+++ emojis/places.js
@@ -1,0 +1,479 @@
+export default [
+  {
+    codes: '1F30F',
+    char: '🌏',
+    name: 'globe showing Asia-Australia',
+    category: 'places',
+  },
+  {
+    codes: '1F310',
+    char: '🌐',
+    name: 'globe with meridians',
+    category: 'places',
+  },
+  { codes: '1F30B', char: '🌋', name: 'volcano', category: 'places' },
+  {
+    codes: '1F5FB',
+    char: '🗻',
+    name: 'mount fuji',
+    category: 'places',
+  },
+  { codes: '1F3E0', char: '🏠', name: 'house', category: 'places' },
+  {
+    codes: '1F3E1',
+    char: '🏡',
+    name: 'house with garden',
+    category: 'places',
+  },
+  {
+    codes: '1F3E2',
+    char: '🏢',
+    name: 'office building',
+    category: 'places',
+  },
+  {
+    codes: '1F3E3',
+    char: '🏣',
+    name: 'Japanese post office',
+    category: 'places',
+  },
+  {
+    codes: '1F3E4',
+    char: '🏤',
+    name: 'post office',
+    category: 'places',
+  },
+  { codes: '1F3E5', char: '🏥', name: 'hospital', category: 'places' },
+  { codes: '1F3E6', char: '🏦', name: 'bank', category: 'places' },
+  { codes: '1F3E8', char: '🏨', name: 'hotel', category: 'places' },
+  {
+    codes: '1F3E9',
+    char: '🏩',
+    name: 'love hotel',
+    category: 'places',
+  },
+  {
+    codes: '1F3EA',
+    char: '🏪',
+    name: 'convenience store',
+    category: 'places',
+  },
+  { codes: '1F3EB', char: '🏫', name: 'school', category: 'places' },
+  {
+    codes: '1F3EC',
+    char: '🏬',
+    name: 'department store',
+    category: 'places',
+  },
+  { codes: '1F3ED', char: '🏭', name: 'factory', category: 'places' },
+  {
+    codes: '1F3EF',
+    char: '🏯',
+    name: 'Japanese castle',
+    category: 'places',
+  },
+  { codes: '1F3F0', char: '🏰', name: 'castle', category: 'places' },
+  { codes: '1F492', char: '💒', name: 'wedding', category: 'places' },
+  {
+    codes: '1F5FD',
+    char: '🗽',
+    name: 'Statue of Liberty',
+    category: 'places',
+  },
+  { codes: '26EA', char: '⛪', name: 'church', category: 'places' },
+  { codes: '26F2', char: '⛲', name: 'fountain', category: 'places' },
+  { codes: '26FA', char: '⛺', name: 'tent', category: 'places' },
+  { codes: '1F301', char: '🌁', name: 'foggy', category: 'places' },
+  {
+    codes: '1F303',
+    char: '🌃',
+    name: 'night with stars',
+    category: 'places',
+  },
+  {
+    codes: '1F304',
+    char: '🌄',
+    name: 'sunrise over mountains',
+    category: 'places',
+  },
+  { codes: '1F305', char: '🌅', name: 'sunrise', category: 'places' },
+  {
+    codes: '1F306',
+    char: '🌆',
+    name: 'cityscape at dusk',
+    category: 'places',
+  },
+  { codes: '1F307', char: '🌇', name: 'sunset', category: 'places' },
+  {
+    codes: '1F309',
+    char: '🌉',
+    name: 'bridge at night',
+    category: 'places',
+  },
+  {
+    codes: '2668 FE0F',
+    char: '♨️',
+    name: 'hot springs',
+    category: 'places',
+  },
+  {
+    codes: '1F3A0',
+    char: '🎠',
+    name: 'carousel horse',
+    category: 'places',
+  },
+  {
+    codes: '1F3A1',
+    char: '🎡',
+    name: 'ferris wheel',
+    category: 'places',
+  },
+  {
+    codes: '1F3A2',
+    char: '🎢',
+    name: 'roller coaster',
+    category: 'places',
+  },
+  {
+    codes: '1F488',
+    char: '💈',
+    name: 'barber pole',
+    category: 'places',
+  },
+  {
+    codes: '1F3AA',
+    char: '🎪',
+    name: 'circus tent',
+    category: 'places',
+  },
+  {
+    codes: '1F682',
+    char: '🚂',
+    name: 'locomotive',
+    category: 'places',
+  },
+  {
+    codes: '1F683',
+    char: '🚃',
+    name: 'railway car',
+    category: 'places',
+  },
+  {
+    codes: '1F684',
+    char: '🚄',
+    name: 'high-speed train',
+    category: 'places',
+  },
+  {
+    codes: '1F685',
+    char: '🚅',
+    name: 'bullet train',
+    category: 'places',
+  },
+  { codes: '1F686', char: '🚆', name: 'train', category: 'places' },
+  { codes: '1F687', char: '🚇', name: 'metro', category: 'places' },
+  {
+    codes: '1F688',
+    char: '🚈',
+    name: 'light rail',
+    category: 'places',
+  },
+  { codes: '1F689', char: '🚉', name: 'station', category: 'places' },
+  { codes: '1F68A', char: '🚊', name: 'tram', category: 'places' },
+  { codes: '1F69D', char: '🚝', name: 'monorail', category: 'places' },
+  {
+    codes: '1F69E',
+    char: '🚞',
+    name: 'mountain railway',
+    category: 'places',
+  },
+  { codes: '1F68B', char: '🚋', name: 'tram car', category: 'places' },
+  { codes: '1F68C', char: '🚌', name: 'bus', category: 'places' },
+  {
+    codes: '1F68D',
+    char: '🚍',
+    name: 'oncoming bus',
+    category: 'places',
+  },
+  {
+    codes: '1F68E',
+    char: '🚎',
+    name: 'trolleybus',
+    category: 'places',
+  },
+  { codes: '1F690', char: '🚐', name: 'minibus', category: 'places' },
+  { codes: '1F691', char: '🚑', name: 'ambulance', category: 'places' },
+  {
+    codes: '1F692',
+    char: '🚒',
+    name: 'fire engine',
+    category: 'places',
+  },
+  {
+    codes: '1F693',
+    char: '🚓',
+    name: 'police car',
+    category: 'places',
+  },
+  { codes: '1F695', char: '🚕', name: 'taxi', category: 'places' },
+  {
+    codes: '1F697',
+    char: '🚗',
+    name: 'automobile',
+    category: 'places',
+  },
+  {
+    codes: '1F698',
+    char: '🚘',
+    name: 'oncoming automobile',
+    category: 'places',
+  },
+  {
+    codes: '1F699',
+    char: '🚙',
+    name: 'sport utility vehicle',
+    category: 'places',
+  },
+  {
+    codes: '1F69A',
+    char: '🚚',
+    name: 'delivery truck',
+    category: 'places',
+  },
+  {
+    codes: '1F69B',
+    char: '🚛',
+    name: 'articulated lorry',
+    category: 'places',
+  },
+  { codes: '1F6B2', char: '🚲', name: 'bicycle', category: 'places' },
+  { codes: '26FD', char: '⛽', name: 'fuel pump', category: 'places' },
+  {
+    codes: '1F6A8',
+    char: '🚨',
+    name: 'police car light',
+    category: 'places',
+  },
+  {
+    codes: '1F6A5',
+    char: '🚥',
+    name: 'horizontal traffic light',
+    category: 'places',
+  },
+  {
+    codes: '1F6A6',
+    char: '🚦',
+    name: 'vertical traffic light',
+    category: 'places',
+  },
+  {
+    codes: '1F6A7',
+    char: '🚧',
+    name: 'construction',
+    category: 'places',
+  },
+  { codes: '2693', char: '⚓', name: 'anchor', category: 'places' },
+  { codes: '26F5', char: '⛵', name: 'sailboat', category: 'places' },
+  { codes: '1F6A4', char: '🚤', name: 'speedboat', category: 'places' },
+  { codes: '1F6A2', char: '🚢', name: 'ship', category: 'places' },
+  {
+    codes: '2708 FE0F',
+    char: '✈️',
+    name: 'airplane',
+    category: 'places',
+  },
+  {
+    codes: '1F681',
+    char: '🚁',
+    name: 'helicopter',
+    category: 'places',
+  },
+  {
+    codes: '1F69F',
+    char: '🚟',
+    name: 'suspension railway',
+    category: 'places',
+  },
+  {
+    codes: '1F6A0',
+    char: '🚠',
+    name: 'mountain cableway',
+    category: 'places',
+  },
+  {
+    codes: '1F6A1',
+    char: '🚡',
+    name: 'aerial tramway',
+    category: 'places',
+  },
+  {
+    codes: '1F6F0 FE0F',
+    char: '🛰️',
+    name: 'satellite',
+    category: 'places',
+  },
+  { codes: '1F680', char: '🚀', name: 'rocket', category: 'places' },
+  {
+    codes: '231B',
+    char: '⌛',
+    name: 'hourglass done',
+    category: 'places',
+  },
+  {
+    codes: '23F3',
+    char: '⏳',
+    name: 'hourglass not done',
+    category: 'places',
+  },
+  { codes: '231A', char: '⌚', name: 'watch', category: 'places' },
+  {
+    codes: '23F0',
+    char: '⏰',
+    name: 'alarm clock',
+    category: 'places',
+  },
+  {
+    codes: '1F55B',
+    char: '🕛',
+    name: 'twelve o’clock',
+    category: 'places',
+  },
+  {
+    codes: '1F567',
+    char: '🕧',
+    name: 'twelve-thirty',
+    category: 'places',
+  },
+  {
+    codes: '1F550',
+    char: '🕐',
+    name: 'one o’clock',
+    category: 'places',
+  },
+  {
+    codes: '1F55C',
+    char: '🕜',
+    name: 'one-thirty',
+    category: 'places',
+  },
+  {
+    codes: '1F551',
+    char: '🕑',
+    name: 'two o’clock',
+    category: 'places',
+  },
+  {
+    codes: '1F55D',
+    char: '🕝',
+    name: 'two-thirty',
+    category: 'places',
+  },
+  {
+    codes: '1F552',
+    char: '🕒',
+    name: 'three o’clock',
+    category: 'places',
+  },
+  {
+    codes: '1F55E',
+    char: '🕞',
+    name: 'three-thirty',
+    category: 'places',
+  },
+  {
+    codes: '1F553',
+    char: '🕓',
+    name: 'four o’clock',
+    category: 'places',
+  },
+  {
+    codes: '1F55F',
+    char: '🕟',
+    name: 'four-thirty',
+    category: 'places',
+  },
+  {
+    codes: '1F554',
+    char: '🕔',
+    name: 'five o’clock',
+    category: 'places',
+  },
+  {
+    codes: '1F560',
+    char: '🕠',
+    name: 'five-thirty',
+    category: 'places',
+  },
+  {
+    codes: '1F555',
+    char: '🕕',
+    name: 'six o’clock',
+    category: 'places',
+  },
+  {
+    codes: '1F561',
+    char: '🕡',
+    name: 'six-thirty',
+    category: 'places',
+  },
+  {
+    codes: '1F556',
+    char: '🕖',
+    name: 'seven o’clock',
+    category: 'places',
+  },
+  {
+    codes: '1F562',
+    char: '🕢',
+    name: 'seven-thirty',
+    category: 'places',
+  },
+  {
+    codes: '1F557',
+    char: '🕗',
+    name: 'eight o’clock',
+    category: 'places',
+  },
+  {
+    codes: '1F563',
+    char: '🕣',
+    name: 'eight-thirty',
+    category: 'places',
+  },
+  {
+    codes: '1F558',
+    char: '🕘',
+    name: 'nine o’clock',
+    category: 'places',
+  },
+  {
+    codes: '1F564',
+    char: '🕤',
+    name: 'nine-thirty',
+    category: 'places',
+  },
+  {
+    codes: '1F559',
+    char: '🕙',
+    name: 'ten o’clock',
+    category: 'places',
+  },
+  {
+    codes: '1F565',
+    char: '🕥',
+    name: 'ten-thirty',
+    category: 'places',
+  },
+  {
+    codes: '1F55A',
+    char: '🕚',
+    name: 'eleven o’clock',
+    category: 'places',
+  },
+  {
+    codes: '1F566',
+    char: '🕦',
+    name: 'eleven-thirty',
+    category: 'places',
+  },
+];
