+++ emojis/skintone.js
@@ -1,0 +1,32 @@
+export default [
+  {
+    codes: '1F3FB',
+    char: '🏻',
+    name: 'light-skin-tone',
+    category: 'skin-tone',
+  },
+  {
+    codes: '1F3FC',
+    char: '🏼',
+    name: 'medium-light-skin-tone',
+    category: 'skin-tone',
+  },
+  {
+    codes: '1F3FD',
+    char: '🏽',
+    name: 'medium-skin-tone',
+    category: 'skin-tone',
+  },
+  {
+    codes: '1F3FE',
+    char: '🏾',
+    name: 'medium-dark-skin-tone',
+    category: 'skin-tone',
+  },
+  {
+    codes: '1F3FF',
+    char: '🏿',
+    name: 'dark-skin-tone',
+    category: 'skin-tone',
+  }
+];
