+++ emojis/symbols.js
@@ -1,0 +1,705 @@
+export default [
+  {
+    codes: '2764 FE0F',
+    char: '❤️',
+    name: 'red heart',
+    category: 'symbols',
+  },
+  {
+    codes: '1F49B',
+    char: '💛',
+    name: 'yellow heart',
+    category: 'symbols',
+  },
+  {
+    codes: '1F49A',
+    char: '💚',
+    name: 'green heart',
+    category: 'symbols',
+  },
+  {
+    codes: '1F499',
+    char: '💙',
+    name: 'blue heart',
+    category: 'symbols',
+  },
+  {
+    codes: '1F49C',
+    char: '💜',
+    name: 'purple heart',
+    category: 'symbols',
+  },
+  {
+    codes: '1F48B',
+    char: '💋',
+    name: 'kiss mark',
+    category: 'symbols',
+  },
+  {
+    codes: '1F48C',
+    char: '💌',
+    name: 'love letter',
+    category: 'symbols',
+  },
+  {
+    codes: '1F498',
+    char: '💘',
+    name: 'heart with arrow',
+    category: 'symbols',
+  },
+  {
+    codes: '1F49D',
+    char: '💝',
+    name: 'heart with ribbon',
+    category: 'symbols',
+  },
+  {
+    codes: '1F496',
+    char: '💖',
+    name: 'sparkling heart',
+    category: 'symbols',
+  },
+  {
+    codes: '1F497',
+    char: '💗',
+    name: 'growing heart',
+    category: 'symbols',
+  },
+  {
+    codes: '1F493',
+    char: '💓',
+    name: 'beating heart',
+    category: 'symbols',
+  },
+  {
+    codes: '1F49E',
+    char: '💞',
+    name: 'revolving hearts',
+    category: 'symbols',
+  },
+  {
+    codes: '1F495',
+    char: '💕',
+    name: 'two hearts',
+    category: 'symbols',
+  },
+  {
+    codes: '1F49F',
+    char: '💟',
+    name: 'heart decoration',
+    category: 'symbols',
+  },
+  {
+    codes: '1F494',
+    char: '💔',
+    name: 'broken heart',
+    category: 'symbols',
+  },
+  {
+    codes: '1F4AF',
+    char: '💯',
+    name: 'hundred points',
+    category: 'symbols',
+  },
+  {
+    codes: '1F4A2',
+    char: '💢',
+    name: 'anger symbol',
+    category: 'symbols',
+  },
+  {
+    codes: '1F4A5',
+    char: '💥',
+    name: 'collision',
+    category: 'symbols',
+  },
+  { codes: '1F4AB', char: '💫', name: 'dizzy', category: 'symbols' },
+  {
+    codes: '1F4A6',
+    char: '💦',
+    name: 'sweat droplets',
+    category: 'symbols',
+  },
+  {
+    codes: '1F4A8',
+    char: '💨',
+    name: 'dashing away',
+    category: 'symbols',
+  },
+  { codes: '1F4A3', char: '💣', name: 'bomb', category: 'symbols' },
+  {
+    codes: '1F4AC',
+    char: '💬',
+    name: 'speech balloon',
+    category: 'symbols',
+  },
+  {
+    codes: '1F4AD',
+    char: '💭',
+    name: 'thought balloon',
+    category: 'symbols',
+  },
+  { codes: '1F4A4', char: '💤', name: 'zzz', category: 'symbols' },
+  {
+    codes: '1F6AE',
+    char: '🚮',
+    name: 'litter in bin sign',
+    category: 'symbols',
+  },
+  {
+    codes: '1F6B0',
+    char: '🚰',
+    name: 'potable water',
+    category: 'symbols',
+  },
+  {
+    codes: '267F',
+    char: '♿',
+    name: 'wheelchair symbol',
+    category: 'symbols',
+  },
+  {
+    codes: '1F6B9',
+    char: '🚹',
+    name: 'men’s room',
+    category: 'symbols',
+  },
+  {
+    codes: '1F6BA',
+    char: '🚺',
+    name: 'women’s room',
+    category: 'symbols',
+  },
+  { codes: '1F6BB', char: '🚻', name: 'restroom', category: 'symbols' },
+  {
+    codes: '1F6BC',
+    char: '🚼',
+    name: 'baby symbol',
+    category: 'symbols',
+  },
+  {
+    codes: '1F6BE',
+    char: '🚾',
+    name: 'water closet',
+    category: 'symbols',
+  },
+  {
+    codes: '1F6C2',
+    char: '🛂',
+    name: 'passport control',
+    category: 'symbols',
+  },
+  { codes: '1F6C3', char: '🛃', name: 'customs', category: 'symbols' },
+  {
+    codes: '1F6C4',
+    char: '🛄',
+    name: 'baggage claim',
+    category: 'symbols',
+  },
+  {
+    codes: '1F6C5',
+    char: '🛅',
+    name: 'left luggage',
+    category: 'symbols',
+  },
+  {
+    codes: '26A0 FE0F',
+    char: '⚠️',
+    name: 'warning',
+    category: 'symbols',
+  },
+  {
+    codes: '1F6B8',
+    char: '🚸',
+    name: 'children crossing',
+    category: 'symbols',
+  },
+  { codes: '26D4', char: '⛔', name: 'no entry', category: 'symbols' },
+  {
+    codes: '1F6AB',
+    char: '🚫',
+    name: 'prohibited',
+    category: 'symbols',
+  },
+  {
+    codes: '1F6B3',
+    char: '🚳',
+    name: 'no bicycles',
+    category: 'symbols',
+  },
+  {
+    codes: '1F6AD',
+    char: '🚭',
+    name: 'no smoking',
+    category: 'symbols',
+  },
+  {
+    codes: '1F6AF',
+    char: '🚯',
+    name: 'no littering',
+    category: 'symbols',
+  },
+  {
+    codes: '1F6B1',
+    char: '🚱',
+    name: 'non-potable water',
+    category: 'symbols',
+  },
+  {
+    codes: '1F6B7',
+    char: '🚷',
+    name: 'no pedestrians',
+    category: 'symbols',
+  },
+  {
+    codes: '1F4F5',
+    char: '📵',
+    name: 'no mobile phones',
+    category: 'symbols',
+  },
+  {
+    codes: '1F51E',
+    char: '🔞',
+    name: 'no one under eighteen',
+    category: 'symbols',
+  },
+  {
+    codes: '2622 FE0F',
+    char: '☢️',
+    name: 'radioactive',
+    category: 'symbols',
+  },
+  {
+    codes: '2623 FE0F',
+    char: '☣️',
+    name: 'biohazard',
+    category: 'symbols',
+  },
+  {
+    codes: '2B06 FE0F',
+    char: '⬆️',
+    name: 'up arrow',
+    category: 'symbols',
+  },
+  {
+    codes: '27A1 FE0F',
+    char: '➡️',
+    name: 'right arrow',
+    category: 'symbols',
+  },
+  {
+    codes: '2B07 FE0F',
+    char: '⬇️',
+    name: 'down arrow',
+    category: 'symbols',
+  },
+  {
+    codes: '2B05 FE0F',
+    char: '⬅️',
+    name: 'left arrow',
+    category: 'symbols',
+  },
+  {
+    codes: '21A9 FE0F',
+    char: '↩️',
+    name: 'right arrow curving left',
+    category: 'symbols',
+  },
+  {
+    codes: '21AA FE0F',
+    char: '↪️',
+    name: 'left arrow curving right',
+    category: 'symbols',
+  },
+  {
+    codes: '2934 FE0F',
+    char: '⤴️',
+    name: 'right arrow curving up',
+    category: 'symbols',
+  },
+  {
+    codes: '2935 FE0F',
+    char: '⤵️',
+    name: 'right arrow curving down',
+    category: 'symbols',
+  },
+  {
+    codes: '1F503',
+    char: '🔃',
+    name: 'clockwise vertical arrows',
+    category: 'symbols',
+  },
+  {
+    codes: '1F504',
+    char: '🔄',
+    name: 'counterclockwise arrows button',
+    category: 'symbols',
+  },
+  {
+    codes: '1F519',
+    char: '🔙',
+    name: 'BACK arrow',
+    category: 'symbols',
+  },
+  {
+    codes: '1F51A',
+    char: '🔚',
+    name: 'END arrow',
+    category: 'symbols',
+  },
+  {
+    codes: '1F51B',
+    char: '🔛',
+    name: 'ON! arrow',
+    category: 'symbols',
+  },
+  {
+    codes: '1F51C',
+    char: '🔜',
+    name: 'SOON arrow',
+    category: 'symbols',
+  },
+  {
+    codes: '1F51D',
+    char: '🔝',
+    name: 'TOP arrow',
+    category: 'symbols',
+  },
+  { codes: '2648', char: '♈', name: 'Aries', category: 'symbols' },
+  { codes: '2649', char: '♉', name: 'Taurus', category: 'symbols' },
+  { codes: '264A', char: '♊', name: 'Gemini', category: 'symbols' },
+  { codes: '264B', char: '♋', name: 'Cancer', category: 'symbols' },
+  { codes: '264C', char: '♌', name: 'Leo', category: 'symbols' },
+  { codes: '264D', char: '♍', name: 'Virgo', category: 'symbols' },
+  { codes: '264E', char: '♎', name: 'Libra', category: 'symbols' },
+  { codes: '264F', char: '♏', name: 'Scorpio', category: 'symbols' },
+  {
+    codes: '2650',
+    char: '♐',
+    name: 'Sagittarius',
+    category: 'symbols',
+  },
+  { codes: '2651', char: '♑', name: 'Capricorn', category: 'symbols' },
+  { codes: '2652', char: '♒', name: 'Aquarius', category: 'symbols' },
+  { codes: '2653', char: '♓', name: 'Pisces', category: 'symbols' },
+  { codes: '26CE', char: '⛎', name: 'Ophiuchus', category: 'symbols' },
+  {
+    codes: '1F500',
+    char: '🔀',
+    name: 'shuffle tracks button',
+    category: 'symbols',
+  },
+  {
+    codes: '1F501',
+    char: '🔁',
+    name: 'repeat button',
+    category: 'symbols',
+  },
+  {
+    codes: '1F502',
+    char: '🔂',
+    name: 'repeat single button',
+    category: 'symbols',
+  },
+  {
+    codes: '23E9',
+    char: '⏩',
+    name: 'fast-forward button',
+    category: 'symbols',
+  },
+  {
+    codes: '23EA',
+    char: '⏪',
+    name: 'fast reverse button',
+    category: 'symbols',
+  },
+  {
+    codes: '1F53C',
+    char: '🔼',
+    name: 'upwards button',
+    category: 'symbols',
+  },
+  {
+    codes: '23EB',
+    char: '⏫',
+    name: 'fast up button',
+    category: 'symbols',
+  },
+  {
+    codes: '1F53D',
+    char: '🔽',
+    name: 'downwards button',
+    category: 'symbols',
+  },
+  {
+    codes: '23EC',
+    char: '⏬',
+    name: 'fast down button',
+    category: 'symbols',
+  },
+  { codes: '1F3A6', char: '🎦', name: 'cinema', category: 'symbols' },
+  {
+    codes: '1F505',
+    char: '🔅',
+    name: 'dim button',
+    category: 'symbols',
+  },
+  {
+    codes: '1F506',
+    char: '🔆',
+    name: 'bright button',
+    category: 'symbols',
+  },
+  {
+    codes: '1F4F6',
+    char: '📶',
+    name: 'antenna bars',
+    category: 'symbols',
+  },
+  {
+    codes: '1F4F3',
+    char: '📳',
+    name: 'vibration mode',
+    category: 'symbols',
+  },
+  {
+    codes: '1F4F4',
+    char: '📴',
+    name: 'mobile phone off',
+    category: 'symbols',
+  },
+  {
+    codes: '267B FE0F',
+    char: '♻️',
+    name: 'recycling symbol',
+    category: 'symbols',
+  },
+  {
+    codes: '1F531',
+    char: '🔱',
+    name: 'trident emblem',
+    category: 'symbols',
+  },
+  {
+    codes: '1F4DB',
+    char: '📛',
+    name: 'name badge',
+    category: 'symbols',
+  },
+  {
+    codes: '1F530',
+    char: '🔰',
+    name: 'Japanese symbol for beginner',
+    category: 'symbols',
+  },
+  {
+    codes: '2B55',
+    char: '⭕',
+    name: 'hollow red circle',
+    category: 'symbols',
+  },
+  {
+    codes: '2705',
+    char: '✅',
+    name: 'check mark button',
+    category: 'symbols',
+  },
+  {
+    codes: '2611 FE0F',
+    char: '☑️',
+    name: 'check box with check',
+    category: 'symbols',
+  },
+  {
+    codes: '2714 FE0F',
+    char: '✔️',
+    name: 'check mark',
+    category: 'symbols',
+  },
+  {
+    codes: '274C',
+    char: '❌',
+    name: 'cross mark',
+    category: 'symbols',
+  },
+  {
+    codes: '274E',
+    char: '❎',
+    name: 'cross mark button',
+    category: 'symbols',
+  },
+  {
+    codes: '27B0',
+    char: '➰',
+    name: 'curly loop',
+    category: 'symbols',
+  },
+  {
+    codes: '27BF',
+    char: '➿',
+    name: 'double curly loop',
+    category: 'symbols',
+  },
+  {
+    codes: '303D FE0F',
+    char: '〽️',
+    name: 'part alternation mark',
+    category: 'symbols',
+  },
+  {
+    codes: '303D',
+    char: '〽',
+    name: 'part alternation mark',
+    category: 'symbols',
+  },
+  {
+    codes: '2733 FE0F',
+    char: '✳️',
+    name: 'eight-spoked asterisk',
+    category: 'symbols',
+  },
+  {
+    codes: '2734 FE0F',
+    char: '✴️',
+    name: 'eight-pointed star',
+    category: 'symbols',
+  },
+  {
+    codes: '2747 FE0F',
+    char: '❇️',
+    name: 'sparkle',
+    category: 'symbols',
+  },
+  {
+    codes: '2753',
+    char: '❓',
+    name: 'question mark',
+    category: 'symbols',
+  },
+  {
+    codes: '2757',
+    char: '❗',
+    name: 'exclamation mark',
+    category: 'symbols',
+  },
+  {
+    codes: '1F534',
+    char: '🔴',
+    name: 'red circle',
+    category: 'symbols',
+  },
+  {
+    codes: '1F535',
+    char: '🔵',
+    name: 'blue circle',
+    category: 'symbols',
+  },
+  {
+    codes: '26AB',
+    char: '⚫',
+    name: 'black circle',
+    category: 'symbols',
+  },
+  {
+    codes: '26AA',
+    char: '⚪',
+    name: 'white circle',
+    category: 'symbols',
+  },
+  {
+    codes: '2B1B',
+    char: '⬛',
+    name: 'black large square',
+    category: 'symbols',
+  },
+  {
+    codes: '2B1C',
+    char: '⬜',
+    name: 'white large square',
+    category: 'symbols',
+  },
+  {
+    codes: '25FC FE0F',
+    char: '◼️',
+    name: 'black medium square',
+    category: 'symbols',
+  },
+  {
+    codes: '25FB FE0F',
+    char: '◻️',
+    name: 'white medium square',
+    category: 'symbols',
+  },
+  {
+    codes: '25FE',
+    char: '◾',
+    name: 'black medium-small square',
+    category: 'symbols',
+  },
+  {
+    codes: '25FD',
+    char: '◽',
+    name: 'white medium-small square',
+    category: 'symbols',
+  },
+  {
+    codes: '1F536',
+    char: '🔶',
+    name: 'large orange diamond',
+    category: 'symbols',
+  },
+  {
+    codes: '1F537',
+    char: '🔷',
+    name: 'large blue diamond',
+    category: 'symbols',
+  },
+  {
+    codes: '1F538',
+    char: '🔸',
+    name: 'small orange diamond',
+    category: 'symbols',
+  },
+  {
+    codes: '1F539',
+    char: '🔹',
+    name: 'small blue diamond',
+    category: 'symbols',
+  },
+  {
+    codes: '1F53A',
+    char: '🔺',
+    name: 'red triangle pointed up',
+    category: 'symbols',
+  },
+  {
+    codes: '1F53B',
+    char: '🔻',
+    name: 'red triangle pointed down',
+    category: 'symbols',
+  },
+  {
+    codes: '1F4A0',
+    char: '💠',
+    name: 'diamond with a dot',
+    category: 'symbols',
+  },
+  {
+    codes: '1F518',
+    char: '🔘',
+    name: 'radio button',
+    category: 'symbols',
+  },
+  {
+    codes: '1F533',
+    char: '🔳',
+    name: 'white square button',
+    category: 'symbols',
+  },
+  {
+    codes: '1F532',
+    char: '🔲',
+    name: 'black square button',
+    category: 'symbols',
+  },
+];
