+++ js/main.mjs
@@ -1,12 +1,12 @@
 import initialSetting from './pages/settings/init.js';
 import makeEmojiContainers from './pages/main/emojiContainers.js';
-import addButtonEvents from './events/buttonEvent/index.js';
-import addSearchInputEvents from './events/inputEvent/search.js';
+import getEmoji from './pages/main/getEmoji.js';
 import makeInitialView from './components/index.js';
+import addButtonEvents from './events/buttonEvent/index.js';
 import addSkintoneButtonEvent from './events/buttonEvent/skintone.js';
-import getEmoji from './getEmoji.js';
 import { addAutoCopyOnEvent } from './events/emojiEvent/autoCopyOn.js';
 import { addAutoCopyOffEvent } from './events/emojiEvent/autoCopyOff.js';
+import addSearchInputEvents from './events/inputEvent/search.js';
 
 function setInit() {
   makeInitialView();
